% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-processing.R
\name{addDominantPFT}
\alias{addDominantPFT}
\title{Find dominant PFT}
\usage{
addDominantPFT(input, do.all = TRUE, do.tree = FALSE, do.woody = FALSE)
}
\arguments{
\item{input}{The ModelObject which is to have the shade tolerance classes combined.}

\item{do.all}{If TRUE calculate the dominant PFT out of all PFTs}

\item{do.tree}{If TRUE calculate the dominant PFT out of all tree PFTs}

\item{do.woody}{If TRUE calculate teh dominant PFT out of all woody PFTs}
}
\value{
The ModelObject with the dominant layers PFTs added
}
\description{
Find the dominant PFT (for each point in space and/or time) and stored in the ModelObject.  Can be of all PFTs, of all tree PFTs and/or of all woody PFTs 
The resulting layers are called "Dominant", "DominantTree" and "DominantWoody" respectively.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

