% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-framework-functions.R
\name{addMapOverlay}
\alias{addMapOverlay}
\title{Make a map overlay for ggplot2}
\usage{
addMapOverlay(map_plot, map_overlay)
}
\arguments{
\item{map_plot}{The ggplot to which the map overlay should be addeded}

\item{map_overlay}{A character string specifying the overlay to be used a string matching maps package dataset}
}
\value{
Returns data.frame suitable for plotting with ggplot::geom_path
}
\description{
Reads coastlines or country boundaries from the rnaturalearth package (rnaturalearthdata for the coastlines) and adds it to a ggplot.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
