% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-import.R
\name{addSeasonalityIndex}
\alias{addSeasonalityIndex}
\title{Adds a seasonality index column}
\usage{
addSeasonalityIndex(input = NULL, colname = "SI", verbose = TRUE)
}
\arguments{
\item{input}{Input data (either data.frame, data.table or ModelObject) with at least the month abbreviations as column names.}

\item{colname}{Name of the new created seasonality index column.}

\item{verbose}{print some information}
}
\value{
same type as input with the seasonality index as new data column
}
\description{
Adds a column of the seasonality index (\code{\link{seasonalityIndex}}) based on Marhan (1970) to the input data
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

