% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStandardSpatialAttributes.R
\name{addStandardSpatialAttributes}
\alias{addStandardSpatialAttributes}
\title{Adds standard spatial attributes to a netCDF file}
\usage{
addStandardSpatialAttributes(nc.file, lat.name = "Lat",
  lon.name = "Lon")
}
\arguments{
\item{nc.file}{The netcdf object to which the attributes should be added}
}
\description{
lala
}
\keyword{internal}
