% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{addToModelRun}
\alias{addToModelRun}
\title{Add an object (either a \code{ModelObject} or a  \code{SpatialComparison}) to a \code{ModelRun} object to be use later.}
\usage{
addToModelRun(object, run)
}
\arguments{
\item{object}{Object to add to the \code{ModelRun}.}

\item{run}{The \code{ModelRun} object to which the object argument should be added}
}
\value{
A ModelRun object the the object argument added
}
\description{
Stores an object in its run for later calculations, plotting, comparisons.
}
\seealso{
ModelRun, ModelRunInfo
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
