% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-runtools.R
\name{addToVegRun}
\alias{addToVegRun}
\title{Add an object (either a \code{VegObject}, \code{BiomeComparison} or \code{RasterComparison}) to a \code{VegRun} object to be use later.}
\usage{
addToVegRun(object, run)
}
\arguments{
\item{object}{Object to add to the \code{VegRun},   The parameters are the slots of a VegRunInfo object. Note that that \code{id} and \code{run.dir} are compulsory, the rest will be filled with dummy/default values if left blank.
compulsory.  Take care with \code{lon.lat.offset} and \code{year.offset} which are initialised to 0 which is unsuitable for some LPJ-GUESS confgurations.}

\item{run}{The \code{VegRun} object to which the object argument shold be added}
}
\value{
A VegRun object the the object argument added
}
\description{
Stores an object in its run for later calculations, plotting, comparisons.
}
\seealso{
VegRun, VegRunInfo
}

