% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfts-adgvm.R
\docType{data}
\name{aDGVM.PFTs}
\alias{aDGVM.PFTs}
\title{aDGVM PFTs (under development)}
\format{A list of \code{PFT} objects that store meta-data for PFTs classified from aDGVM model output.}
\usage{
aDGVM.PFTs
}
\description{
The list contains PFTs which could be used to classify individuals coming from the aDGVM
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{datasets}

