% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-aDGVM1.R
\name{aDGVM1QuantFromFilename}
\alias{aDGVM1QuantFromFilename}
\title{Helper function to raster::trim the ".out" or the ".out.gz" from an aDGVM1 filename to get the variable in question}
\usage{
aDGVM1QuantFromFilename(var.filename)
}
\arguments{
\item{var.filename}{The string of the filename to be raster::trimmed}
}
\value{
A string less the last fou.
}
\description{
Returns NULL if the last characters are not ".out" or ".out.gz
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
