% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome-schemes.R
\name{aDGVMBiomeRules}
\alias{aDGVMBiomeRules}
\title{Rules to classify tropical biomes from aDGVM1 output}
\usage{
aDGVMBiomeRules(x)
}
\arguments{
\item{x}{Numerical vector of vegetation over values for a particular location. 
Certain fractions and quantities should have been pre-calculated.}
}
\value{
Biomes code (1-5, baren, C4 grassland, C3 grassland, woodland, forest)
}
\description{
No reference yet
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}

Glenn Moncrief \email{glenn@saeon.ac.za}
}
\keyword{internal}
