% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-misc.R
\name{AGBtoTotalCarbon}
\alias{AGBtoTotalCarbon}
\title{Calculate total Carbon given above ground biomass.}
\usage{
AGBtoTotalCarbon(AGB)
}
\arguments{
\item{AGB}{Above ground biomass (carbon)}
}
\description{
Equation from Baccini et al . 2012. 
Used when reading the original Baccini et al. 2012 and Avitabile et al. 2015 dataset
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
