% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateSpatial.R
\name{aggregateSpatial}
\alias{aggregateSpatial}
\title{Spatially aggregate a Field}
\usage{
aggregateSpatial(x, method = "mean", na.rm = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{x}{Field (or data.table) to be averaged}

\item{method}{The method with which to spatially aggregate.  Can be "weighted.mean", "w.mean", "mean", 
"weighted.sum", "w.sum", "sum", "mode", "median", "max", "min", "sd", "var" and "cv" (= coefficient of variation: sd/mean).}

\item{na.rm}{Logical, passed to aggregation function to control whether or not to remove NAs before aggregating - i.e. the standard R usage.}

\item{verbose}{If TRUE give some progress update about the averaging.}

\item{...}{Extra arguments passed to addArea function if a weighted method is being used. Note in particular the lon_centres and lat_centres arguments
if you are using a regular but sparsely populated grid.}
}
\value{
A Field or data.table depending on the input object
}
\description{
Spatially aggregate all gridcells of a Field (or a data.table assuming it has columns "Lon" and "Lat").  Can use a variety of aggregation methods, 
see argument `method`
}
\details{
For summing variables which are defined on a per \eqn{m^2} use the method = "w.sum" (or "weighted.sum) to scale from \eqn{m^-2} to gridcell
totals before adding (gridcell area calculated by addArea() function).  For averaging, use method = "w.mean" (or "weighted.mean") to weight by the 
gridcell area.
}
\examples{
\donttest{
 
# Get an example Field
africa.dir <- system.file("extdata", "LPJ-GUESS_Runs", "CentralAfrica", package = "DGVMTools")
africa.Source <- defineSource(name = "LPJ-GUESS", dir = africa.dir,  format = GUESS)
field <- getField(source = africa.Source, quant = "cmass")

# calculate mean over gridcells and look at the resulting data table
spatial.mean <- aggregateSpatial(x = field, method = "mean", verbose = TRUE)
print(spatial.mean@data)

# do weighted.sum and look at the resulting data table
spatial.wsum <- aggregateSpatial(x = field, method = "w.sum", verbose = TRUE)
print(spatial.wsum@data) 

# plot the weight and unweighted sums as a time series
# (also modify ylab to be more meaningful)
p <- plotTemporal(list(spatial.wsum))
p <- p + ylab("Vegetation Carbon Mass (kgC)")
print(p)

}
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
