% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateSubannual.R
\name{aggregateSubannual.uncompiled}
\alias{aggregateSubannual.uncompiled}
\title{Sub-annual aggregation}
\usage{
aggregateSubannual.uncompiled(
  x,
  method = "mean",
  target = "Year",
  na.rm = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{x}{data.table or Field}

\item{method}{A character string describing the method by which to aggregate the data.  Can currently be "mean", "mode", "median", "sum", "max", "min", "sd", "var" and "cv" (= coefficient of variation sd/mean).
For technical reasons these need to be implemented in the package in the code however it should be easy to implement more, please just contact the author!}

\item{target}{A character string defining the subannual period to which the data should be aggregate. Can be "Month", "Season" or "Year" (also "Annual" is valid). 
Default is year.}

\item{na.rm}{Logical, passed to aggregation function to control whether or not to remove NAs before aggregating - i.e. the standard R usage.}

\item{verbose}{If TRUE give some progress update about the averaging.

Input data can be a Field or data.table with appropriate columns.}
}
\value{
A Field or data.table depending on the input object
}
\description{
Aggregates data with sub-annual time resolution to a coarser time resolution.  For example, going from monthly to annual.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
