% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateYears.R
\name{aggregateYears}
\alias{aggregateYears}
\title{Aggregate years}
\usage{
aggregateYears(x, method = "mean", na.rm = TRUE, verbose = FALSE)
}
\arguments{
\item{x}{A Field or data.table (with a "Year" column)}

\item{method}{A character string describing the method by which to aggregate the data.  Can currently be "mean", "mode", "median", "sum", "max", "min", "sd", "var" and "cv" (= coefficient of variation: sd/mean).
For technical reasons these need to be implemented in the package in the code however it should be easy to implement more, please just contact the author!}

\item{na.rm}{Logical, passed to aggregation function to control whether or not to remove NAs before aggregating - i.e. the standard R usage.}

\item{verbose}{If TRUE give some progress update about the averaging.}
}
\value{
A Field or data.table depending on the input object
}
\description{
Aggregated all available years in a Field or a data.table object (with years denoted by column "Years").
}
\examples{
\donttest{
 
# Get an example Field
test.dir <- system.file("extdata", "LPJ-GUESS_Runs", "CentralEurope", package = "DGVMTools")
test.Source <- defineSource(name = "LPJ-GUESS", dir = test.dir,  format = GUESS)
field <- getField(source = test.Source, quant = "lai")

# calculate of mean of all years
mean.allyears <- aggregateYears(x = field, method = "mean", verbose = TRUE)
print(mean.allyears@data)
print(plotSpatial(mean.allyears))

#  calculate standard deviation of all years
sd.allyears <- aggregateYears(x = field, method = "sd", verbose = FALSE)
print(sd.allyears@data)
print(plotSpatial(sd.allyears))


}
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
