% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsorted.R
\docType{methods}
\name{Array-methods}
\alias{Array-methods}
\alias{as.array,ModelObject-method}
\alias{as.array}
\title{Array methods}
\usage{
as.array(x, ...)

\S4method{as.array}{ModelObject}(x, ...)
}
\arguments{
\item{x}{\code{\linkS4class{ModelObject}}}

\item{...}{Other arguments, not currently used}
}
\value{
an lon/lat(/time) array - or a list of arrays - of the modelObjects input data.table.
}
\description{
Converts a \code{\linkS4class{ModelObject}} to multi-dimensional array(s) all parameters are passed along to \code{\link{modelObject2Array}}.
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}
