% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-aDGVM1.R
\name{availableQuantities_aDGVM1}
\alias{availableQuantities_aDGVM1}
\title{List all aDGVM1 *.out files in a run directory}
\usage{
availableQuantities_aDGVM1(source, names = TRUE, verbose = FALSE)
}
\arguments{
\item{source}{A aDGVM1 source object}

\item{names}{Logical, if TRUE return the namse of the quantities, if FLASE return the quanties themseleves}
}
\value{
A list of all the .out files present, with the ".out" removed.
}
\description{
Simply lists all aDGVM1 output variables (stored as .out files) available in a directory. 
Also ignores some common red herrings like "aDGVM1.out" and "*.out"
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
