% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-aDGVM2.R
\name{availableQuantities_aDGVM2}
\alias{availableQuantities_aDGVM2}
\title{List aDGVM2 quantities avialable}
\usage{
availableQuantities_aDGVM2(source, names, id, adgvm2.scheme)
}
\arguments{
\item{source}{A Source object defining the aDGVM2 model run}

\item{id}{An id of the model run}

\item{adgvm2.scheme}{A number that defines if pop-files (=1) or trait-files (=2) are used.}
}
\value{
A list of all the .out files present, with the ".out" removed.
}
\description{
Simply lists all quantitied aDGVM2  output variables
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
