% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-GUESS.R
\name{availableQuantities_GUESS}
\alias{availableQuantities_GUESS}
\title{List all LPJ-GUESS *.out files in a run directory}
\usage{
availableQuantities_GUESS(source, names = TRUE)
}
\arguments{
\item{source}{A GUESS source object}

\item{names}{Logical, if TRUE return the namse of the quantities, if FLASE return the quanties themseleves}
}
\value{
A list of all the .out files present, with the ".out" removed.
}
\description{
Simply lists all LPJ-GUESS output variables (stored as .out files) available in a directory. 
Also ignores some common red herrings like "guess.out" and "*.out"
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
