% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-NetCDF.R
\name{availableQuantities_NetCDF}
\alias{availableQuantities_NetCDF}
\title{List all LPJ-GUESS *.out files in a source directory}
\usage{
availableQuantities_NetCDF(source, names)
}
\arguments{
\item{source}{A \linkS4class{Source} containing the meta-data about the NetCDF source}

\item{names}{A boolean, if TRUE return a character vector of names of available quantities, if FALSE return a list of the actual Quantities.}
}
\value{
A list of all the .out files present, with the ".out" removed.
}
\description{
Simply lists all LPJ-GUESS output variables (stored as .out files) available in a directory. 
Also ignores some common red herrings like "guess.out" and "*.out"
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
