% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-runtools.R
\name{averageVegObjects}
\alias{averageVegObjects}
\title{Average VegObjects provided as a list}
\usage{
averageVegObjects(list.of.vegobjects, run = NULL)
}
\arguments{
\item{list.of.vegobjects}{The VegObjects that you want to average, stored in a list.}

\item{run}{Either a VegRun object which provides the metadata about the newly created "run" (representing, say, the ensemble mean) or \code{NULL}, 
in which case the function just returns a data.table with the average values.
 
The function currently does no checks that the spatial extent, temporal extent and VegQuant are consistent between runs.  The resulting VegObject simple takes these things
from the first object from list.of.vegruns.  It is therefore up to the user not do anything silly and average quantites, or spatial or temporal extents that don't makes sense.}
}
\value{
Either a VegObject (if a VegRun has been supplied to provide the necessary meta-data), or a data.table (if no meta-data has been supplies in the form of a VegRun)
}
\description{
Useful for averaging model ensembles, replicate simulations etc.  Fails with an error message if the supplied VegObjects don't have the same fields
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

