% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{benchmarkSpatial}
\alias{benchmarkSpatial}
\title{Benchmark ModelRuns against a spatial dataset}
\usage{
benchmarkSpatial(runs, layer.name, dataset, tag = NULL, diff.cuts = NULL,
  perc.diff.cuts = seq(-100, 200, 5), spatial.extent = NULL,
  showR2 = TRUE, showFitLine = TRUE, histo.plot.range = NULL,
  correction.dt = NULL, summary.plot.dir = NULL, canvas.options = list(dpi
  = 72, type = "png", width = 1800, height = 1000, bg = "transparent"), ...)
}
\arguments{
\item{runs}{A list of ModelRun objects (each with the RasterComparison object already created)}

\item{layer.name}{Layer in the vegetation model output to compare to the data}

\item{dataset}{The spatial dataset to which we are comparing (represented as a DataObject object)}

\item{tag}{A string with which to tag the resulting plots to specify the analysis (to differentiate them from other plots),
for example "ForcingDatasetComparison" or "SoilDepths.  Or whatever you want.}

\item{diff.cuts}{Numeric vector of cuts for plotting the absolute difference.  If not specified it is derived as the maximum
possible range based on the Quantity object in the dataset.}

\item{perc.diff.cuts}{Numerice vector of cuts for plotting the percentage differences.  If not specified defaults to -100\% to +200\%
in 5\% intervals.}

\item{spatial.extent}{If specified (as a raster::Extent object), the spatial extent to plot.  Note that the statistics will *not* reflect
this extent}

\item{showR2}{A logical, if TRUE, put the R2.eff values on the plots.}

\item{showFitLine}{A logical, if TRUE, put the a linear fit line on the scatter plots}

\item{histo.plot.range}{A two-value numeric vector defining the range for the histo plots}

\item{correction.dt}{A data.table with columns "Lon", "Lat" and "Correction" to apply a multiplicative correction to the model data before performing the benchmark 
to account for (for example) land cover.  NA's can be used to completely mask out areas from the comparison.}

\item{summary.plot.dir}{A directory (full path as a character string) so save the plots which compare many runs}

\item{canvas.options}{A list of options (to be given to the \code{Cairo}) function to define the canvas. See the \code{Cairo} dicumentation}

\item{...}{Arguments to be passed to plotSpatial when making the spatial plots.}
}
\value{
Returns a DataObject, and also makes plots (in the individual run@run.dir and in the summary.plot.dir argument).
}
\description{
Take a list of model runs and compares each run to a DataObject (which should obviously have spatial dimensions).  
Comparsion plots to each individual run are saved in the directory for that particular run, and overall plots (simultaneously showing all the runs 
next to the data) are saved in the \code{summary.plor.dir} argument.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

