% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000rvc-classes.R
\docType{class}
\name{BiomeComparison-class}
\alias{BiomeComparison-class}
\title{Result of comparing a modelled biome map to a data biome map}
\description{
This class stores the rasters (model, data), the biome scheme used and the Cohen's Kappa statistics resulting when two rasters of with identical biome categorisations are compared using \code{compareBiomes}
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular biome compariosn.  Recommended to be alphanumeric because it is used to construct file names.}

\item{\code{Kappa}}{The overall Cohen's Kappa obtained when comparing modelled biomes to data biomes.}

\item{\code{individual.Kappas}}{The individual Cohen's Kappas for each biome obtained when comparing modelled to data biomes}
}}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

