% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-classes.R
\docType{class}
\name{BiomeScheme-class}
\alias{BiomeScheme-class}
\title{A biome scheme}
\description{
This class stores the information about a biome scheme.  It describes what how the model output (in the form of a data.table) must be prepared, and then the rules which are used to classify the biomes.
}
\details{
This is ultimately not flexible enough for all conceivable biome schemes from all models and will need to be somehow generalised and expanded at some point.  This is most certainly a challenge for another day!
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular biome scheme.  Recommended to be alphanumeric because it is used to construct file names. (Inherited from Quantity via "contains")}

\item{\code{name}}{A character string that can be more descriptive of the biome scheme. (Inherited from Quantity via "contains")}

\item{\code{type}}{A character string defining the type of Quantity, here should always be "categorical" (Inherited from Quantity via "contains")}

\item{\code{units}}{A list of character strings giving the names of categories (biomes). (Inherited from Quantity via "contains")}

\item{\code{colours}}{A function that returns the colour scale for this BiomeScheme. (Inherited from Quantity via "contains")}

\item{\code{aggregate.method}}{A character string defining the default method for how to aggregate the quantity, should be set to "categorical" for BiomeSchemes (Inherited from Quantity via "contains")}

\item{\code{model}}{Either a the string "Standard" to denote that this is a  standard quantity to be compared across all model and data, or vector of model names to denote to which models this BiomeScheme can generally be applied to. (Inherited from Quantity via "contains")}

\item{\code{rules}}{A function which is applied to every row of the data.table and describes the biome classification rules.}

\item{\code{totals.needed}}{List of vegetation totals needed to calculate biomes, for example c("Tree", "Grass")}

\item{\code{max.needed}}{List of maximum values needed to calculate biomes, for example c("PFT", "Tree")}

\item{\code{fraction.of.total}}{List of vegetation fraction of totals needed to calculate biomes}

\item{\code{fraction.of.tree}}{List of vegetation fraction of tree total needed to calculate biomes}

\item{\code{fraction.of.woody}}{List of vegetation fraction of woody total needed to calculate biomes}

\item{\code{needGDD5}}{If TRUE the biome rules require GDD5 for the classification}

\item{\code{data.reference}}{Character string giving a reference where the data for this biome scheme comes from}

\item{\code{published.reference}}{Character string giving a reference where this model output classification scheme was published}
}}

\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
