% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000rvc-tools.R
\name{byIDfromList}
\alias{byIDfromList}
\title{Retrieves an object from a list based on it's \code{id} slot}
\usage{
byIDfromList(id, list)
}
\arguments{
\item{id}{The id sought (must be string)}

\item{list}{The list to be scanned (must be a list)}
}
\value{
The matching object from the list or NULL.  If NULL it will give a message and a warning, the following code will probably fail.
}
\description{
Looks through a list of arbitary object until it finds one which has a slot called "id" whose value matches the imput argument.
The idea is that you can use this fucntion to pull a particular PFT from a list of PFT objects, or a a particular run from a list of ModelRun objects etc.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

