% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{calcNashSutcliffe}
\alias{calcNashSutcliffe}
\title{Calculate Nash-Sutcliff Model Efficiency}
\usage{
calcNashSutcliffe(vector2, vector1)
}
\arguments{
\item{vector2}{A numeric vector of data (same size as vector2)}

\item{vector1}{A numeric vector of data}
}
\value{
A numeric
}
\description{
Calculates Nash-Sutcliff Model Efficiency between two datasets (represented as two equally sized numeric vectors)
}
\details{
No check currently done on vector lengths. See:
Nash, J. E. and J. V. Sutcliffe (1970), River flow forecasting through conceptual models part I - A discussion of principles, Jounral of Hydrology, 10 (3), 282-290.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
