% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridarea.R
\name{calcNewModelObj}
\alias{calcNewModelObj}
\title{calcNewModelObj: Simple calculation of new ModelObjects}
\usage{
calcNewModelObj(run = NULL, targets = NULL, operator = NULL,
  quant = NULL, verbose = TRUE)
}
\arguments{
\item{run}{the run the data should be taken from}

\item{targets}{a list with elements 'x' and 'y' with components ModelObject name and optionally column name. If neither x nor y has a column name the columns of x and y must be equal.}

\item{operator}{which arithmetic should be performed: addition ('+'), substraction ('-'), multiplication ('*') or division ('/')}

\item{quant}{new Quantity definition to use, if NULL it will be guessed}

\item{verbose}{print some messages}
}
\value{
hopefully a new ModelObject
}
\description{
Simple calculation of new ModelObjects by operation '+', '-', '*' or '/'
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

