% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{calcNMSE}
\alias{calcNMSE}
\title{Calculate Normalised Mean Square Error}
\usage{
calcNMSE(mod, obs, weights)
}
\arguments{
\item{mod}{A numeric vector of observed values}

\item{obs}{A numeric vector of modelled values (same size as mod)}

\item{weights}{A numeric vector of weights the values, typically the gridcell areas (same size as obs)}
}
\value{
A numeric
}
\description{
Calculates NMSE between two datasets (represented as two equally sized numeric vectors)
}
\details{
No check currently done on vector lengths
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
