% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{calcR2}
\alias{calcR2}
\title{Calculate Coefficient of Determination}
\usage{
calcR2(vector2, vector1)
}
\arguments{
\item{vector2}{A numeric vector of data (same size as vector2)}

\item{vector1}{A numeric vector of data}
}
\value{
A numeric
}
\description{
Calculates Coefficient of Determination between two datasets (represented as two equally sized numeric vectors)
}
\details{
No check currently done on vector lengths.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
