% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{categoricalComparison}
\alias{categoricalComparison}
\title{Comparison between two datasets of categorical variables}
\usage{
categoricalComparison(
  x,
  layers1,
  layers2,
  additional,
  verbose = TRUE,
  area = TRUE,
  tolerance = 0.01
)
}
\arguments{
\item{x}{A data.table containing the spatial-temporal-annual columns and two columns containg the data to be compared}

\item{layers1}{A character string giving the first layer to compare (should be a column in x)}

\item{layers2}{A character string giving the second layer to compare (should be a column in x)}

\item{additional}{A list of functions define additions metrics, see the custom.metrics argument of \code{\link{compareLayers}}}

\item{verbose}{A logical, if TRUE print out all the Kappa scores}

\item{area}{A logical, if true weight the metrics by gridcell area (not currently implemented)

Note that there are many other slots in a Statistics object which will not be filled in the resulting object because they are for continuous as opposed to categorical data}
}
\value{
A named list of metric statistics
}
\description{
Calculates a Statistics object (which contains the Cohen's Kappa scores) given data.table of with two layers of categorical data (eg. biomes, land cover classes).
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
