% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo.R
\name{cdo}
\alias{cdo}
\title{CDO wrapper}
\usage{
cdo(fun, ifile, ofile, grid = NULL, return.raster = FALSE, verbose = TRUE,
  remove.temps = TRUE, ...)
}
\arguments{
\item{fun}{The CDO operator to be used (right now must be remaplaf)}

\item{ifile}{The input, either a RasterLayer/Stack/Brick or a character string pointing to netCDF file on disk}

\item{ofile}{Character sting for the output file (path on disk)}

\item{grid}{Character string specifying the grid for the output file. Can be a code (sag, "n31' or "r192x145", or a file on disk), see CDO documentation}

\item{return.raster}{If TRUE, read the resulting netCDF file and return it is a a raster}

\item{verbose}{If TRUE print some progress and diagnostic output}

\item{remove.temps}{Logical, if TRUE remove the intermediate .nc files from disk

Basically just calls CDO on the commandline.  If the input arguments is an R RasterLayer/Stack/Brick, then is uses writeRaster() to write a netCDF file for processing (in the directory of the output file)}

\item{varname}{If operating on a raster (not a netCDF file) the name for the variable in the resulting netCDF file.}
}
\value{
NULL, unless return.raster is TRUE (j which case it returns a raster of the result of the cdo processing)
}
\description{
Right now only defined for remaplaf, but in principle it is easy to extend to other operators.  Operator chaining is not supported.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
