% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdo.R
\name{check.cdo}
\alias{check.cdo}
\title{Check for CDO}
\usage{
check.cdo()
}
\value{
TRUE if cdo present, false if not
}
\description{
Checks if the CDO (climate data operators) are available on the system
}
\keyword{internal}

