% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-framework-functions.R
\name{checkDimensionValues}
\alias{checkDimensionValues}
\title{Check values of a particular dimension}
\usage{
checkDimensionValues(fields, input.values = NULL, dimension)
}
\arguments{
\item{fields}{The list of Fields to be plotted (should have been check by santiseFieldsForPlotting first)}

\item{input.values}{A list of the values that have been requested to be plotted}

\item{dimension}{A character string specifying which dimension is to be checked (one of "Lon", "Lat", "Year", "Month", "Day" or "Season")}
}
\value{
Returns character vector of the layers to be plotted
}
\description{
This is an internal helper function which either stops if a dimension is requested to be plotted but is not present
dimension, it gives a warning
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
