% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000rvc-classes.R
\name{checkModelRun}
\alias{checkModelRun}
\title{Checks validity of a \code{ModelRunInfo}.}
\usage{
checkModelRun(object)
}
\arguments{
\item{object}{The \code{ModelRunInfo} object to check for vailidity.}
}
\value{
Empty string if the essential slots are fine, a string containing an error message if not.
}
\description{
Called internally as the validity slot of the \code{ModelRunInfo}.  It checks that the essential slots are filled with sensible values ie a run.dir that
exists on the file system; a model type which is valid and an \code{id} that is a non-empty character string.  It doesn't check that this is alphanumeric, this would be a useful addition.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

