% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{checkSource}
\alias{checkSource}
\title{Checks validity of a \code{Source}.}
\usage{
checkSource(object)
}
\arguments{
\item{object}{The \code{Source} object to check for vailidity.}
}
\value{
Empty string if the essential slots are fine, a string containing an error message if not.
}
\description{
Called internally as the validity slot of the \code{Source}.  It checks that the essential slots are filled with sensible values ie a dir that
exists on the file system; the format type is specified and an \code{id} that is a non-empty character string.  It doesn't check that this is alphanumeric, this would be a useful addition.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
