% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSTAMatches.R
\name{checkSTAMatches}
\alias{checkSTAMatches}
\title{Checks if STAInfo matches between Fields}
\usage{
checkSTAMatches(sta.requested, sta.found, verbose)
}
\arguments{
\item{sta.requested}{The STAInfo requested (in the getField() call)}

\item{sta.found}{The STAInfo found the in the existing Field (the one which will have been read from disk)}

\item{verbose}{If TRUE give a lot of information for debugging/checking.}
}
\value{
Logical
}
\description{
This is an internal function, used when checking preprocessed Fields in getField().
}
\keyword{internal}
