% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonSTAInfo.R
\name{commonSTAInfo}
\alias{commonSTAInfo}
\title{Checks common for STAInfo object}
\usage{
commonSTAInfo(sta.objects, logical = FALSE)
}
\arguments{
\item{sta.objects}{A list of Field and/or STAInfo objects which should be compared to determine the common STAInfo}

\item{logical}{Logical, if TRUE return a simple TRUE/FALSE rather than an STAInfo object of common info.}
}
\value{
Either a logical or an STAInfo object, depending on the value of 'logical'
}
\description{
Given a list of STAInfo or Field objects, this functions determines the common spatial-temporal-annual info.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
