% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-benchmarking.R
\name{compareBiomes}
\alias{compareBiomes}
\title{Compare a run to a biome map}
\usage{
compareBiomes(run, variable, period, scheme, plot = TRUE, ...)
}
\arguments{
\item{run}{The VegRun to be compared to the data}

\item{variable}{The VegQuant (or name) of the variable from which the biomes are to be derived. This is not flexible enough, 
and should be folded into the scheme.}

\item{period}{The temporal extent over which the model output should be averaged for calculating the biomes}

\item{scheme}{The biome scheme we are comparing. Also not flexible enough, here should provide a raster (or SpatialDataset?) 
containing the data}

\item{plot}{Logical, if true make a biome plot.}

\item{...}{Additional parameters supplied to the plotting function (at time of writing the plotting function is plotBiomes(),
but should be changed soon to plotVegMaps())}
}
\value{
A Biome comparison object
}
\description{
Takes a run, a time period and a variable from which to calculate a modelled biome map, and biome scheme, 
and compares the model to the data
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

