% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-benchmarking.R
\name{compareRunToSpatialDataset}
\alias{compareRunToSpatialDataset}
\title{Compare a layer from a VegObject to a spatial dataset}
\usage{
compareRunToSpatialDataset(dataset, vegobject, layer, diff.cuts = NULL,
  histo.plot.range = NULL, doScatterPlots = TRUE, doHistoPlots = TRUE,
  quant = NULL, ignore.raster = NULL, ...)
}
\arguments{
\item{dataset}{The data set as a SpatialDataset object}

\item{vegobject}{The VegObject which contains the layer to be compared from the model (and, helpfully, meta-data about the run)}

\item{layer}{The name of the layer to be compared to the spatial dataset}

\item{diff.cuts}{A sequence of bins for the histo plot (defaults to the cuts)}

\item{histo.plot.range}{A two-value numeric vector defining the range for the histo plots}

\item{doScatterPlots}{Logical, if TRUE make a scatter plot (default is TRUE)}

\item{doHistoPlots}{Logical, if TRUE make histograms plots (default is TRUE)}

\item{quant}{A VegQuant object to override the one from vegobject (if necessary)}

\item{ignore.raster}{A raster used to mask out gridcells wich are not to be compared (ie any gridcells set to NA in the ignore.raster
are set to NA in thr model and data rasters)}

\item{...}{Parameters passed to plotVegMaps() for the spatial plots.}
}
\value{
An object of class RasterComparison
}
\description{
Produces standard plot to compare model output to a spatial dataset.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

