% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-benchmarking.R
\name{compareTwoRastersStats}
\alias{compareTwoRastersStats}
\title{Statistically compare two rasters}
\usage{
compareTwoRastersStats(data.raster, model.raster)
}
\arguments{
\item{data.raster}{The data as a raster}

\item{model.raster}{The model output as a raster}
}
\value{
An object of class RasterComparison

The choice of statistical measures might not be the best but can easily be improved.  No spatial correlations are considered.
}
\description{
Returns the R^2, RMSE and Pearson's correlation coefficient between two rasters, as well as the rasters themselves 
(on a common extent and with the NA areas masked from both data datasets), 
a raster of the difference and their percentage difference (relative to the data).
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

