% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-benchmarking.R
\name{compareVegSpatialObject}
\alias{compareVegSpatialObject}
\title{Compare runs to each other}
\usage{
compareVegSpatialObject(runs, veg.spatial.id, layer, expand.layer = TRUE,
  base.run.id = NULL, abs.value.cuts = NULL, plot.diff = TRUE,
  diff.cuts = NULL, plot.perc.diff = TRUE, perc.diff.cuts = NULL,
  special = NULL, doIndividual = FALSE, tag = NULL, ...)
}
\arguments{
\item{runs}{List of VegRun objects to be compared}

\item{veg.spatial.id}{Character string holding the id of the VegObject to be compared}

\item{layer}{The layer to be compared across all VegRuns}

\item{expand.layer}{Logical, if TRUE expand the layer arguments use \code{expandLayers}}

\item{base.run.id}{A character string to which (if specified) all the other runs are compared by plotting the difference 
and percentage difference relative to this base run}

\item{abs.value.cuts}{A numeric vector which (if specified), defines the cuts for the plot of the absolute values of each 
run on one figure}

\item{plot.diff}{Logical, if TRUE and a base.run.id supplied, plot the absolute difference of each run relative to the base run.}

\item{diff.cuts}{A numeric vector which (if specified), defines the cuts for the difference plot of each run compared to the base run}

\item{plot.perc.diff}{Logical, if TRUE and a base.run.id supplied, plot the percentage difference of each run relative to the base run.}

\item{perc.diff.cuts}{A numeric vector which (if specified), defines the cuts for the percentage difference plot of each run compared to the base run}

\item{special}{A character string which, if specified, is used to give special plotting instructions to \code{plotVegMaps}, 
see the documentation of that function for details.}

\item{doIndividual}{Logical, if TRUE, make individual comparison plot for, say, all PFTs, as well as putting them all on one plot.}

\item{tag}{A character string to identify this plot/analysis in comparison to others, eg. "SoilDepthTests" or "NewSLAForumalation"}

\item{...}{Further arguments passed to the \code{plotVegMaps} function.}
}
\description{
Given a list of runs and an id of a spatial VegObjects, compare the specified layers between all runs, and, 
if a base line run is specified, also plot comaprisons relative to that
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

