% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adgvm-runtools.R
\name{convertMonthlyScheme1}
\alias{convertMonthlyScheme1}
\title{Simple classification of monthly output for aDGVM}
\usage{
convertMonthlyScheme1(runid, fire, directory)
}
\arguments{
\item{runid}{Character string identifying the run}

\item{fire}{Character string denoting if fire was on or off}

\item{directory}{Character string specifying the run directory}
}
\description{
Original from Simon Scheiter, ***THERE IS CURRENTLY NO INTERGATED VERSION OF THIS WHICH FITS INTO DGVM TOOLS*
Someone needs to do that...    
Classifies only as tree or grass based on only the pop file, doesn't open the trait file.
}
\author{
Simon Scheiter \email{simon.scheiter@senckenberg.de}
}
\seealso{
\code{getQuantity_aDGVM_Scheme1}
}
\keyword{internal}

