% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-adgvm.R
\name{convertYearlyScheme1}
\alias{convertYearlyScheme1}
\title{Simple classification of yearly output for aDGVM}
\usage{
convertYearlyScheme1(runid, fire, directory)
}
\arguments{
\item{runid}{Character string identifying the run}

\item{fire}{Character string denoting if fire was on or off}

\item{directory}{Character string specifying the run directory}
}
\description{
Original from Simon Scheiter, a modified version has been produced by M. Forrest to fit into the DGVM framework and is probably now
redundant.  Classifies only as tree or grass based on only the pop file, doesn't open the trait file.
}
\author{
Simon Scheiter \email{simon.scheiter@senckenberg.de}
}
\seealso{
\code{getQuantity_aDGVM_Scheme1}
}
\keyword{internal}

