% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-adgvm.R
\name{convertYearlyScheme2}
\alias{convertYearlyScheme2}
\title{Slightly more complex classification of yearly output for aDGVM output}
\usage{
convertYearlyScheme2(runid, fire, directory)
}
\arguments{
\item{runid}{Character string identifying the run}

\item{fire}{Character string denoting if fire was on or off}

\item{directory}{Character string specifying the run directory}
}
\description{
Original from Simon Scheiter, a modified version has been produced by M. Forrest to fit into the DGVM framework and is probably now
redundant.  Classifies as evergreen tree, deciduous tree or grass based on the pop file and the trait file.
}
\seealso{
\code{getQuantity_aDGVM_Scheme2}
}
\author{
Simon Scheiter \email{simon.scheiter@senckenberg.de}
}
\keyword{internal}
