% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyLayers.R
\name{copyLayers}
\alias{copyLayers}
\title{Copy layers from one VegObject or Field to another}
\usage{
copyLayers(from, to, layer.names, new.layer.names = NULL,
  keep.all.to = TRUE, keep.all.from = TRUE, dec.places = NULL)
}
\arguments{
\item{from}{The Field/DataObject that the layers are to be copied from.}

\item{to}{The Field/DataObject that the layers are to be copied to.}

\item{layer.names}{The layers to be copied from the "from" argument}

\item{new.layer.names}{The new names that the layers should have in the 'to' object. Use this to avoid naming conflict whereby, for 
example, if a layer "Total" is copied to an object which already has a "Total" layer then they layers will be names "Total.x" and "Total.y".}

\item{keep.all.to}{Boolean, if set to FALSE, all points in the 'to' object which don't have corresponding points in the 'from' object are removed.}

\item{keep.all.from}{Boolean, if set to FALSE, all points in the 'from' object which don't have corresponding points in the 'to' object are removed.}

\item{dec.places}{Numeric, how many decimal places to rounds the coordinates to inorder to get a match.  Default is no rounding (value is NULL) and if dine for most regularing spaced grids.
But setting this can be useful to force matching of coordinates with many decimal places which may have lost a small amount of precision and so don't match exactly.}
}
\value{
A Field (or data.table) comprising the 'to' object with the new layers added
}
\description{
This function does not check the Lon, Lat and Year columns are identical.  Any points in the 'from' object which are not in the 'to' object are ignored, 
and any points in the 'to' object which don't have corresponding points in the 'from' object are assigned NA, unless keep.all.to is set to FALSE .
}
\details{
This function allows layers (reminder, they are implemented as columns in a data.table) to be copied from one Field/DataObject to another.  
This is particularly useful for colouring or facetting a plot of one variable by another one say.  To give a more concrete example, one could use a biome classification 
to split (facet) a data-vs-model scatter plot.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
