% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-spatial.R
\name{correct.map.offset}
\alias{correct.map.offset}
\title{Fixes a spatial lines object where some of eastern Russia transposed to the other side of the world}
\usage{
correct.map.offset(spl)
}
\arguments{
\item{spl}{SpatialLines object to fix}
}
\value{
a the SpatialLines object
}
\description{
Fixes a spatial lines object where some of eastern Russia transposed to the other side of the world
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

