% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tools.R
\name{countCategoricalData}
\alias{countCategoricalData}
\title{Aggregate categorical data to a larger resolution by counting the number of small gridcells in a larger gridcell.}
\usage{
countCategoricalData(original.data, output.raster, categories = NULL)
}
\arguments{
\item{original.data}{A raster containing the original categorical data (integer values)}

\item{output.raster}{A raster used to specify the resolution and extent of the returned data.}

\item{categories}{A numeric vector of the integer codes defining the categories in the original data.
Can be calculated automatically if not supplied, but can be very slow so it is recommended to supply this meta data 
for large input rasters.

Uncertainty in this method will be zero if the gridlines in the larger output raster lie exactly on the gridlines of the smaller input raster.
Otherwise it will be low if resolution oforiginal.data is very much lower than the output.raster, and will increase 
This is because the gridcells in the original raster which stradle two output grid cells are placed into a gridcell in the 
closer output raster cell, not divided with weighted contributions to each}
}
\value{
A RasterStack recording (one layer for each category in the original data) how many of the original data gridcells with a particular
fell into each output raster gridcell,
}
\description{
Given a categorical raster (with \emph{n} factors) and a lower resolution template, return a RasterStack with the \emph{n} layers, 
where the value in each cell of the \emph{n} layers in the number of original data gridcells with that value folling in that high resolution
gridcell.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

