% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-classes.R
\docType{class}
\name{DataObject-class}
\alias{DataObject-class}
\title{Contains one dataset (ie. measured, not model-generated).}
\description{
A key class of the package.  A \code{DataObject} stores the data and metadata for one dataset (temporal or spatial).
For example
}
\details{
Generally these are not created directly by the user, but rather by functions like \code{getMODISTreecover} or \code{getSaatchi2011} or so, but the use can create them directly if desired.
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular DataObject.  Recommended to be alphanumeric because it is used to construct file names.}

\item{\code{name}}{A character string to more fully describe this DataObject}

\item{\code{data}}{A data.table object.  This is used because is it very much faster for calculations that data.frame or raster layers.}

\item{\code{quant}}{A Quantity object to define what output this DataObject contains}

\item{\code{spatial.extent}}{A SpatialExtent object which describes the area covered by this DataObject.  Particularly useful if the data has been spatially averaged.}

\item{\code{temporal.extent}}{A TemporalExtent object which describes the time periog covered by this DataObject.  Particularly useful if the data has been temporally averaged.}

\item{\code{correction.layer}}{A character string defining a multiplicative corection layer that can be used for the data set}

\item{\code{comparisons}}{A list of SpatialComparison objects, one for each ModelObject which has been compared to this DataObject}
}}

\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
