% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-classes.R
\docType{class}
\name{DatasetInfo-class}
\alias{DatasetInfo-class}
\title{Class to hold the metadata for a dataset}
\description{
This class describes a vegetation run, including its location on disk, the model used, the PFT set used, an unique id and a description, offsets to apply to the longitudes and latitudes to make the co-rordinates gridcell centered and so on.
It is not primarily intended to be used by itself. Instead it is inherited by \code{Source} object (due to this inheritance the slots can be accessed directly)
and included in a \code{Field} in the \code{run} slot (not inherited, so needs to be access be \code{@source}).
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular model un.  Recommended to be alphanumeric because it is used to construct file names. (Mandatory)}

\item{\code{name}}{A character string giving the name of this dataset (eg MODIS GPP). (Mandatory)}

\item{\code{quant}}{A Quantity object describing the physical quantity thatthe dataset describes}

\item{\code{fill.col}}{A string to define an R colour used when plotting this run as a histogram or scatter plot or so}

\item{\code{line.col}}{A string to define an R colour used when plotting this runs as a line graph}

\item{\code{line.width}}{A numeric to define the width of a line representing this model run}

\item{\code{line.type}}{A numeric to define the line style representing this model run}
}}

\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
