% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineLayer.R
\name{defineLayer}
\alias{defineLayer}
\title{Define a Layer object}
\usage{
defineLayer(id, name, colour, properties = list(type = "Undefined"),
  add.to)
}
\arguments{
\item{id}{A unique character string to identify this particular data.  Recommended to be alphanumeric because it is used to construct file names and to use underscores and full stops (periods) 
for separating characters where need be.   For example, "Standard_v4.0" or "Saatchi2011.HD"  (can be derived from \code{name} if a \code{name} is supplied, otherwise mandatory)}

\item{format}{A Format object to describe the type of source.  Can be GUESS, aDGVM, aDGVM2 or DGVMData (note no quotes since these are actual R objects not strings).  (Mandatory) DEPRECATED: alternatively a character string to identify the format of the files of this data sorce. This can be anything, but should be in order for "getField()" to work corrected 
MF TODO UPDATE Note that that \code{id}, \code{name}, \code{format}, and \code{dir} are compulsory, the rest will be filled with dummy/default values if left blank.
Take care with \code{lon.lat.offset} and \code{year.offset} which are initialised to 0 which is unsuitable for some LPJ-GUESS configurations.}
}
\value{
A Layer object including metadata defined by empty data slots
}
\description{
This function is  \linkS4class{Layer} represents a sub-component or one particular aspect of a \linkS4class{Field} object, 
for example one PFT or one carbon pool.
}
\details{
This function is preferred to a \code{new("Layer",...)} initialisation because it does both the initialisation 
and also optionally adds the Field to to a \linkS4class{Format} object with the \code{add.to} argument.

Note that no actual data is stored in the resultant \linkS4class{Layer}, only metadata.
}
\seealso{
Layer
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
