% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{defineModelRun}
\alias{defineModelRun}
\title{Define a ModelRun object, setting up all the required metadata (only).}
\usage{
defineModelRun(...)
}
\arguments{
\item{...}{The parameters are the slots of a ModelRunInfo object. Note that that \code{id} and \code{run.dir} are compulsory, the rest will be filled with dummy/default values if left blank.
compulsory.  Take care with \code{lon.lat.offset} and \code{year.offset} which are initialised to 0 which is unsuitable for some LPJ-GUESS confgurations.}
}
\value{
A ModelRun object, with metadata defined by empty data slots.s
}
\description{
This function is preferred to a \code{new("ModelRunInfo",...)} initialisation followed by  \code{new("ModelRun",...)} initialisation because it does both intialisations in one step and also performs some extra checks and initialisations of defaults.
}
\details{
Note that initialliy, no actual data from the run is stored, only metadata. The data, stored as ModelObjects and added to a ModelRun object, are added built with later with other commands and added to \code{ModelRun} command using \code{addToModelRun}
}
\seealso{
ModelRun, ModelRunInfo
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
