% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determinePFTs.R
\name{determinePFTs}
\alias{determinePFTs}
\title{Get PFTs present}
\usage{
determinePFTs(x, ...)
}
\arguments{
\item{x}{The Source object for which to determine the PFTs present}

\item{...}{Extra arguments to the format-specific functions that this function wraps around.}
}
\description{
Get PFTs present
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
