% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-aDGVM.R
\name{determinePFTs_aDGVM}
\alias{determinePFTs_aDGVM}
\title{Detemine PFTs present in an aDGVM2; currently only returns a warning.}
\usage{
determinePFTs_aDGVM(x, variables)
}
\arguments{
\item{x}{A Source objects describing a DGVMData source}

\item{variables}{Some variable to look for to detremine the PFTs present in the run.  Not the function automatically searches:
"lai", "cmass", "dens" and "fpc".  If they are not in your output you should define another per-PFT variable here.  Currently ignored.}
}
\description{
Detemine PFTs present in an aDGVM2; currently only returns a warning.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
