% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-aDGVM1.R
\name{determinePFTs_aDGVM1}
\alias{determinePFTs_aDGVM1}
\title{Detemine PFTs present in an aDGVM1 run}
\usage{
determinePFTs_aDGVM1(x, variables)
}
\arguments{
\item{x}{A Source objects describing an aDGVM1(-SPITFIRE) run}

\item{variables}{Some variable to loom for to detremine the PFTs present in the run.  Not the function automatically searches:
"lai", "cmass", "dens" and "fpc".  If they are not in your output you should define another per-PFT variable here.}
}
\description{
Detemine PFTs present in an aDGVM1 run
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
