% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-GUESS.R
\name{determinePFTs_GUESS}
\alias{determinePFTs_GUESS}
\title{Detemine PFTs present in an LPJ-GUESS run}
\usage{
determinePFTs_GUESS(x, variables)
}
\arguments{
\item{x}{A Source objects describing an LPJ-GUESS(-SPITFIRE) run}

\item{variables}{Some variable to loom for to detremine the PFTs present in the run.  Not the function automatically searches:
"lai", "cmass", "dens" and "fpc".  If they are not in your output you should define another per-PFT variable here.}
}
\description{
Detemine PFTs present in an LPJ-GUESS run
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
