% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-PEcAn.R
\name{determinePFTs_PEcAn}
\alias{determinePFTs_PEcAn}
\title{Detemine PFTs present in an PEcAn source}
\usage{
determinePFTs_PEcAn(x, variables)
}
\arguments{
\item{x}{A Source objects describing a PEcAn source}

\item{variables}{Some variable to look for to detremine the PFTs present in the run.  Not the function automatically searches:
"lai", "cmass", "dens" and "fpc".  If they are not in your output you should define another per-PFT variable here.  Currently ignored.}
}
\description{
Detemine PFTs present in an PEcAn source
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
