% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-FireMIP.R
\name{determineQuantities_FireMIP}
\alias{determineQuantities_FireMIP}
\title{List all LPJ-GUESS *.out files in a source directory}
\usage{
determineQuantities_FireMIP(source)
}
\arguments{
\item{source}{A path to a directory on the file system containing some .out files}
}
\value{
A list of all the .out files present, with the ".out" removed.
}
\description{
Simply lists all LPJ-GUESS output variables (stored as .out files) available in a directory. 
Also ignores some common red herrings like "guess.out" and "*.out"
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
