% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-ggplot2.R
\name{dgvm.ggplot.theme}
\alias{dgvm.ggplot.theme}
\title{ggplot2 themes}
\usage{
dgvm.ggplot.theme(x)
}
\arguments{
\item{x}{string: "temporal", "spatial", or "scatter"}
}
\value{
theme list
}
\description{
Get the internal ggplot2 theme definitions
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}
