% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-benchmarking.R
\name{doKappa}
\alias{doKappa}
\title{Kappa comparison between two biome rasters}
\usage{
doKappa(dt, id, labels = NULL, verbose = TRUE)
}
\arguments{
\item{dt}{A two-columned data.table containing the biomes (or other categorical data) represented as integer codes}

\item{id}{A character string to identify this comparison, typically the id of the biome scheme.}

\item{labels}{A vector of character strings to describe the categories over which Kappa is compared (typically a list of biomes)
following the order of the integer codes used in the data (see \code{stack} argument)}

\item{verbose}{A logical, if TRUE print out all the Kappa scores}
}
\value{
A Biome comparison object
}
\description{
Calculates a BiomeComparison object (which contains the Cohen's Kappa scores) given a stack containing two biome maps.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

