% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{doSpatialAverage}
\alias{doSpatialAverage}
\title{Spatially average a ModelObject, DataObject or data.table}
\usage{
doSpatialAverage(input.obj, verbose = FALSE, area.weighted = TRUE)
}
\arguments{
\item{input.obj}{A ModelObject, DataObject or data.table to be averaged}

\item{verbose}{If TRUE give some progress update about the averaging.}

\item{area.weighted}{If TRUE area-weight the gridcells}
}
\value{
A data.table
}
\description{
Spatially average all gridcells of a ModelObject, DataObject or data.table (assuming the data.table has columns "Lon" and "Lat"). 
Can use area-weighting to take into account the difference different areas of gridcells (even though they are constant in Lon,Lat)
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}

