% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-processing.R
\name{doTemporalAverage}
\alias{doTemporalAverage}
\title{Time average a data.table}
\usage{
doTemporalAverage(input.dt, verbose = FALSE)
}
\arguments{
\item{input.dt}{data.table}

\item{verbose}{If TRUE give some progress update about the averaging.}
}
\value{
A data.table
}
\description{
Time average all availables years (denoted by column "Years") of a data.table object.  This function does not select the years, that should be done first.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{export}
