% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object-processing.R
\name{doTemporalAverage.uncompiled}
\alias{doTemporalAverage.uncompiled}
\title{Time average a data.table}
\usage{
doTemporalAverage.uncompiled(input.dt, verbose = FALSE)
}
\arguments{
\item{input.dt}{data.table}

\item{verbose}{If TRUE give some progress update about the averaging.}
}
\value{
A data.table
}
\description{
Time average all availables years (denoted by column "Years") or a data.table object
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
