% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-utility-functions.R
\name{equivalentQuantities}
\alias{equivalentQuantities}
\title{Check if two Quantity objects are equivalent}
\usage{
equivalentQuantities(quant1, quant2)
}
\arguments{
\item{quant1}{First Quantity to compare}

\item{quant2}{Second Quantity to compare}
}
\description{
Performs a simple check on the units and types of two quantities
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}
\keyword{internal}
