% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpj-pfts.R
\docType{data}
\name{euro.PFTs}
\alias{euro.PFTs}
\title{LPJ-GUESS European PFTs}
\format{A list of \code{PFT} objects that store meta-data for commonly used European PFTs from LPJ-GUESS.}
\usage{
euro.PFTs
}
\description{
The list contain the standard LPJ-GUESS european PFTs (see Hickler at al. 2012) and few more.  It includes some metadata 
and a preferred plot colour.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\seealso{
\code{PFT}-class
}
\keyword{datasets}

