% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{expandLayers}
\alias{expandLayers}
\title{Expand plotting/processing layers from convenient short hand codes.}
\usage{
expandLayers(layers, input.data, PFT.set = NULL, type = "unknown",
  include.woody = TRUE)
}
\arguments{
\item{layers}{Character vector of layers to expand}

\item{input.data}{The data to which the layers will be applied, may be a \code{ModelObject}, a Raster* object, data.table or data.frame, a Spatial*DataFrame.}

\item{PFT.set}{List of a superset of PFTs that might be found in this run (only necessary if *not* passing a \code{ModelObject}).}

\item{type}{Character string identifying if this is a monthly  (= "monthly") or per PFT (="pft") variable.  Ignored for \code{ModelObjects} which supply this data themselves.
The function attempts to determine this argument if it is not provided.}

\item{include.woody}{If TRUE and "lifeform" is included in the layer list "Woody" is also returned}
}
\value{
A list of layers
}
\description{
Expands characters strings with particular meaning (eg. "lifefoms", "pfts", "seasons", etc...).  
This allows the user to simply specify, for example, "lifeforms" to a plotting command instead of c("Tree","Grass","Shrub") when plotting or processing.
}
\details{
Supported short-hand strings are:
\itemize{
  \item{"seasons"}{ expands to c("DJF","MAM","JJA","SON")}
  \item{"PFTs"}{ expands to a vector of each of the ids of PFTs present in the data (requires the \code{PFT.set} argument).}
  \item{"lifeforms"}{ expands all the different PFT lifeforms present as determined by the lifeform slots of the \code{PFTs} argument (and optionally "Woody").  For example c("Tree", "Grass", "Woody")}
  \item{"zones"}{ expands all the different PFT climactic zones present as determined by the zone slots of the \code{PFTs} argument.  For example c("Temperature", "Boreal", "Tropical")}
  \item{"leafforms"}{ expands all the different PFT leaf forms present as determined by the leafforms slot of the \code{PFTs} argument.  For example c("Needleleved", "Broadleaved")}
  \item{"phenologies"}{ expands all the different PFT leaf forms present as determined by the phenology slot of the \code{PFTs} argument.  For example c("Evergreen", "Summergreen", "Raingreen", "GrassPhenology)}
}
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

