% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent-methods.R
\name{extent-methods}
\alias{extent-methods}
\alias{extent}
\alias{extent,data.table-method}
\alias{extent,Field-method}
\alias{extent,Comparison-method}
\title{Extent methods}
\usage{
\S4method{extent}{data.table}(x)

\S4method{extent}{Field}(x)

\S4method{extent}{Comparison}(x)
}
\arguments{
\item{x}{A Field, Comaprison or a data.table.}
}
\value{
A raster::extent
}
\description{
Methods for returning a raster::extent object from a Field, Comparison or a data.table.  The data.table, or the data.table in the "data" slot of the Field and Comparison, 
must have columns named "Lon" and "Lat" otherwise these methods fail.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
