% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent-methods.R
\docType{methods}
\name{extent}
\alias{extent}
\alias{extent,DataObject-method}
\alias{extent,ModelObject-method}
\alias{extent,data.table-method}
\title{Extent methods}
\usage{
extent(x, ...)

\S4method{extent}{data.table}(x)

\S4method{extent}{ModelObject}(x)

\S4method{extent}{DataObject}(x)
}
\arguments{
\item{x}{A ModelObject, DataObject or a data.table DGVMTools object}

\item{...}{Other arguments, not currently used}
}
\value{
A raster::extent
}
\description{
Methods for returning a raster::extent object from a ModelObject, DataObject or a data.table.  The data.table, or the data.table in the "data" slot of the ModelObject and DataObject, 
must have columns named "Lon" and "Lat" otherwise these methods fail.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.table}: Returns a raster::extent object from a data.table.

\item \code{ModelObject}: Returns a raster::extent object from a ModelObject.

\item \code{DataObject}: Returns a raster::extent object from a DataObject.
}}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

