% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000rvc-tools.R
\name{extentFromDT}
\alias{extentFromDT}
\title{Extract the extent from a data.table}
\usage{
extentFromDT(this.dt)
}
\arguments{
\item{this.dt}{A data.table (which is assumed to have columns "Lon" and "Lat)

Note that it guesses the gridcell spacing.}
}
\value{
A raster::extent object
}
\description{
Returns an raster::extent object based on the "Lon" and "Lat" columns of the data.table
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

