% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridarea.R
\name{extract.seq}
\alias{extract.seq}
\title{Extract a regular sequence from unique values in a vector}
\usage{
extract.seq(x, force.regular = FALSE, descending = FALSE)
}
\arguments{
\item{x}{vetor of values (e.g. longitudes or latitudes)}

\item{force.regular}{force a regular spaced vector with the smallest distance.}

\item{descending}{sort in descending order if TRUE (default: ascendind)}
}
\value{
data.frame of gridcells with columns c("Lon", "Lat", "area")
}
\description{
Returns, if possible, an regular spaced sorted vector vector. If more than half of the unique values have unequal distances, the sorted values are returned.
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}
