% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop-methods.R
\name{extractRasterExtent}
\alias{extractRasterExtent}
\title{Get an extent from anything}
\usage{
extractRasterExtent(y)
}
\arguments{
\item{y}{Something to crop to  (ie the \code{y} argument of a crop command)

TODO One day this should be converted to terra::SpatExtent.  Probably whenever raster is really deprecated and shall be purged from DGVMTools.}
}
\value{
A raster::Extent object
}
\description{
Internal function to try to get a raster::Extent object from any potential y argument to a \code{crop} command.
}
\keyword{internal}
