% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-methods.R
\name{FieldToArray}
\alias{FieldToArray}
\title{Convert a Field to Arrays}
\usage{
FieldToArray(
  x,
  start.date = NULL,
  calendar = "365_day",
  add.missing.cols = TRUE,
  add.missing.rows = TRUE,
  global.extent = FALSE,
  cname = FALSE,
  invertlat = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{x}{the data.table of a \code{\linkS4class{Field}}}

\item{add.missing.cols}{logical, if TRUE fill empty longitude columns (should be regularly spaced or you'll get silliness), default is TRUE}

\item{add.missing.rows}{logical, if TRUE fill empty latitude rows (should be regularly spaced or you'll get silliness), default is TRUE}

\item{global.extent}{logical, if TRUE extend the array to be the entire global extent,  default is FALSE}

\item{cname}{the column name to convert, if not set a list is returned}

\item{invertlat}{start in the north}

\item{verbose}{print some information}
}
\value{
an array or a list or arrays
}
\description{
Convert a Field (or the data.table from a Field) into a list of multi-deminsional arrays (indexed Lon, Lat and Time),
with one array in the list per Layer in the input Field.
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}, Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
