% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FireMIP.R
\docType{data}
\name{FireMIP.PFTs}
\alias{FireMIP.PFTs}
\title{LPJ-GUESS(-SPITFIRE) Global PFTs}
\format{A list of \code{PFT} objects that store meta-data for commonly used PFTs from LPJ-GUESS (and LPJ-GUESS-SPITFIRE)}
\usage{
FireMIP.PFTs
}
\description{
The list contain the standard LPJ-GUESS global PFTs and few more.  It includes some metadata 
and a preferred plot colour.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\seealso{
\code{PFT}-class
}
\keyword{datasets}

