% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome-schemes.R
\name{FireMIPBiomeRules}
\alias{FireMIPBiomeRules}
\title{Rules to classify FireMIP biomes from fractionalcover of PFTs}
\usage{
FireMIPBiomeRules(x)
}
\arguments{
\item{x}{Numerical vector of FPC values for a particular location. 
Certain fractions and quantities should have been pre-calculated.}
}
\value{
Biomes code (1-10)
}
\description{
Unpublished but possibly useful in principle.  Simpler classes that Smith et al 2014,
but more complex that Forrest et al 2015.
}
\keyword{internal}
