% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome-schemes.R
\name{Forrest2015MegaBiomeRules}
\alias{Forrest2015MegaBiomeRules}
\title{Rules to classify coarses "mega biomes" as per Forrest et al. 2015}
\usage{
Forrest2015MegaBiomeRules(lai)
}
\arguments{
\item{lai}{Numerical vector of LAI values for a particular location. 
Certain fractions and quantities should have been pre-calculated.}
}
\value{
Biomes code (1-8, ordering as in the Forrest et al. 2015 figure)
}
\description{
# This is described in Forrest et al. 2015 Climates of the Past.  Basically it calculates the Smith et al. 2014 biomes, 
but assigns them to broader categories as per Harrison and Prentice 2006
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
