% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome-schemes.R
\docType{data}
\name{FPCMegabiomes.scheme}
\alias{FPCMegabiomes.scheme}
\title{Meta-data for an FPC-based mega biome scheme}
\format{An object of class \code{BiomeScheme} of length 1.}
\usage{
FPCMegabiomes.scheme
}
\description{
Joerg stuff here
}
\keyword{internal}
