% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-import.R
\name{getAnnualClimate}
\alias{getAnnualClimate}
\title{Get Annual values of climate drivers}
\usage{
getAnnualClimate(run = NA, file = NA, operation = "mean",
  temporal.extent = NA, spatial.extent = NA, full = FALSE,
  data.name = NA, lon.name = NA, lat.name = NA, time.name = NA,
  forceReCalculation = FALSE, write = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{run}{a VegRun object, if a VegSpatial object should be returned, otherwise NA}

\item{file}{path to the netCDF file}

\item{operation}{Any of sum, mean, gdd5}

\item{temporal.extent}{Either a TemporalExtent object, a vector of years (min/max are choosen) or 'NA' (whole netCDF timespan is used)}

\item{spatial.extent}{Either a SpatialExtent object of a raster extent}

\item{full}{should the full annual dataset be returned or an multiannual average (default).}

\item{data.name}{Name of the data field in the netcdf file, if NA it will be guessed}

\item{lon.name}{Name of the longitude vector in the netcdf file, if NA it will be guessed}

\item{lat.name}{Name of the latitude vector in the netcdf file, if NA it will be guessed}

\item{time.name}{Name of the time vector in the netcdf file, if NA it will be guessed}

\item{forceReCalculation}{Recalculate the desired values (default: FALSE)}

\item{write}{write the calculated values for fture speedup (default: TRUE)}

\item{verbose}{print some messages}

\item{...}{further so far ignored parameters}
}
\value{
data.table if no VegRun was given, otherwise a VegObject
}
\description{
Get Annual values of climate drivers from netCDF files, files must be CF-compliant
}
\examples{
message("See templates/Example.ggplot.R")
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

