% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Baccini2012.R
\name{getBaccini2012}
\alias{getBaccini2012}
\title{Read Baccini et al. 2012 total biomass map}
\usage{
getBaccini2012(location = "/data/forrest/Biomass/Baccini2012",
  resolution = "original")
}
\arguments{
\item{location}{A character string specifying the the location of the original data files}

\item{resolution}{A character string specifying the resolution, currently supported
are "HD" and "T63" and "original"

Should be moved to DGVMDatasets, not distributed with DGVMTools}
}
\value{
A \code{SpatialDataset} object.
}
\description{
Total biomass across the tropics.  
Original dataset is 1km, also provided here at 0.5 degrees or T63.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

