% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Crowther2015.R
\name{getCrowther2015}
\alias{getCrowther2015}
\title{Read the tree count/tree density of Crwother et al. (2015)}
\usage{
getCrowther2015(location = "", resolution = "HD")
}
\arguments{
\item{location}{path to the files}

\item{resolution}{string of resolutions. So far 'original' (tree count, very large 1km^2!) and 'HD' (trees/ha, 0.5deg) are available.}
}
\value{
A \code{DataObject}
}
\description{
Read the tree count/tree density of Crwother et al. (2015)
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}
