% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDimInfo.R
\name{getDimInfo}
\alias{getDimInfo}
\title{Returns the spatio-temporal information}
\usage{
getDimInfo(x, info = "names")
}
\arguments{
\item{x}{A DataObject, Field, Comparison or data.tables}

\item{info}{A character string to define what spatio-temporal info you want.  Can be:

\describe{

 \item{"names"}{to give the names of the spatio-temporal dimensions (a character vector)}
 \item{"range"}{to give the range of each dimension (list)}
 \item{"size"}{to give the number of points in a dimension}
 \item{"values"}{to give the actual value for each point in a dimension}
 \item{"full"}{to return a data.table with the full spatio-temporal information}

}}
}
\value{
See above
}
\description{
This function returns information about the spacio-temporal dimensions of a DataObject, Field, Comparison or data.table
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
