% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-runtools.R
\name{getExtentFromDT}
\alias{getExtentFromDT}
\title{Get the spatial extent from a data.table}
\usage{
getExtentFromDT(dt)
}
\arguments{
\item{dt}{A data.table}
}
\value{
Either a raster::extent (for a grid of points) or a numerical vector (for a single site)
}
\description{
Returns either a raster::extent (for a grid of points) or a numerical vector (for a single site) from a data.table 
assuming column name "Lon" and "Lat" are present.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

