% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-aDGVM1.R
\name{getField_aDGVM1}
\alias{getField_aDGVM1}
\title{Get a Field for aDGVM1}
\usage{
getField_aDGVM1(source, quant, target.STAInfo, file.name, verbose, ...)
}
\arguments{
\item{source}{A \code{Source} containing the meta-data about the aDGVM1 run}

\item{quant}{A string the define what output file from the aDGVM1 run to open, for example "anpp" opens and read the "anpp.out" file}

\item{target.STAInfo}{The spatial-temporal target domain}

\item{file.name}{Character string holding the name of the file.  This can be left blank, in which case the file name is just taken to be 
"<quant@id>.out" (also "<quant@id>.out.gz")}

\item{verbose}{A logical, set to true to give progress/debug information}
}
\value{
A list containing firstly the data.tabel containing the data, and secondly the STA.info
}
\description{
An internal function that reads data from an aDGVM1 run.  It actually call one of three other functions depending on the type of quantity specified.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
