% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-aDGVM2.R
\name{getField_aDGVM2}
\alias{getField_aDGVM2}
\title{Get a Field for aDGVM2}
\usage{
getField_aDGVM2(
  source,
  quant,
  layers = NULL,
  target.STAInfo,
  file.name,
  verbose,
  adgvm2.scheme,
  adgvm2.daily
)
}
\arguments{
\item{source}{A \code{source} containing the meta-data about the aDGVM2 run}

\item{quant}{A Quantity object to define what quantity from the aDGVM2 run to extract}

\item{layers}{Ignored for aDGVM}

\item{target.STAInfo}{STAInfo object specifying the spatial-temporal-annual extent required.  Note that at this stage only the years are selected/}

\item{file.name}{Character string holding the name of the file.  This can be left blank, in which case the file name is automatically generated}

\item{verbose}{A logical, set to true to give progress/debug information}

\item{adgvm2.scheme}{A number that defines if pop-files (=1) or trait-files (=2) are used.}

\item{adgvm2.daily}{A logical, set to true to read daily data (only for \code{adgvm2.scheme=1} and if daily data are provided in pop file)}
}
\value{
A list containing firstly the data.table containing the data, and secondly the STA.info
}
\description{
An internal function that reads data from an aDGVM2 run. It actually calls one of two other functions depending on the type of quantity specified.
}
\seealso{
\code{\link{getQuantity_aDGVM2_Scheme1}, \link{getQuantity_aDGVM2_Scheme2}}
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
