% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-FireMIP.R
\name{getField_FireMIP}
\alias{getField_FireMIP}
\title{Get a Field for FireMIP}
\usage{
getField_FireMIP(source, quant, target.STAInfo, verbose)
}
\arguments{
\item{verbose}{A logical, set to true to give progress/debug information}

\item{run}{A \code{Source} containing the meta-data about the LPJ-GUESS run}

\item{variable}{A string the define what output file from the LPJ-GUESS run to open, for example "anpp" opens and read the "anpp.out" file}

\item{first.year}{The first year (as a numeric) of the data to be returned}

\item{last.year}{The last year (as a numeric) of the data to be returned}
}
\value{
A list containing firstly the data.tabel containing the data, and secondly the STA.info
}
\description{
An internal function that reads data from an FireMIP run.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
