% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-NetCDF.R
\name{getField_NetCDF}
\alias{getField_NetCDF}
\title{Get a Field for NetCDF}
\usage{
getField_NetCDF(
  source,
  quant,
  layers = NULL,
  target.STAInfo,
  file.name,
  verbose = FALSE,
  nc.verbose = FALSE,
  vars.to.ignore = c("time_bnds", "lon_bnds", "lat_bnds", "lat", "lon"),
  calendar = "standard"
)
}
\arguments{
\item{source}{A \linkS4class{Source} containing the meta-data about the NetCDF source}

\item{quant}{A Quantity object to specify what quantity should be opened.}

\item{layers}{A character string (or a vector of character strings) specifying which variables from the NetCDF file are to be read.
NULL (default) means read all.}

\item{target.STAInfo}{An STAInfo object defining the spatial-temporal-annual extent over which we want the data}

\item{file.name}{Character string holding the name of the file.  This can be left blank, in which case the file name is automatically generated}

\item{verbose}{A logical, set to true to give progress/debug information}

\item{nc.verbose}{A logical, set to true to give progress/debug information from the ncdf4 package functions.  This can be a lot, 
so it is handy to control that separately.}

\item{vars.to.ignore}{A list of character strings specifying which cvariables *not* to read.  These are typically metadata about the coordinate data
(hence the defaults ""time_bnds", "lon_bnds", "lat_bnds", "lat" and "lon"), but you can also use this argument to ignore a particular data variable from a file.}

\item{calendar}{Character string, sometimes the calendar string on the time axis can be incorrect or missing.  Here you can manually provide it.
Note: A common error in paleo files is "standard" instead of "proleptic_gregorian". Specifically, if you have dates with years 1582 
(the start of the Gregorian calendar) and it includes leap years the calendar needs to be set to "proleptic_gregorian".}
}
\value{
A list containing firstly the data.table containing the data, and secondly the STAInfo for the data that we have
}
\description{
An internal function that reads data from an NetCDF .nc file.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
