% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-PEcAn.R
\name{getField_PEcAn}
\alias{getField_PEcAn}
\title{Get a Field for PEcAn}
\usage{
getField_PEcAn(source, quant, target.STAInfo, file.name, runid,
  verbose = FALSE)
}
\arguments{
\item{source}{A \code{Source} containing the meta-data about the PEcAn source}

\item{quant}{A Quantity object to specify what quantity should be opened.}

\item{file.name}{Character string holding the name of the file.  This can be left blank, in which case the file name is automatically generated}

\item{runid}{The PEcAn run id (as a string) to choose the relevant run from the ensemble.
From the original PEcAn documentation:
"Custom character vector of full paths to NetCDF
  files. If `NULL` (default), this list is constructed
  automatically by looking for `YYYY.nc` files in
  `file.path(outdir, runid)`."}

\item{verbose}{A logical, set to true to give progress/debug information}

\item{target.sta.info}{An STAInfo object defining the spatial-temporal-annual extent over which we want the data}
}
\value{
A Field with correct STAInfo idefined.
}
\description{
An internal function that reads data from an PEcAn .nc file.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
