% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HandPBiomes.R
\name{getHandPBiomes}
\alias{getHandPBiomes}
\title{Read HandPBiomes et al. 2015 total biomass map}
\usage{
getHandPBiomes(location = "/data/forrest/Biomes/", resolution = "HD",
  classification = "Smith2014")
}
\arguments{
\item{location}{A character string specifying the the location of the original data files}

\item{resolution}{A character string specifying the resolution, currently supported
are "HD" and "T63" and "original"

Should be moved to DGVMDatasets, not distributed with DGVMTools}
}
\value{
A \code{DataObject} object.
}
\description{
Total biomass across the tropics.  
Original dataset is 1km, also provided here at 0.5 degrees or T63.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
