% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOD17A3.R
\name{getMOD17A3}
\alias{getMOD17A3}
\title{Read MODIS GPP map}
\usage{
getMOD17A3(location = "/data/forrest/Productivity/MODIS/V55/processed/",
  resolution = "original", average = TRUE, var)
}
\arguments{
\item{location}{A character string specifying the the location of the original data files}

\item{resolution}{A character string specifying the resolution, currently supported
are "1km-Turkey"}

\item{average}{Logical, if true return average of all years, if false return all years}

\item{var}{A character string with value either "npp" for net primary productivity, or "gpp" for gross primary productivity.

Should be moved to DGVMDatasets, not distributed with DGVMTools}
}
\value{
A \code{DataObject} object.
}
\description{
Blah
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
