% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODISGPP.R
\name{getMODISGPP}
\alias{getMODISGPP}
\title{Read MODIS GPP map}
\usage{
getMODISGPP(location = "/data/forrest/Productivity/MODIS/V55",
  resolution = "original", average = TRUE)
}
\arguments{
\item{location}{A character string specifying the the location of the original data files}

\item{resolution}{A character string specifying the resolution, currently supported
are "1km-Turkey"}

\item{average}{Logical, if true return average of all years, if false return all years

Should be moved to DGVMDatasets, not distributed with DGVMTools}
}
\value{
A \code{DataObject} object.
}
\description{
Blah
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

