% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPFTs.R
\name{getPFTs}
\alias{getPFTs}
\title{Get the PFTs present in a run}
\usage{
getPFTs(input, PFT.data)
}
\arguments{
\item{input}{The data from which to retrieve the PFTs present (data.table, Field, or Raster*-object)}

\item{PFT.data}{A list of PFTs which might be present (ie. a superset of those actually present)}
}
\value{
A list of PFT object which are actually present
}
\description{
Extract the PFTs present in some data, given the set of possible PFTs.  The data can be represented as a data.table, Field, or Raster*-object.  The 
It does thsi based on the names of the layers/columns of the data
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
