% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-aDGVM2.R
\name{getQuantity_aDGVM2_Scheme1}
\alias{getQuantity_aDGVM2_Scheme1}
\title{Read aDGVM2 quantities from pop file}
\usage{
getQuantity_aDGVM2_Scheme1(
  run,
  variable,
  target.sta,
  file.name = file.name,
  adgvm2.daily
)
}
\arguments{
\item{run}{A  \code{\linkS4class{Source}} containing the meta-data about the aDGVM2 run}

\item{variable}{A character string specifying which variable/quantity to get, can be "agb“, "aGPP_std“, "basalarea“, "bgb“, "canopyheight_std“, "LAI_std“, meanheight“, "nind“, "pind“, "vegC_std“, "vegcover_std"}

\item{target.sta}{STAInfo object containing the space-time-annual extent required.}

\item{file.name}{Character string holding the name of the file.  This can be left blank, in which case the file name is automatically generated}
}
\description{
An internal function to read quantities from aDGVM2 pop files. Quantities are provided for trees, C4 grasses and C3 grasses. Trait files are not opened
}
\seealso{
\code{\link{getQuantity_aDGVM2_Scheme2}}
}
\author{
Simon Scheiter \email{simon.scheiter@senckenberg.de}, Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
