% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-aDGVM.R
\name{getQuantity_aDGVM_Scheme1}
\alias{getQuantity_aDGVM_Scheme1}
\title{An internal function to read quantities from aDGVM2 pop files. Quantities are provided for trees, C4 grasses and C3 grasses. Trait files are not opened}
\usage{
getQuantity_aDGVM_Scheme1(run, variable, first.year, last.year,
  adgvm.daily)
}
\arguments{
\item{run}{A \code{source}}

\item{variable}{A character string specifying which variable/quantity to get, can be "agb","nind","meanheight","basalarea","pind","firefreq","vegcover_std","vegC_std","LAI_std","aGPP_std","canopyheight_std","burntfraction_std"}

\item{first.year}{First year of data to read}

\item{last.year}{Last year of data to read}
}
\description{
An internal function to read quantities from aDGVM2 pop files. Quantities are provided for trees, C4 grasses and C3 grasses. Trait files are not opened
}
\seealso{
\code{getQuantity_aDGVM_Scheme2}
}
\author{
Simon Scheiter \email{simon.scheiter@senckenberg.de}, Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
