% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-aDGVM.R
\name{getQuantity_aDGVM_Scheme2}
\alias{getQuantity_aDGVM_Scheme2}
\title{Slightty more complex classification of yearly output for aDGVM}
\usage{
getQuantity_aDGVM_Scheme2(run, variable, first.year, last.year)
}
\arguments{
\item{run}{A \code{Source}}

\item{variable}{A character string specifying which variable to get, only "agb" currently supported}

\item{first.year}{First year of data to read (not currently used)}

\item{last.year}{Last year of data to read (not currently used)}
}
\description{
Original from Simon Scheiter, this version modified by M. Forrest to fit into the DGVM framework. 
Classifies as evergreen tree, deciduous tree, evergreen shrub, deciduous shrub, C3 grass, C4 grass.
}
\seealso{
\code{getQuantity_aDGVM_Scheme1}
}
\author{
Simon Scheiter \email{simon.scheiter@senckenberg.de}, Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
