% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Saatchi2011.R
\name{getSaatchi2011}
\alias{getSaatchi2011}
\title{Read Saatchi et al. 2015 total biomass map}
\usage{
getSaatchi2011(location = "/data/forrest/Biomass/", resolution = "HD")
}
\arguments{
\item{location}{A character string specifying the the location of the original data files}

\item{resolution}{A character string specifying the resolution, currently supported
are "HD" and "T63" and "original"

Should be moved to DGVMDatasets, not distributed with DGVMTools}
}
\value{
A \code{DataObject} object.
}
\description{
Total biomass across the tropics.  
Original dataset is 1km, also provided here at 0.5 degrees or T63.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
