% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-datasets.R
\name{getSaatchi2011_example}
\alias{getSaatchi2011_example}
\title{Read Saatchi et al. 2011 total biomass map}
\usage{
getSaatchi2011_example(resolution = "HD")
}
\arguments{
\item{resolution}{A character string specifying the resolution, currently supported
are "HD" and "T63"

This dataset has been used in various publications, but just because we include it here 
doesn't mean we specifically recommend it's use.  It is included here as an example, judgement of the scientific merit is left 
to the user.}
}
\value{
A \code{DataObject} object.
}
\description{
An example dataset of distribution total biomass across the tropics.  
Original dataset is 1km, provided here at 0.5 degrees or T63.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
