% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simard2011.R
\name{getSimard2011}
\alias{getSimard2011}
\title{Read dataset et al. 2015 total biomass map}
\usage{
getSimard2011(location = "/data/forrest/CanopyHeight/Simard2011/",
  resolution = "HD")
}
\arguments{
\item{location}{A character string specifying the the location of the original data files}

\item{resolution}{A character string specifying the resolution, currently supported
are "HD".

Should be moved to DGVMDatasets, not distributed with DGVMTools}
}
\value{
A \code{DataObject} object.
}
\description{
Global average top of canopy height
Original dataset is 1km, but provided here at 0.5 degrees.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

