% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-aDGVM.R
\name{getStandardQuantity_aDGVM}
\alias{getStandardQuantity_aDGVM}
\title{Returns the data from one aDGVM output variable as a \code{data.table}.}
\usage{
getStandardQuantity_aDGVM(
  run,
  quant,
  target.sta,
  file.name = file.name,
  adgvm.file.type,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{run}{A \code{\linkS4class{Source}} containing the meta-data about the aDGVM run from which the data is to be read.  Most importantly it must contain the run.dara nd the offsets.}

\item{quant}{A Quantity to define what output file from the aDGVM run to open}

\item{file.name}{An optional character string (or a list of character strings) holding the name of the file(s)
This can be left blank, in which case the file name is automatically generated.}

\item{verbose}{A logical, set to true to give progress/debug information}
}
\value{
a data.table (with the correct tear offset and lon-lat offsets applied)
}
\description{
This function can retrieve a 'Standard' vegetation quantity (returned as a data.table) with standard definition and units
to compare to other models and to data.  This must be implemented for each and every Standard quantity 
for each and every model to ensure completeness.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}

Glenn Moncrieff \email{glenn@saeon.ac.za}
}
\keyword{internal}
