% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-GUESS.R
\name{getStandardQuantity_LPJ}
\alias{getStandardQuantity_LPJ}
\title{Returns the data from one LPJ-GUESS output variable as a \code{data.table}.}
\usage{
getStandardQuantity_LPJ(
  run,
  quant,
  layers = NULL,
  target.sta,
  file.name = file.name,
  verbose = FALSE
)
}
\arguments{
\item{run}{A \code{\linkS4class{Source}} containing the meta-data about the LPJ-GUESS run from which the data is to be read.  Most importantly it must contain the run.dara nd the offsets.}

\item{quant}{A Quantity to define what output file from the LPJ-GUESS run to open}

\item{file.name}{Character string holding the name of the file.  This can be left blank, in which case the file name is just taken to be 
"<quant@id>.out" (also "<quant@id>.out.gz")}

\item{verbose}{A logical, set to true to give progress/debug information}
}
\value{
a data.table (with the correct tear offset and lon-lat offsets applied)
}
\description{
This function can retrieve a 'Standard' vegetation quantity (returned as a data.table) with standard definition and units
to compare to other models and to data.  This must be implemented for each and every Standard quantity 
for each and every model to to ensure completeness.
\code{\linkS4class{Source}} object, it will return that to save time.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
