% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{getSTInfo}
\alias{getSTInfo}
\title{Returns the spatio-temporal information}
\usage{
getSTInfo(x, info = "names")
}
\arguments{
\item{x}{A DataObject or ModelObject}

\item{info}{A character string to define what info you want.  Can be "names" to give the names of the spatio-temporal dimensions or ... (to be defined)}
}
\value{
A ModelObject or DataObject
}
\description{
This function returns information about the spacio-temporal dimensions of a DataObject, ModelObject or data.table
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

