% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thurner2013.R
\name{getThurner2013}
\alias{getThurner2013}
\title{Read Thurner et al. 2013 total biomass map}
\usage{
getThurner2013(location = "/data/forrest/Biomass/", resolution = "HD")
}
\arguments{
\item{location}{A character string specifying the the location of the original data files}

\item{resolution}{A character string specifying the resolution, currently supported
are "HD" and "T63" and "original"

Should be moved to DGVMDatasets, not distributed with DGVMTools}
}
\value{
A \code{DataObject} object.
}
\description{
Total biomass across the nothern temperate and boreal zones.
Original dataset is 1km, also provided here at 0.5 degrees or T63.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
