% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpj-runtools.R
\name{getVegQuantity_LPJ}
\alias{getVegQuantity_LPJ}
\title{Returns the data from one LPJ-GUESS output variable as a \code{data.table}.}
\usage{
getVegQuantity_LPJ(run, var.string, store.internally = FALSE,
  verbose = FALSE)
}
\arguments{
\item{run}{A \code{VegRun} containing the meta-data about the LPJ-GUESS run from which the data is to be read.  Most importantly it must contain the run.dara nd the offsets.}

\item{var.string}{A string the define what output file from the LPJ-GUESS run to open, for example "anpp" opens and read the "anpp.out" file}

\item{store.internally}{A logical defining whether to attach the resulting to the \code{data.table} to the \code{VegRun} object for later use}

\item{verbose}{A logical, set to true to give progress/debug information}
}
\value{
a data.table (with the correct tear offset and lon-lat offsets applied)
}
\description{
\code{getVegQuantity_LPJ} returns a \code{data.table} containing the full data (not averaged spatially or temporally) data from an LPJ-GUESS
output variable.  Normally it will read the file from disk, but if that has already been done, and the \code{data.table} has been saved to the 
\code{VegRun} object, it will return that to save time.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

