% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-utility-functions.R
\name{\%/0\%}
\alias{\%/0\%}
\title{Safe division}
\usage{
x \%/0\% y
}
\arguments{
\item{x}{numerator}

\item{y}{denominator

Handy little thing.}
}
\description{
Function to divide two number but return 0 if the denominator is 0
}
\keyword{internal}
