% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridarea.R
\name{gridarea1d}
\alias{gridarea1d}
\title{grid cell area along a vector of latitudes}
\usage{
gridarea1d(lat, dlon, scale = 1, ellipse = FALSE)
}
\arguments{
\item{lat}{vetor of latitudes}

\item{dlon}{longitudinal extent}

\item{scale}{multiplicator. If 1 (default) unit m^2}

\item{ellipse}{TRUE (polar and equatorial radius differ) or
FALSE (default, polar and equatorial radius are the same)}
}
\value{
vector of gridcell area is m^2
}
\description{
Returns the area in square meters along a vetor of latitudes by equal longitude distance.
}
\details{
The function returns a vector of area in square meters along a vector
of latitudes. These must not be of equal distance. However, for the longitude
will be equal along the given latitude vector. The latitude is assumed
to be the gridcell midpoint and the northern and southern edges are
calculated as half the distance to the next element in the latitude vector.
}
\examples{
dlon <- 0.5
lat <- seq(-89.75,89.75,0.5)
sum(gridarea1d(lat,dlon))*720*1.e-12
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}
\keyword{internal}
