% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome-schemes.R
\docType{data}
\name{Hickler2012.scheme}
\alias{Hickler2012.scheme}
\title{Meta-data describing the Hickler et al. 2012 european biome scheme for LPJ-GUESS output.}
\format{An object of class \code{BiomeScheme} of length 1.}
\usage{
Hickler2012.scheme
}
\description{
Note 'substution' is probably a redundant slot, this is a part of the data, 
not the classication of the model output.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}

