% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome-schemes.R
\name{Hickler2012Rules}
\alias{Hickler2012Rules}
\title{Rules to classify biomes as per Hickler et al. 2012}
\usage{
Hickler2012Rules(x)
}
\arguments{
\item{x}{Numerical vector of LAI values for a particular location. 
Certain fractions and quantities should have been pre-calculated.}
}
\value{
Biomes code (1-13, ordering as in Hickler et al 2012 figure)
}
\description{
Based on LAI, GDD5 and lon/lat only, see paper for details.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
