% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-misc.R
\name{is.DataObject}
\alias{is.DataObject}
\title{Check is an object is a \code{DataObject}.}
\usage{
is.DataObject(input, spatial = FALSE, temporal = FALSE, site = FALSE)
}
\arguments{
\item{input}{Any R object to bec checked}

\item{spatial}{check if input is a spatial object}

\item{temporal}{check if input is a temporal object}

\item{site}{check if input is a site object}
}
\value{
logical
}
\description{
Returns TRUE if an object is a \code{DataObject}, and FALSE otherwise
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

