% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equality-methods.R
\docType{methods}
\name{is.equal}
\alias{is.equal}
\alias{is.equal,Quantity,Quantity-method}
\title{Checks two DGVMTools metadata objects for equality}
\usage{
is.equal(a, b)

\S4method{is.equal}{Quantity,Quantity}(a, b)
}
\arguments{
\item{a}{a DGVMTools metadata object}

\item{b}{another DGVMTools metadata object of the same type}
}
\value{
logical
}
\description{
checks if two objects have the same quantity, spatial or temporal extent
}
\section{Methods (by class)}{
\itemize{
\item \code{a = Quantity,b = Quantity}: Checks two Quantity objects for equality
}}

\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}
