% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsorted.R
\name{is.leapyear}
\alias{is.leapyear}
\title{check if a given year is a leap year}
\usage{
is.leapyear(year, always = FALSE, doy = FALSE)
}
\arguments{
\item{year}{year (integer or vector)}

\item{always}{use leap years even before 1582}

\item{doy}{return days of year instead logical}
}
\value{
logical or integer, if doy os TRUE
}
\description{
check if a given year is a leap year
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

