% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-methods.R
\name{is.object-methods}
\alias{is.object-methods}
\alias{is.Field}
\alias{is.ComparisonLayer}
\alias{is.Source}
\alias{is.SourceInfo}
\alias{is.Quantity}
\alias{is.PFT}
\alias{is.SpatialComparison}
\alias{is.Format}
\title{Check if an object is a particular type of DGVMTools object.}
\usage{
is.Field(input, spatial = FALSE, temporal = FALSE)

is.ComparisonLayer(input)

is.Source(input)

is.SourceInfo(input)

is.Quantity(input)

is.PFT(input)

is.SpatialComparison(input)

is.Format(input)
}
\arguments{
\item{input}{Any R object to be checked}

\item{spatial}{check if input is a spatial object}

\item{temporal}{check if input is a temporal object}
}
\value{
logical
}
\description{
Returns TRUE if an object of the particular object type and FALSE otherwise
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}

Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}
