% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-misc.R
\name{is.Quantity}
\alias{is.Quantity}
\title{Check is an object is a \code{Quantity}.}
\usage{
is.Quantity(input)
}
\arguments{
\item{input}{Any R object to be checked}
}
\value{
logical
}
\description{
Returns TRUE if an object is a \code{Quantity}, and FALSE otherwise
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

