% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000rvc-tools.R
\name{is.RasterComparison}
\alias{is.RasterComparison}
\title{Check if an object is a \code{RasterComparison}.}
\usage{
is.RasterComparison(input)
}
\arguments{
\item{input}{Any R object to be checked}
}
\value{
logical
}
\description{
Returns TRUE if an object is a \code{RasterComparison}, and FALSE otherwise
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

