% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000rvc-tools.R
\name{is.VegObject}
\alias{is.VegObject}
\title{Check is an object is a \code{VegObject}.}
\usage{
is.VegObject(input, spatial = FALSE, temporal = FALSE, site = FALSE)
}
\arguments{
\item{input}{Any R object to bec checked}

\item{spatial}{check if input is a spatial object}

\item{temporal}{check if input is a temporal object}

\item{site}{ckeck if input is a site object}
}
\value{
logical
}
\description{
Returns TRUE if an object is a \code{VegObject}, and FALSE otherwise
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

