% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000rvc-tools.R
\name{is.VegQuant}
\alias{is.VegQuant}
\title{Check is an object is a \code{VegQuant}.}
\usage{
is.VegQuant(input)
}
\arguments{
\item{input}{Any R object to be checked}
}
\value{
logical
}
\description{
Returns TRUE if an object is a \code{VegQuant}, and FALSE otherwise
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

