% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpj-runtools.R
\name{listAllLPJOutput}
\alias{listAllLPJOutput}
\title{List all LPJ-GUESS *.out files in a run directory}
\usage{
listAllLPJOutput(run.directory)
}
\arguments{
\item{run.directory}{A path to a directory on the file system containing some .out files}
}
\value{
A list of all the .out files present, with the ".out" removed.
}
\description{
Simply lists all LPJ-GUESS output variables (stored as .out files) available in a directory. 
Also ignores some common red herrings like "guess.out" and "*.out"
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

