% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{lm_eqn}
\alias{lm_eqn}
\title{Make linear fit equation string}
\usage{
lm_eqn(linear.model)
}
\arguments{
\item{linear.model}{An object of class lm, should have been made with a simple \code{y ~ x} formula}
}
\value{
A character string
}
\description{
Makes a string (form: y = ax \+ b, r^2 = r^2) for putting on plots from a linear model (lm)
}
\details{
Make sure the model is \code{y ~ x} or this function doesn't really make sense
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
