% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-runtools.R
\name{LondonCentre}
\alias{LondonCentre}
\title{London centre a gridcell}
\usage{
LondonCentre(lon)
}
\arguments{
\item{lon}{A longitude value to transform data.table}
}
\value{
The transformed longitude value
}
\description{
Transform a longitude value to lie in the range (-180,180) instead of (0,36)
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}

