% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities.R
\name{lookupQuantity}
\alias{lookupQuantity}
\title{Function to get an Quantity based on an ID string *and* the scope (i.e. a particular model or the 'Standard' properties)}
\usage{
lookupQuantity(quant.id, model.str = "Standard")
}
\arguments{
\item{quant.id}{String holding the id of the \code{Quantity} you want}

\item{model.str}{String holding the name of the model or "Standard"}
}
\description{
lalal
}

