% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-quantities.R
\name{lookupVegQuant}
\alias{lookupVegQuant}
\title{Function to get an VegQuant based on an ID string *and* the scope (i.e. a particular model or the 'Standard' properties)}
\usage{
lookupVegQuant(quant.id, model.str = "Standard")
}
\arguments{
\item{model.str}{String holding the name og the model or "Standard"}

\item{quant.str}{String holding the id of the \code{VegQuant} you want}
}
\description{
lalal
}

