% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpj-quantities.R
\name{lookupVegQuantity}
\alias{lookupVegQuantity}
\title{Function to get an LPJ-GUESS quantity based on a string}
\usage{
lookupVegQuantity(quant.str, verbose = FALSE)
}
\arguments{
\item{quant.str}{String holding the id of the \code{VegQuant} you want}

\item{verbose}{Logical, whether or not to announce which \code{VegQuant} has been found}
}
\description{
NEED to generalise this to deal with other lists of \code{VegQuants}, not just assuming all are LPJ-GUESS
}

