% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-import.R
\name{lpj.radiation}
\alias{lpj.radiation}
\title{Calculate the incoming solar radiation based of sunshine/cloudcover}
\usage{
lpj.radiation(lat, doy, rad.frac, albedo = 0.17, cloudcover = FALSE,
  leap = FALSE)
}
\arguments{
\item{lat}{latitude (vector)}

\item{doy}{day of the year (vector)}

\item{rad.frac}{sunshine/cloudcover. Must have the shape lat x doy}

\item{albedo}{overall land surface albedo value to be applied (default is the the standard LPJ-GUESS value of 0.17)}

\item{cloudcover}{logical, if rad.frac is cloudcover (TRUE), default FALSE}

\item{leap}{use February 29th or not}
}
\value{
incoming solar radiation in W m^-2 as vector or matrix, depending on input shape
}
\description{
incoming net solar radiation (W m^-2) reduced by albedo,
cloud coverage or sun shine fraction, latitude and
day of the year as done in function daylengthinsoleet
(driver.cpp) LPJ-GUESS v2.1
}
\references{
This code is based on the original LPJ-GUESS C++ source code (\href{http://iis4.nateko.lu.se/lpj-guess/}{LPJ-GUESS homepage}).
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}
