% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-LPJ-GUESS.R
\name{LPJQuantFromFilename}
\alias{LPJQuantFromFilename}
\title{Helper function to raster::trim the ".out" or the ".out.gz" from an LPJ-GUESS filename to get the variable in question}
\usage{
LPJQuantFromFilename(var.filename)
}
\arguments{
\item{var.filename}{The string of the filename to be raster::trimmed}
}
\value{
A string less the last fou.
}
\description{
Returns NULL if the last characters are not ".out" or ".out.gz
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
