% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-misc.R
\name{lsos}
\alias{lsos}
\title{Improved list of objects}
\usage{
lsos(pos = 1, pattern, order.by = "Size", decreasing = TRUE,
  head = TRUE, n = 10)
}
\arguments{
\item{pos}{???}

\item{pattern}{Character, to match only certain objects}

\item{decreasing}{Logical, whether to list the objects in decreasing size}

\item{head}{Logical, if true just print the first \code{n} objects}

\item{n}{Integer, how many objects to list if selecting just the head}

\item{orderby}{Character, can be "Type", "Size", "PrettySize", Rows" or "Columns"}
}
\description{
Because rasters and \code{ModelObject}s and so can be quite large, this function is useful for checking how large are the objects you have stored in memory.
}
\keyword{Credit:}
\keyword{Taken}
\keyword{from:}
\keyword{http://stackoverflow.com/questions/1358003/tricks-to-manage-the-available-memory-in-an-r-session}
\keyword{internal}

