% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFieldID.R
\name{makeFieldID}
\alias{makeFieldID}
\title{Make an ID string for a \code{Field}}
\usage{
makeFieldID(field, source, var.string, sta.info, first.year, last.year,
  year.aggregate.method, subannual.resolution = NULL,
  spatial.aggregate.method = NULL, subannual.aggregate.method = NULL,
  spatial.extent.id = NULL, subannual.original = NULL)
}
\arguments{
\item{var.string}{Character string to describe the variable, eg "lai" or "corrected.cmass" or "npp.diff"}

\item{first.year}{The first year if it has been cropped from the orginal duration, otherwise NULL}

\item{last.year}{The first year if it has been cropped from the orginal duration, otherwise NULL}

\item{year.aggregate.method}{Character, method by which this Field was temporally aggregated (if not temporally averaged leave blank of use "none")}

\item{spatial.aggregate.method}{Character, method by which this Field was spatially aggregated (if not spatially averaged leave blank of use "none")}

\item{spatial.extent.id}{The spatial extent of this object if it has been cropped from the orginal simulation extent, otherwise NULL}

\item{unit.string}{Character string to describe the units in which the data are measured}
}
\value{
A character string
}
\description{
Given a string for the quantity and temporal and spatial extents and averaging flags, build an appropriate (and unique) ID string
for use in the \code{id} slot of a \code{Field} and for filenames etc.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
