% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-utility-functions.R
\name{makeMapOverlay}
\alias{makeMapOverlay}
\title{Make a map overlay for ggplot2}
\usage{
makeMapOverlay(map.overlay, all.lons, interior.lines, xlim, ylim)
}
\arguments{
\item{map.overlay}{A character string specifying the overlay to be used a string matching maps package dataset}

\item{all.lons}{A numeric vector of all the longitudes to be plotted, this is used to determine if it the over lay should be on longitues (-180,180) or (0,360).}

\item{interior.lines}{A logical, if TRUE include the internal country lines}

\item{xlim}{A numeric vector of length 2 to giving the longitude window that the overlay should cover}

\item{ylim}{A numeric vector of length 2 to giving the latitide window that the overlay should cover}
}
\value{
Returns data.frame suitable for plotting with ggplot::geom_path
}
\description{
Take a string and derives an approriate data.frame that can be used to add a map overlay 
(eg coast or country lines from the maps and mapdata packages) with the ggplot::geom_path function.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
