% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvc-plotting.R
\name{makeOverlay}
\alias{makeOverlay}
\title{Returns a SpatialLines object (of, for example, country or continent outlines) for overlaying on an spplot with for example plotVegMaps}
\usage{
makeOverlay(which)
}
\arguments{
\item{which}{A string specifiying "hires" or "lowres" for a full overlay of countries,
of "hires-continents" or "lowres-contents" for just the continent outlines.}
}
\value{
a SpatialLines object
}
\description{
Returns a SpatialLines object (of, for example, country or continent outlines) for overlaying on an spplot with for example plotVegMaps
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}, Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

