% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlotTitle.R
\name{makePlotTitle}
\alias{makePlotTitle}
\title{Make a plot titles}
\usage{
makePlotTitle(fields)
}
\arguments{
\item{fields}{A list of fields which will be plotted and the so the plot title should describe}
}
\value{
A list of two character strings.  The first (called "title") is the main title.  The second (called "subtitle" is the subtitle)
}
\description{
Build an appropriate plot title (and subtitle) from the meta-data from the fields (whihc will be plotted)
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
