% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-spatial.R
\name{makeStatsOverlay}
\alias{makeStatsOverlay}
\title{Make Stats Overlay}
\usage{
makeStatsOverlay(stat.results, stats = "NME", panel = NULL, plot.extent,
  x.pos.frac = 0.15, y.pos.frac = 0.15, cex = 1.5, digits = 3)
}
\arguments{
\item{stat.results}{An object of type SpatialComparison (from which the statistics are taken)}

\item{stats}{Character string (or list of character strings) specifying which stats to show.  Should be the name of a slot of a SpatialComparison object, for example "NME" or "Kappa"}

\item{panel}{An integer to specify which panel the stats are drawn on (leave as NULL for all panels)}

\item{plot.extent}{A raster::extent of a SpatialEXtent object used to give the size of the plot and therefore how to scale the fractional coordinates into longitude and latitude values for the position}

\item{x.pos.frac}{Numeric, the x position (in fraction of plot size) of the stats}

\item{y.pos.frac}{Numeric, the y position (in fraction of plot size) of the stats}

\item{cex}{The text size for the stats text}

\item{digits}{The number of digits to which to round the stats}
}
\description{
Make a layout object to display stats on a map when using either spplot() or plotSpatial().
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
