% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-plotting.R
\name{makeVariableIDString}
\alias{makeVariableIDString}
\title{Make a variable ID}
\usage{
makeVariableIDString(quantity.id, layer = NULL, run.id = NULL,
  special = NULL)
}
\arguments{
\item{layer}{Character string the particular subset of the quantity being plotted (eg. "Evergreen" or a PFT id)}

\item{run.id}{Character string, the ID of the run}

\item{special}{Character string, a free choice to provide extar information or differentiate this from other varibles (eg. "LanduseCorrected")}

\item{quantity.str}{Character string for the quantity plotted (eg. "lai")}
}
\value{
A character string to describe a variable in a file name.  For example "lai.Evergreen.Test1.LanduseCorrected".
}
\description{
Helper function to build an appropriate variable ID (for use in filenames) from some variables.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}

