% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-runtools.R
\name{makeVegObjectID}
\alias{makeVegObjectID}
\title{Make an ID string for a \code{VegObject}}
\usage{
makeVegObjectID(var.string, temporal.extent = NULL, spatial.extent = NULL,
  temporally.averaged = FALSE, spatially.averaged = FALSE)
}
\arguments{
\item{var.string}{Character string to describe the variable, eg "lai" or "corrected.cmass" or "npp.diff"}

\item{temporal.extent}{The temporal extent of this object if it has been cropped from the orginal duration, otherwise NULL}

\item{spatial.extent}{The spatial extent of this object if it has been cropped from the orginal simulation extent, otherwise NULL}

\item{temporally.averaged}{Logical, should be TRUE if temporal averaging has been done}

\item{spatially.averaged}{Logical, should be TRUE if spatial averaging has been done}
}
\value{
A character string
}
\description{
Given a string for the quantity and temporal and spatial extents and averaging flags, build an appropriate (and unique) ID string
for use in the \code{id} slot of a \code{VegObject} and for filenames etc.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

