% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-framework-functions.R
\name{makeYAxis}
\alias{makeYAxis}
\title{Make y-axis}
\usage{
makeYAxis(objects)
}
\arguments{
\item{objects}{The list of Fields to be plotted (should have been check by santiseFieldsForPlotting first)}
}
\value{
Returns the y-axis as a chatacter string,
}
\description{
This is an internal helper function to build a y-axis for Temporal and Subannual plots, possibly with multiple Quantities
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
