% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-netcdf-utility-functions.R
\name{makeYearsAndDaysFromDailyTimeAxis}
\alias{makeYearsAndDaysFromDailyTimeAxis}
\title{makeYearsAndDaysFromDailyTimeAxis 
Currently not used, see function processDailyNCAxis in this file
CAN PROBABLY BE DELETED}
\usage{
makeYearsAndDaysFromDailyTimeAxis(nc_file, time_axis_string, start.date)
}
\description{
makeYearsAndDaysFromDailyTimeAxis 
Currently not used, see function processDailyNCAxis in this file
CAN PROBABLY BE DELETED
}
\keyword{internal}
