% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-utility-functions.R
\name{matchPFTCols}
\alias{matchPFTCols}
\title{Make a PFT colour list}
\usage{
matchPFTCols(values, pfts, others = list(Total = "black", None =
  "grey75", Tree = "brown", Grass = "green", Shrub = "red"))
}
\arguments{
\item{values}{List of values (as chararacters) for which you want standard colours.}

\item{pfts}{A list of PFT objects (which should contain PFTs with ids provided in 'values)}

\item{others}{A list of other name-colour combinations, for example to plot 'Total' as black, "None" as grey, or whatever.  Some defaults are defined.}
}
\value{
Returns a named list of colours, where the names are the values that the colours will represent
}
\description{
This is a helper function for when plotting PFTs by colour.  It takes a list of PFT ids (other things like "Total" or "Tree" can also be specified) and returns a list 
of colours with the names of the PFT (which is how ggplot likes colours to be specified).
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
