% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification-schemes.R
\name{MeditBiomeRules}
\alias{MeditBiomeRules}
\title{Rules to classify coarses Mediterranean biomes}
\usage{
MeditBiomeRules(x)
}
\arguments{
\item{x}{Numerical vector of LAI values for a particular location. 
Certain fractions and quantities should have been pre-calculated.}
}
\value{
Biomes code (1-8, ordering as in the Forrest et al. 2015 figure)
}
\description{
No reference yet
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
