% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome-schemes.R
\name{MegaBiomeRules_dev}
\alias{MegaBiomeRules_dev}
\title{Rules to classify coarser "mega biomes" from LAI, latitude and GDD5}
\usage{
MegaBiomeRules_dev(x)
}
\arguments{
\item{x}{Numerical vector of LAI values for a particular location. 
Certain fractions and quantities should have been pre-calculated.}
}
\value{
Biomes code (1-13)
}
\description{
Unpublished but possibly useful in principle.  Simpler classes that Smith et al 2014,
but more complex that Forrest et al 2015.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
