% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge-methods.R
\name{merge}
\alias{merge}
\alias{merge,Field,Field-method}
\title{Generic method for coercing to raster}
\usage{
merge(x, y, source = NULL, ...)

\S4method{merge}{Field,Field}(x, y, source = NULL, ...)
}
\description{
Generic method for coercing to raster
}
