% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-framework-functions.R
\name{mergeFieldsForPlotting}
\alias{mergeFieldsForPlotting}
\title{Merge data from Field for plotting}
\usage{
mergeFieldsForPlotting(fields, add.Quantity = FALSE, add.Site = FALSE,
  add.Region = FALSE, add.Years = FALSE)
}
\arguments{
\item{fields}{The list of Fields to be plotted (should have been check by santiseFieldsForPlotting first)}

\item{add.Quantity}{Logical, if TRUE add a column with the Quantity name of each Field}

\item{add.Site}{Logical, if TRUE add a column with the with a string containing the Lon and Lat of each site (gridcell).}

\item{add.Region}{Logical, if TRUE add a column with the with a string containing the Region (as defined by the spatial.extent.id) of each Field.}

\item{add.Years}{Logical, if TRUE add a column (with name "Years") with a string containing the first and last years of each Field
(as defined by first.year and last.year).}
}
\value{
Returns a data.table
}
\description{
This is an internal helper function which pulls out the data from a bunch of Fields to be plotted, adds columns to describe the characteristics 
of each Field (ie Quantity, Site code etc) and returns it all in one big melted data.table.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
