% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo.makeDriver.ED2.r
\name{meteo.makeDriver.ED2}
\alias{meteo.makeDriver.ED2}
\title{Make meteo driver .h5 files for the ED2 model. Based mainly on the code by Marcos Longo.
TODO: make this a bit more elegant looking.}
\usage{
meteo.makeDriver.ED2(meteo, height, name, path.header, path.output,
  verbose = TRUE)
}
\arguments{
\item{meteo}{list with variables following my naming conventions}

\item{height}{[m] height above ground level}

\item{name}{site name, will be used as prefix of all meteo files}

\item{path.header}{the header prefix, for the ED2 header file (eg. the directory on your HPC where these data will be located)}

\item{path.output}{output folder where all h5 files will be saved}

\item{verbose}{Give some process feedback to the user or not}
}
\description{
Make meteo driver .h5 files for the ED2 model. Based mainly on the code by Marcos Longo.
TODO: make this a bit more elegant looking.
}
\author{
Wim Verbruggen (wim.verbruggen@ugent.be)

Marcos Longo
}
