% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsorted.R
\name{mmdd2doy}
\alias{mmdd2doy}
\title{date in the form of MonthDay to day of year conversion}
\usage{
mmdd2doy(mmdd, leap = FALSE)
}
\arguments{
\item{mmdd}{string or number interpreted as MMDD}

\item{leap}{in a leap year of not.}
}
\value{
Day of the year as integer (1. Januar is day 1).
}
\description{
date in the form of MonthDay to day of year conversion
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

