% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsorted.R
\name{modelObject2Array}
\alias{modelObject2Array}
\title{Convert a ModelObject to a multi-dimensional array}
\usage{
modelObject2Array(d, cname = FALSE, invertlat = FALSE)
}
\arguments{
\item{d}{the data.table of a \code{\linkS4class{ModelObject}}}

\item{cname}{the column name to convert, if not set a list is returned}

\item{invertlat}{start in the north}
}
\value{
a array or a list or arrays
}
\description{
Convert a ModelObject to a multi-dimensional array
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

