% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-processing.R
\name{newLayer}
\alias{newLayer}
\title{Combine layers of a ModelObject (or a data.table)}
\usage{
newLayer(input, layers, method = NULL, PFT.data = NULL)
}
\arguments{
\item{input}{The ModelObject for which to aggregate layers.}

\item{layers}{The new layers to produce}

\item{method}{The method to use to aggregate the layers ie. "mean" or "sum".  However, it is often most sensible to leave it unspecified and use the default (see below)}

\item{PFT.data}{If calling the function on a data.table, it is neccessary to specify a PFT set here.  Normally this function will be called on  ModelObject so it is not neccessary.}
}
\value{
A ModelObject (or data.table) with the new layers added
}
\description{
This is very useful and important function.  It aggregates different layers 
of a ModelObject according the the desired method (or a sensible default).  Please note some special features of the arguments, designed for convenience, which are described below.
}
\details{
Whilst the \code{method} argument can be specified for maximum flexibility, the recommended usage is not specify this argument.
In this case, the function uses the default of the Quantity object of the ModelObject, which should be the sensible option.  
For example, per-PFT variables (such as lai or cmass) should typically be summed, 
but monthly variables (such as soil water content) should be average to calculate the seasonal means.

For convenience, both \code{layers} will be expanded using \code{expandLayers}.
This allows all lifeforms totals in a ModelObject to be calculated using a simple call such as

 \code{veg.obj <- newLayer(veg.obj, c("lifeforms"))}
 
See documention of \code{expandLayers} for details.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\seealso{
expandslayers getVegFractions
}

