% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-aDGVM1.R
\name{openaDGVM1OutputFile_FireMIP}
\alias{openaDGVM1OutputFile_FireMIP}
\title{Open an aDGVM1 .out file}
\usage{
openaDGVM1OutputFile_FireMIP(run, quant, target.sta,
  file.name = file.name, verbose = FALSE,
  soil_water_capacities = "none", ...)
}
\arguments{
\item{run}{A \code{Source} containing the meta-data about the aDGVM1 run}

\item{quant}{A Quantity to define what output file from the aDGVM1 run to open.}

\item{file.name}{Character string holding the name of the file.  This can be left blank, in which case the file name is just taken to be 
"<quant@id>.out" (also "<quant@id>.out.gz")}

\item{verbose}{A logical, set to true to give progress/debug information}

\item{first.year}{The first year (as a numeric) of the data to be return}

\item{last.year}{The last year (as a numeric) of the data to be return}
}
\value{
a data.table (with the correct tear offset and lon-lat offsets applied)
}
\description{
\code{openaDGVM1OutputFile} returns a data.table object given a string defining a vegetation quantity 
from the run (eg. "lai", to read the file "lai.out") and  \code{Source} object which defines where the run is on disk and the offsets to apply
}
\details{
Note that the files can be gzipped on UNIX systems, but this might fail on windows systems.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
