% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FireMIP.R
\name{openFireMIPOutputFile}
\alias{openFireMIPOutputFile}
\title{Open a FireMIP output file}
\usage{
openFireMIPOutputFile(run, quantity, temporal.extent = NULL,
  spatial.extent = NULL, verbose = TRUE)
}
\arguments{
\item{run}{A ModelRun object to define the run we want to open.}

\item{quantity}{A Quantity object to define which variable we want to look up}

\item{temporal.extent}{A TemporalExtent object to define which years we want to read}

\item{spatial.extent}{A SpatialExtent object to define the spatial extent we want to read}

\item{verbose}{Logical, if TRUE spew forth a lot of info.}
}
\value{
A data.table
}
\description{
Opens a .nc file from the FireMIP output and sorts out the meta-data and dimensions and all that messy stuff.  
Returns a data.table, because it is intended to be called by getModelObject(), but of course the data.table could be used directly if you wish
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
