% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-GUESS.R
\name{openLPJOutputFile_FireMIP}
\alias{openLPJOutputFile_FireMIP}
\title{Open an LPJ-GUESS .out file}
\usage{
openLPJOutputFile_FireMIP(
  run,
  quant,
  layers = NULL,
  target.sta,
  file.name = file.name,
  verbose = FALSE,
  soil_water_capacities = "none",
  ...
)
}
\arguments{
\item{run}{A \code{\linkS4class{Source}}  containing the meta-data about the LPJ-GUESS run}

\item{quant}{A Quantity to define what output file from the LPJ-GUESS run to open.}

\item{layers}{A character string (or a vector of character strings) specifying which layer columns are to be read.  NULL (default) means read all.}

\item{file.name}{Character string holding the name of the file.  This can be left blank, in which case the file name is just taken to be 
"<quant@id>.out" (also "<quant@id>.out.gz")}

\item{verbose}{A logical, set to true to give progress/debug information}
}
\value{
a data.table (with the correct tear offset and lon-lat offsets applied)
}
\description{
\code{openLPJOutputFile} returns a data.table object given a string defining a vegetation quantity 
from the run (eg. "lai", to read the file "lai.out") and \code{\linkS4class{Source}} object which defines where the run is on disk and the offsets to apply
}
\details{
Note that the files can be gzipped on UNIX systems, but this might fail on windows systems.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
