% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-netcdf-utility-functions.R
\name{parseStartDate}
\alias{parseStartDate}
\title{Parse start date}
\usage{
parseStartDate(start.date.string)
}
\description{
Parses a "units" attribute string of the time axis of a netCDF file.  Returns a three-vector of year,month,time.
NOTE: Accepts years < 0 and years > 9999, unlike POSIX and "Date" dates.
}
\keyword{internal}
