% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/periods.R
\docType{class}
\name{Period-class}
\alias{Period-class}
\alias{all.periods}
\alias{periods}
\alias{all.months}
\alias{all.seasons}
\alias{annual}
\title{Time periods - eg. a month or a season or a year.}
\format{
An object of class \code{list} of length 17.

List of\code{Period} objects

An object of class \code{list} of length 12.

An object of class \code{list} of length 4.

An object of class \code{list} of length 1.
}
\usage{
all.periods

periods

all.months

all.seasons

annual
}
\description{
A simple S4 class to define a month, a season or a year used for aggregating monthly values to seasonal or annual values and for making monthly plots with nice labels etc.
}
\details{
Since these are for the most part standard (ie. people commonly need the months, the seasons (DJF, MAM, JJA, SON) and annual) these are defined in simple list that might be commonly used and can be looped through.  These are:
\describe{
  \item{all.months}{which contains all the months.}
  \item{all.seasons}{which contains all the seasons.}
  \item{annual}{which only contains only the annual period.}
  \item{all.periods}{which contains all of the above.} 
}
 
However other periods can be defined for specific growing seasons etc.
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular time period.  Recommended to be alphanumeric because it is used to construct file names.}

\item{\code{name}}{A character string to describe the spatial extent. Used for building plot labels, not file names, so doesn't need to be alphanumeric and can so can be prettier.}

\item{\code{abbreviation}}{Not really sure if this is necessary, think is it covered by the "id" slot}

\item{\code{index}}{The index of the months of the year for this period.  For example, Jan has index = 1, DJF has index =c (12,1,2) and annual has index = c(1,2,3,4,5,6,7,8,9,10,11,12)}

\item{\code{padded.index}}{As slot \code{index} but zero-padded strings for use in when encoding/decoding file names.}

\item{\code{contains}}{Holds the names of the other periods included in this period.  For example for Jan just contains "Jan", but DJF it contains c("Dec","Jan,"Feb)}

\item{\code{days}}{Number of days in this time period during a normal (non-leap) year}

\item{\code{days.leap}}{Number of days on this period during a leap year}

\item{\code{col}}{A colour (as a character string) for this particular period}
}}

\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{datasets}
