% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-ggplot2.R
\name{plotGGHist}
\alias{plotGGHist}
\title{Plot a histogram}
\usage{
plotGGHist(input, column = "value", colors = NA, bars = TRUE,
  lines = FALSE, bins = 10, long.title = TRUE, plot = TRUE, ...)
}
\arguments{
\item{input}{a VegSpatial object or a list of several.}

\item{column}{The column(s) to display. Default: 'value'.}

\item{colors}{Colors for the diffent VegSpatial objects or columns. Must have the same length otherwise colors are choosing automatically.}

\item{bars}{draw bars of counted occurence (default TRUE).}

\item{lines}{also draw density lines (default FALSE).}

\item{bins}{number of bins for histogram (default: 25)}

\item{long.title}{If the description (default) should be used as titles or the shorter id.}

\item{plot}{If FALSE only the data is returned, without drawing the map.}

\item{...}{Ignored further parameters}
}
\description{
Plot the given columns as histogram
}

