% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-ggplot2.R
\name{plotGGScatter}
\alias{plotGGScatter}
\title{Create a various forms of scatterplots}
\usage{
plotGGScatter(x, y, x.column = "value", y.column = "value", limit = NULL,
  flip = TRUE, density = FALSE, wrap.column = NULL, color.column = NULL,
  colors = "orange", sym.col = FALSE, alpha = 0.1, lines = NA,
  labels = NA, label.pos = "bottomright", equal.axis = TRUE, wrap = 2,
  plot = TRUE, verbose = TRUE)
}
\arguments{
\item{x}{A ModelObject or a list of them.}

\item{y}{A ModelObject, dataframe/data.table or raster.}

\item{x.column}{Column name of x to use for the scatter.}

\item{y.column}{Column name of y to use for the scatter.}

\item{limit}{A data.frame with the possible columns min, max, xmin, xmax, ymin, ymax to exclude data from x and/or y. If a data.column 'inclusive' is present and TRUE, the given values will remain in the data.}

\item{flip}{Should x and y be on the other axis. Default TRUE, means x is on the y-axis and y on the x-axis.}

\item{density}{Use a ggplot density_2d function instead of a point cloud. Values: 'polygon', 'raster' and everything else than FALSE will result in 'hexagonal'.}

\item{wrap.column}{A column name in x, which should be used for faceting.}

\item{color.column}{A column name for coloring the points (NOT working currently).}

\item{colors}{A data.frame for colors to use mandatory column names: XXX}

\item{sym.col}{Should the colors be symmetrical (NOT working currently).}

\item{alpha}{transparency value for points (ignored in density mode).}

\item{lines}{Character vector of any of '1:1', 'lm' and/or 'gam'. Or a data.frame with the columns 'lines', 'color', 'type', 'width'.}

\item{labels}{Character string of metrics to be added. Possible values 'rsq', 'rmse', 'me', 'eq'}

\item{label.pos}{Position of the labels as character consiting 'top' or 'bottom' and 'left', 'center', 'right'}

\item{equal.axis}{Should the x and y- axis be scaled equally.}

\item{wrap}{number of column, when wrapping}

\item{plot}{Should a ggplot-object be returned (default) or the data.table.}

\item{verbose}{print some stupid messages.}
}
\value{
a data.table or a ggplot2-object.
}
\description{
Create a point cloud (scatterplot) or density plot with several metrics and line fits.
}
\examples{
message("See templates/Example.ggplot.R")
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}
\references{
Mayer, D. G. and Butler, D. G.: Statistical validation, Ecological Modelling, 68(1-2), 21-32, \href{doi:10.1016/0304-3800(93)90105-2}{dx.doi.org/10.1016/0304-3800(93)90105-2}, 1993.
}

