% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-plotting.R
\name{plotHistoComparison}
\alias{plotHistoComparison}
\title{Make a histograms by comparing two rasters}
\usage{
plotHistoComparison(model, data, run, data.name, quant, breaks, plot.range,
  stat.results = NULL)
}
\arguments{
\item{model}{Raster of the model}

\item{data}{Raster of the data}

\item{run}{The \code{VegRun} object for the run plotted (optional)}

\item{data.name}{Character string for the data}

\item{quant}{The quantity plotted (as \code{VegQuant} object)}

\item{breaks}{A numerical vector of the breakpoints for the histograms}

\item{plot.range}{A numerical vector with two elements defining the range to plot on the histogram.}

\item{stat.results}{The \code{RasterComparion} object if it has already been calculated

The plot is saved to the run directory of the run object}
}
\description{
Takes either two rasters or a \code{RasterComparion} object and makes a histograms.
}
\details{
Probably should be made a method of \code{RasterComparion}, 
with further additions to that class necessary.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

