% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{plotScatterComparison}
\alias{plotScatterComparison}
\title{Plot scatter of model vs data}
\usage{
plotScatterComparison(data.obj, showFitLine = TRUE, showStats = TRUE,
  labels = NULL, alpha = 0.05, text.size = 6)
}
\arguments{
\item{data.obj}{The DataObject for which to plot the residuals}

\item{showFitLine}{Boolean, if TRUE shows a linear fit and the fit equation}

\item{showStats}{Boolean, if TRUE show some stats do quantify how well the model fits the data}

\item{labels}{Character vector of labels (one for each run).  If not provided uses the run ids}

\item{alpha}{Numeric for the transparency of tre points (range (0,1], default = 0.05))}

\item{text.size}{Numeric to scale the text size on the plot (default = 6)}
}
\value{
A ggplot2 plot.
}
\description{
Function will makes multiple scatter plots on one page if more than one ModelRun has been compared to the the DataObject.
}
\details{
This function should be called after a call to \code{benckmarkSpatial} for a DataObject.  It plots the scatters for each model run to which it was compared .  
It is called automatically by \code{benckmarkSpatial}, but can be called again for better flexibility.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

