% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatialComparison.R
\name{plotSpatialComparison}
\alias{plotSpatialComparison}
\title{Plot a comparison between two spatial layers}
\usage{
plotSpatialComparison(
  comparisons,
  type = c("difference", "percentage.difference", "values", "nme"),
  limits = NULL,
  legend.title,
  panel.bg.col = "white",
  override.cols = NULL,
  symmetric.scale = TRUE,
  do.phase = FALSE,
  metrics = c(),
  metric_size = waiver(),
  metric_x_pos = 0.025,
  metric_y_pos = 0.975,
  ...
)
}
\arguments{
\item{comparisons}{The data to plot, must be a Comparison or a list of Comparisons}

\item{type}{A character specifying what type of plot to make. Can be "difference" (default, for a difference plot), "percentage.difference", "values" 
(actual values, side-by-side) or "nme" (for the Normalised Mean Error, not yet implemented)}

\item{limits}{A numeric vector with two members (lower and upper limit) to limit the plotted values.}

\item{legend.title}{A character string or expression to override the default legend title. Set to NULL for no legend title.  The default legend title is the \code{units}
of the \linkS4class{Quantity} of the first \linkS4class{Comparison} provided in the \code{comparisions} argument.  This argument allows general flexibility, but it is particularly handy
to facilitate expressions for nicely marked up subscript and superscript.}

\item{panel.bg.col}{Colour string for the panel background, default to "white" for absolute values plots, and a grey for difference plots.}

\item{override.cols}{A colour palette function to override the defaults.}

\item{symmetric.scale}{If plotting a differences, make the scale symmetric around zero (default is TRUE)}

\item{do.phase}{Logical, only applies to plotting Comparison objects of type "seasonal".
If TRUE plot the the seasonal phase, if FALSE (the default), plot the seasonal concentration.}

\item{metrics}{A character vector specifying the metrics to put on the plots. For spatial data these can be: "ME", "NME", "NMSE", "RMSE", "NME_2", "NMSE_2", "NME_3", 
"NSME_3", "r2_eff", "r", "r2", "m", "c".}

\item{metric_size}{A numeric value for the size of the metric text, note that this will be also be scaled by the \code{text.multiplier} argument.}

\item{metric_x_pos, metric_y_pos}{A number value specifying the x/y location of the metric text as a fraction of the plot area.  Note, it uses the
overall range of the plotting for calculating this (not the specific ranges of the axes), so using facets or grids with "free" scales will mess this up.}

\item{...}{Parameters passed to \link{plotSpatial}}
}
\value{
Returns a ggplot object
}
\description{
This function is for plotting maps from Comparison objects (or a list of those Comparisons).  Three types of comparisons plots are supported: 'difference' - 
a difference map; "values" - the absolute values plotted in panels and "percentage.difference" - the percentage differences.
}
\details{
A wrapper for around \link{plotSpatial} to plot the spatial Comparisons as maps.  Extra arguments to \link{plotSpatial} can also be specified.
}
\seealso{
\link{plotSpatial},  \link{compareLayers}
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
