% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTemporalComparison.R
\name{plotTemporalComparison}
\alias{plotTemporalComparison}
\title{Plot a comparison between two Temporal layers}
\usage{
plotTemporalComparison(
  comparisons,
  type = c("difference", "percentage.difference", "values", "nme"),
  limits = NULL,
  panel.bg.col = "white",
  override.cols = NULL,
  symmetric.scale = TRUE,
  percentage.difference.limit = 300,
  do.phase = FALSE,
  ...
)
}
\arguments{
\item{comparisons}{The data to plot, must be a Comparison or a list of Comparisons}

\item{type}{A character specifying what type of plot to make. Can be "difference" (default, for a difference plot), "
percentage.difference", "values" (actual values, side-by-side).}

\item{limits}{A numeric vector with two members (lower and upper limit) to limit the plotted values.}

\item{panel.bg.col}{Colour string for the panel background, default to "white" for absolute values plots, and a grey for difference plots.}

\item{override.cols}{A colour palette function to override the defaults.}

\item{symmetric.scale}{If plotting a differences, make the scale symmetric around zero (default is TRUE)}

\item{percentage.difference.limit}{If percentage difference to be plotted, what to limit the scale to.  Default is 300.}

\item{do.phase}{Logical, only applies to plotting Comparison objects of type "seasonal".
If TRUE plot the the seasonal phase, if FALSE (the default), plot the seasonal concentration.}

\item{...}{Parameters passed to \code{plotTemporal()}}
}
\value{
Returns a ggplot object
}
\description{
This function is for plotting maps from Comparison objects (or a list of those Comparisons).  Three types of comparisons plots are supported: 'difference' - 
a difference map; "values" - the absolute values plotted in panels and "percentage.difference" - the percentage differences.
}
\details{
A wrapper for around \code{plotTemporal()} to plot the temporal Comparisons as maps.  Extra arguments to \code{plotTemporal} can also be specified.
}
\seealso{
\code{plotTemporal},  \code{compareLayers}
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
