% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotXYComparison.R
\name{plotXYComparison}
\alias{plotXYComparison}
\title{Plot a scatter comparison between two layers}
\usage{
plotXYComparison(
  comparisons,
  type = c("points", "hex", "bin2d"),
  fit_line_col = NULL,
  perfect_line_col = NULL,
  matchLimits = TRUE,
  metrics = c(),
  metric_size = waiver(),
  metric_x_pos = 0.025,
  metric_y_pos = 0.975,
  text.multiplier = NULL,
  col.by = NULL,
  ...
)
}
\arguments{
\item{comparisons}{The data to plot, must be a Comparison or a list of Comparisons}

\item{type}{A character specifying what type of plot to make. Can be "points" (default, for geom_points), "hex" (for hex binning), "bin2d" (for square binning).
There might be more useful options to add later.}

\item{fit_line_col}{A colour for the fit line through the data (default is NULL meaning no fit line).}

\item{perfect_line_col}{A colour for the perfect one-to-one line (default is NULL meaning no one-to-one line).}

\item{matchLimits}{Logical, determins if the X and X axes should have the same range. Default is TRUE.
(only works for \code{type = "points"}). Can meaningfully take the values of spatiotemporal dimensions which are in the dataset 
such as "Day", "Month", "Season", "Year", "Lon" and "Lat". 
By default \code{col.by} is set to NULL, which doesn't distinguish the points by colour.}

\item{metrics}{A character vector specifying the metrics to put on the plots. For spatial data these can be: "ME", "NME", "NMSE", "RMSE", "NME_2", "NMSE_2", "NME_3", 
"NSME_3", "r2_eff", "r", "r2", "m", "c".}

\item{metric_size}{A numeric value for the size of the metric text, note that this will be also be scaled by the \code{text.multiplier} argument.}

\item{metric_x_pos, metric_y_pos}{A number value specifying the x/y location of the metric text as a fraction of the plot area.  Note, it uses the
overall range of the plotting for calculating this (not the specific ranges of the axes), so using facets or grids with "free" scales will mess this up.}

\item{text.multiplier}{A number specifying an overall multiplier for the text on the plot.
Make it bigger if the text is too small on large plots and vice-versa.}

\item{col.by}{Character strings defining the aspects of the data which which should be used to set the colour of the points}

\item{...}{Arguments passed to \code{ggplot2::facet_wrap()}.  See the ggplot2 documentation for full details but the following are particularly useful.
\describe{
 \item{"nrow"}{The number of rows of facets}
 \item{"ncol"}{The number of columns of facets}
 \item{"scales"}{Whether the scales (ie. x and y ranges) should be fixed for all facets.  Options are "fixed" (same scales on all facets, default)
 "free" (all facets can their x and y ranges), "free_x" and "free_y"  (only x and y ranges can vary, respectively).}
 \item{"labeller"}{A function to define the labels for the facets.  This is a little tricky, please look to the ggplot2 documentation.
 But basically what you want is to define a named character vector, the names are the previous facet names and the values are the new names.  
 Then make this into a function by passing it to the ggplot function "as.labeller", and then that becomes your 'labeller' argument.} 
}}
}
\value{
Returns a ggplot object
}
\description{
This produces X-Y scatter plots from Comparison objects (or a list of those Comparisons).  These plots can be rendered with points or point densities
 with square binsor hexagonal bins.
}
\details{
A wrapper for around \link{plotSpatial} to plot the spatial Comparisons as maps.  Extra arguments to \link{plotSpatial} can also be specified.
}
\seealso{
\link{plotSpatial},  \link{compareLayers}
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
