% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-methods.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,PFT-method}
\alias{print,Quantity-method}
\alias{print,Period-method}
\alias{print,Source-method}
\alias{print,STAInfo-method}
\alias{print,Field-method}
\alias{print,Comparison-method}
\alias{print,Statistics-method}
\alias{print,BiomeScheme-method}
\alias{print,Format-method}
\title{Print a DGVMTools object}
\usage{
\S4method{print}{PFT}(x)

\S4method{print}{Quantity}(x)

\S4method{print}{Period}(x)

\S4method{print}{Source}(x)

\S4method{print}{STAInfo}(x)

\S4method{print}{Field}(x)

\S4method{print}{Comparison}(x)

\S4method{print}{Statistics}(x)

\S4method{print}{BiomeScheme}(x)

\S4method{print}{Format}(x)
}
\arguments{
\item{x}{Any DGVMTools object to be printed}

\item{...}{other arguments to be passed on to a further print functions, not currently used.}
}
\description{
Print a DGVMTools object in a reasonably nice and space efficient way.
}
\details{
Simple stuff, implemented as S4 methods.  Further tweaking of the formatting may be in order, feedback from users welcome.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
