% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-netcdf-utility-functions.R
\name{processDailyNCAxis}
\alias{processDailyNCAxis}
\title{processDailyNCAxis
This function processes a time axis which is assumed to be daily
It compares the time axis to the years requested and return the start and count values for reading up the netCDF file
and the actual years and days corresponding to this selection.}
\usage{
processDailyNCAxis(
  axis.values,
  start.year,
  start.day,
  target.STAInfo,
  calendar,
  year.offset
)
}
\description{
UNUSED, and is DEPRECATED in favour of processDailyRelativeNCAxis().  However this function may yet find a use for absolute time axes.
}
\keyword{internal}
