% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-netcdf-utility-functions.R
\name{processDailyRelativeNCAxis}
\alias{processDailyRelativeNCAxis}
\title{processDailyRelativeNCAXis
This function processes a relative time axis with daily intervals
It calculates the day, month and year corresponding to the valies on the axis, and also crops according to the requested years
Finally it returns a data.tables with all this information including the netCDF indices (cropped to the years)}
\usage{
processDailyRelativeNCAxis(
  axis.values,
  start.year,
  start.month,
  start.day,
  target.STAInfo,
  calendar,
  year.offset
)
}
\description{
processDailyRelativeNCAXis
This function processes a relative time axis with daily intervals
It calculates the day, month and year corresponding to the valies on the axis, and also crops according to the requested years
Finally it returns a data.tables with all this information including the netCDF indices (cropped to the years)
}
\keyword{internal}
