% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Globcover2009.R
\name{processGlobcover2009}
\alias{processGlobcover2009}
\title{Process the Globcover2009 data}
\usage{

  processGlobcover2009(location = "/data/forrest/LandUseLandCover/Globcover2009/",
  target.grid = raster::raster(nrows = 360, ncols = 720, xmn = -180, xmx =
  180, ymn = -90, ymx = 90))
}
\arguments{
\item{location}{A character string specifying the location of the original Globcover2009 data}

\item{target.grid}{A RasterLayer definging the target extent and resoutuion of the processed data}
}
\value{
A RasterBrick with one layer for each landover catorgory
}
\description{
Read the Globcover2009 landcover dataset at its original resution and process it by counting the number of original resolution pixels to fall into each 
gridcell of the target grid for each landcover. Then return the fraction of each landcover type per class in a RasterBrick.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
