% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-methods.R
\name{promoteToRaster}
\alias{promoteToRaster}
\title{Prepares data as a raster object for plotting.}
\usage{
promoteToRaster(input.data, layers = "all", tolerance = 0.01,
  grid.topology = NULL)
}
\arguments{
\item{input.data}{Field, Comparison or data.table}

\item{layers}{The columns to be selected included in the final Raster* object.  Use NULL or "all" if all layers are required.}

\item{tolerance}{Tolerance (in fraction of gridcell size) for unevenly spaced Lon and Lats, 
when converting gridded table to a raster.}

\item{grid.topology}{A character string defining the grid topology when going from a table to raster, used in a call to SpatialPixels}
}
\value{
A RasterLayer (or RasterBrick)
}
\description{
Converts a Field, Comparison or data.table to Raster object, also subsetting the requested layers.  
This is generally called in the as.Raster and the setAs functions but can be useful in and of itself.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
