% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-classes.R
\docType{class}
\name{Quantity-class}
\alias{Quantity-class}
\title{Class to hold the data for a vegetation quantity}
\description{
Defines specific vegetation quantities like "lai" (LAI) or "mwcont_upper" (mean monthly water).  
Includes metadata about the quantity (the units it is measured in for exampel), values for default plot proprties (color scales, plot ranges) and the aggragating method - ie whether the quantity shoudd generally be summed (for example monthly burnt area or C mass per PFT) or averaged (for example monthly soil water content).
Note that the used will probably need to define their own, and modify these for their own analysis and plots
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular vegetation quantity, should match with the name of a particular model output variable.  Recommended to be alphanumeric because it is used to construct file names.}

\item{\code{name}}{A longer character string to provide a more complete description of this quantity}

\item{\code{type}}{A character string defining if this quantity is defined per PFT ("PFT"), per month ("monthly"), or something else.  The first two have a specific meaning to DGVMTools, but in principle the use can define anything.}

\item{\code{units}}{A character string defining the units this quantity is defined in.  Possibly formally link to udunits2?}

\item{\code{colours}}{A function that returns a colour scale suited for plotting this quantity.}

\item{\code{cuts}}{A numerical sequence defining the plot range and colour breakpoint when plotting the quantity}

\item{\code{aggregate.method}}{A character string defining the default method for how to aggregate the quantity, either "sum" or "average"}

\item{\code{model}}{Either a the string "Standard" to denotethat this is a  standard quantity to be compared across all model and data, or vector of model names to denote to which models this Quantity is applicable.}
}}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

