% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GFED4.R
\name{readGFED4}
\alias{readGFED4}
\title{Read the original GFED4 files}
\usage{
readGFED4(location = "/data/forrest/Fire/GFED4/",
  list.of.spatial.extents = NULL, temporal.extent = new("TemporalExtent", id
  = "GFED4", name = "Full GFED4 period", start = 1996, end = 2013),
  temporally.average = TRUE, spatially.aggregate = FALSE,
  spatial.resolution = "QD", temporal.resolution = "Annual",
  output.files = NULL, maxmemory = 1e+08, units = "fraction",
  archive = FALSE)
}
\arguments{
\item{location}{A character string specifying the the location of the original data files}

\item{list.of.spatial.extents}{A list of objects of type SpatialExtent, the funtion will return a dataset for each one.
This is intended so that one can retrieve time series for particular regions, for example the GFED regions as used in the GFED publications. 
If not specified the function just returns the global data.}

\item{temporal.extent}{A TemporalExtent object defining the time period for the data, the default are the currently available full monthly GFED4 
years (1996-2013)}

\item{temporally.average}{Logical, whether or not to temporal average the data to return yearly, monthly or daily averages (depending on the \code{temporal.resolution} 
argument)}

\item{spatially.aggregate}{Logical, whether or not to sun the burned area across the spatial domain}

\item{spatial.resolution}{Character string defing the desired spatial resolution, can be "QD" (0.25 degree, the original resolution) or "HD" (0.5 degree)}

\item{temporal.resolution}{Character string defing the desired temporal resolution, can be "Annual", "Monthly" or "Daily"}

\item{output.files}{A vector of character strings for the output files.  MOVE TO A DIFFERENT FUNCTION}

\item{maxmemory}{Integer to specify the maximum memory to use raster processing.  If you have a lot of memory, consider increasing this so that calculations 
are al done in memory for better performance.}

\item{units}{A string specifying the units in which to return the data: can be "fraction" (fractional area burnt, weighted by gridcell area but no by sea-land mask), 
"ha" (hectares) or "kmsq" (square kilometers)}

\item{archive}{Logical, whether or not to write the data to disk. MOVE TO A DIFFERENT FUNCTION.}
}
\value{
A raster (if spatial aggregating not requested) or a data.table (if spatially aggregating requested)
}
\description{
Read the original GFED4 data files, and optionally aggregates/averages and crops the data in time and space.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
