% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Globcover2009.R
\name{readGlobcover2009}
\alias{readGlobcover2009}
\title{Read the original Globcover2009 data}
\usage{
readGlobcover2009(location)
}
\arguments{
\item{location}{Character string specifying the directory in which the original data is stored}
}
\value{
DataObject object
}
\description{
This function simply returns the original data, but packaged into a DataObject object with metadata.  This data is at 300m and is categorical so processing it to more manageable resolutions is non-trivial and time consuming.  
Therefore, this function does not do this it just returns the original data (still stored on disk, not in memory).  If you want to process by 
by aggregeting to fractions of a larger, more workable resolution, please see \code{processGlobcover2009}
}
\seealso{
code{processGlobcover2009}, \code{countCategoricalData}
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
