% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-datasets.R
\name{readHandPBiomes}
\alias{readHandPBiomes}
\title{Read Haxeltine and Prentice 1996 PNV biome map}
\usage{
readHandPBiomes(resolution = "HD", classification = "Smith2014")
}
\arguments{
\item{resolution}{A character string specifying the resolution, currently supported
are "HD", "T63" and "original" (original is at half degree so equivalent to "HD")}

\item{classification}{A character string specifying how the biomes should be classified. 
Options are "original" (direct as per the H&P 1996 map), "Smith2014" (following Smith et al. 2014 and the ordering
following the figure in that paper) and "Forrest2015" (an aggregated megabiomes scheme based and Harrsion and Presntice 2006,
see the supplementary material in Forrest et al. 2015 for details and ordering).

This dataset has been used in various publications, but just because we include it here 
doesn't mean we specifically recommend it's use.  It is included here as an example, judgement of the scientific merit is left 
to the user.}
}
\description{
An example dataset of distribution of global potential natural vegetation classified into biomes.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
