% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hyde.R
\name{readHYDE}
\alias{readHYDE}
\title{Read HYDE land use dataset}
\usage{
readHYDE(location = "/home/forrest/Data/LandUseLandCover/HYDE/netCDF", layer,
  temporal.extent, verbose = FALSE)
}
\arguments{
\item{location}{A character string defining where on the disk the data is.}

\item{layer}{A character string, should be one of "CropFraction", "PastureFraction" or "PopulationDensity"}

\item{temporal.extent}{A TemporalExtent object to define the period you want the data for.  It will be averaged between these years.}

\item{verbose}{Logical, if TRUE spew forth a lot of output to the screen}
}
\value{
A DataObject
}
\description{
Reads either croplands fraction, pasture fraction or population density from the HYDE land use dataset (0.5 degree)
}
\details{
This functions reads some netCDF files (not the original HYDE data), produced by the author and available on request.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

