% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOut.ED2.r
\name{readOut.ED2}
\alias{readOut.ED2}
\title{Read ED2 data}
\usage{
readOut.ED2(path, sites, type, yearlim = "all", variables = "all",
  verbose = FALSE)
}
\arguments{
\item{path}{path to the h5 files generated by ED2}

\item{sites}{site(s) or prefix(es)}

\item{type}{Q, T, D, etc.}

\item{variables}{(optional) variables to read in, as defined by the *names* of edvars below (eg. NPP), if "all" then all of edvars will be read}

\item{verbose}{(optional) echo's some progress on screen}
}
\value{
list of the form ed$site$var
}
\description{
Read ED2 data
}
