% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readOut.LPJG.r
\name{readOut.LPJG}
\alias{readOut.LPJG}
\title{Read LPJ-GUESS output}
\usage{
readOut.LPJG(path.run, sites, delete_nye = FALSE)
}
\arguments{
\item{path.run}{Path to the LPJ-GUESS run, with all the output in a "output" subdirectory}

\item{sites}{Sites we are interested in}

\item{delete_nye}{[T/F] because there are some significant spikes on that day due to yearly allocation}
}
\value{
Dataframe with all output
}
\description{
Read LPJ-GUESS output
}
