% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVar.ED2.r
\name{readVar.ED2}
\alias{readVar.ED2}
\title{Read ED2 variables}
\usage{
readVar.ED2(path, sites, type, vars = "all", yearlim = "all",
  include_units = TRUE, verbose = FALSE)
}
\arguments{
\item{path}{path to the h5 files folder generated by ED2}

\item{sites}{site(s) or prefix(es)}

\item{type}{Q, T, D, etc.}

\item{include_units}{include units description?}

\item{verbose}{(optional) echo's some progress on screen}

\item{variables}{variables to read in, as defined by the names in the h5 files}
}
\value{
list of the form ed$site$var
}
\description{
Read ED2 variables
}
