% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{removeFromModelRun}
\alias{removeFromModelRun}
\title{Remove an object (either a \code{ModelObject} from a \code{ModelRun} object to reclaim memory.}
\usage{
removeFromModelRun(object.id, run)
}
\arguments{
\item{object.id}{The id of the object to remove  to the \code{ModelRun},}

\item{run}{The \code{ModelRun} object to which the object argument should be added}
}
\value{
A ModelRun with the object removed
}
\description{
Processing a lot of different files and storing the full output internally (in the ModelRun) may use too much memory. You can get around this by not storing them internally, 
but if you need to do this, for example because you want to average many spatial or temporal extents from one output file, this function lets you remove ModelObject to free the space again.
}
\seealso{
ModelRun, ModelRunInfo
}

