% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameLayers.R
\name{renameLayers}
\alias{renameLayers}
\title{Rename layers}
\usage{
renameLayers(x, old, new)
}
\arguments{
\item{x}{The Field object whose layers should be renamed.}

\item{old}{When \code{new} is provided, character names or numeric positions of column names to change. When \code{new} is not provided, the new column names, which must be the same length as the number of columns. See examples.}

\item{new}{Optional. New column names, must be the same length as columns provided to \code{old} argument. 
  
Syntax copied from data.table::setnames(), for which this function is essentially a wrapper.}
}
\value{
A Field (but not this is not strictly necessary since the Field object is changed in place)
}
\description{
Used to rename layers in a Field object.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
