% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample-methods.R
\docType{methods}
\name{resample}
\alias{resample}
\alias{resample.DGVM}
\alias{resample.DGVM,ModelObject,Raster,character,character-method}
\alias{resample.DGVM,DataObject,Raster,character,character-method}
\alias{resample.DGVM,ModelObject,character,character,character-method}
\title{Resample methods}
\usage{
resample.DGVM(x, y, method, filename)

\S4method{resample.DGVM}{ModelObject,Raster,character,character}(x, y, method,
  filename)

\S4method{resample.DGVM}{DataObject,Raster,character,character}(x, y, method,
  filename)

\S4method{resample.DGVM}{ModelObject,character,character,character}(x, y,
  method, filename)
}
\arguments{
\item{x}{A ModelObject or DataObject to be resample}

\item{y}{A raster define the rather grid for the filename}

\item{method}{The resampling method, can be "bil" or "nn" for raster.}

\item{filename}{If supplied, the resulting resampled raster is written to disk.}

\item{...}{Other arguments, not currently used}
}
\value{
A vector of character strings of the names of the layers
}
\description{
Methods for resampling ModelObjects and Data objects to other resolutions based on Raster* objects or netCDF files on disk 
(especially useful for guassian grids with irregularly spaced latitudes)
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
