% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-runtools.R
\name{sanitiseNamesForRaster}
\alias{sanitiseNamesForRaster}
\title{Make names appropriate for raster layers}
\usage{
sanitiseNamesForRaster(input)
}
\arguments{
\item{input}{A data.table, VegObject, Spatial*-object or vector of characters string to have the names (or itself in the case of a character string) 
sanitised to be suitable for use as Raster*object layer names. Also takes a Raster*-object, but this is return directly.}
}
\value{
An object of the same type as was put in.
}
\description{
This "sanaitises" names of an object for use as Raster*-object laey names.  This just involves replacing "-" with "." in the names of the object or the character strings.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

