% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-framework-functions.R
\name{santiseComparisonsForPlotting}
\alias{santiseComparisonsForPlotting}
\title{Sanitise input Comparisons for plotting}
\usage{
santiseComparisonsForPlotting(comparisons)
}
\arguments{
\item{comparisons}{The input to a plotXXXXComparison() functions to be checked}
}
\value{
Returns a list of DGVMTools::Comparison objects or NULL if a problem was found
}
\description{
This is an internal helper function which checks the inputs to a plotXXXXComparison() function (which should be a single Comparison or a list of Comparisons) 
and returns a list of Comparisons
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
