% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-framework-functions.R
\name{santiseDimensionsForPlotting}
\alias{santiseDimensionsForPlotting}
\title{Sanitise STAInfo for plotting}
\usage{
santiseDimensionsForPlotting(fields, require = NULL)
}
\arguments{
\item{fields}{The list of Fields to be plotted (should have been check by santiseFieldsForPlotting first)}
}
\value{
Returns character vector of the layers to be plotted
}
\description{
This is an internal helper function which checks the dimensions of the Fields to be plotted
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
