% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-framework-functions.R
\name{santiseLayersForPlotting}
\alias{santiseLayersForPlotting}
\title{Sanitise input layers for plotting}
\usage{
santiseLayersForPlotting(fields, layers)
}
\arguments{
\item{fields}{The list of Fields to be plotted (should have been check by santiseFieldsForPlotting first)}

\item{layers}{The layers requested to be plotted}
}
\value{
Returns character vector of the layers to be plotted
}
\description{
This is an internal helper function which checks the layers requested to be plotted against the layers in the the fields to be plotted.  If layers is NULL, then 
it returns all layers present in any fields
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
