% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-import.R
\name{seasonalityIndex}
\alias{seasonalityIndex}
\title{Calculates the seasonality index}
\usage{
seasonalityIndex(x, wgt = c(31, 28.2425, 31, 30, 31, 30, 31, 31, 30, 31, 30,
  31))
}
\arguments{
\item{x}{values, e.g. monthly precipitation values}

\item{wgt}{weight for the angle width of each value, e.g. days per month (default)}
}
\value{
the seasonality index following Markhan (1970)
}
\description{
Calculates a seasonality index based on a monthly vector by default, originally developed for precipitation. With specifying an additional wheighting vector different to months can be used.
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}
\references{
Charles G. Markhan (1970) Seasonality of precipitation in the United States, Annals of the Association of American Geographers, 60:3, 593-597, DOI: 10.1111/j.1467-8306.1970.tb00743.x
}

