% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectDays.R
\name{selectDays}
\alias{selectDays}
\title{Select days from a Field}
\usage{
selectDays(x, days)
}
\arguments{
\item{x}{The Field or data.table from which the days should be selected.  Note that a data.table should have a column called "Day".}

\item{days}{A numeric vector of the days to be selected}
}
\value{
A Field or data.table depending on the type of the input x.
}
\description{
Select days from a Field
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
