% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectMonths.R
\name{selectMonths}
\alias{selectMonths}
\title{Select months from a Field}
\usage{
selectMonths(x, months)
}
\arguments{
\item{x}{The Field or data.table from which the days should be selected.  Note that a data.table should have a column called "Month".}

\item{months}{A numeric vector of the months to be selected (Can also be abbreviations, "Jan", "Feb" etc... )}
}
\value{
A Field or data.table depending on the type of the input x.
}
\description{
Select months from a Field
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
