% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSeasons.R
\name{selectSeasons}
\alias{selectSeasons}
\title{Select seasons from a Field}
\usage{
selectSeasons(x, seasons)
}
\arguments{
\item{x}{The Field or data.table from which the seasons should be selected.  Note that a data.table should have a column called "Season".}

\item{seasons}{A character vector of the seasons to be selected, ie some of "MAM", "JJA", "SOM" and "DDF" (or potentionally a custom defined season)}
}
\value{
A Field or data.table depending on the type of the input x.
}
\description{
Select seasons from a Field
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
