% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectYears.R
\name{selectYears}
\alias{selectYears}
\title{Select years from a Field}
\usage{
selectYears(x, first, last)
}
\arguments{
\item{x}{The Field or data.table from which the years should be selected.  Note that a data.table should have a column called "Year".
should have columns "Lon" and "Lat" included.}

\item{first}{The first year to be selected (numeric)}

\item{last}{The last year to be selected (numeric)}
}
\value{
A Field or data.table depending on the type of the input x.
}
\description{
Select years from a Field
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
