% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.ED2.r
\name{sensitivity.ED2}
\alias{sensitivity.ED2}
\title{Sensitivity analysis ED2 (one dimensional version)
Note that this will look for T-output files only! This assumes a certain directory structure of the ED2 output as well:}
\usage{
sensitivity.ED2(param.var, param.eval, site, path, verbose = FALSE)
}
\arguments{
\item{param.var}{The parameter that is being varied (prefix of the folders)}

\item{param.eval}{The parameter(s) of which we want to see the response (ED2 output names)}

\item{path}{Path to the sensitivity analysis}
}
\value{
A sensitivity analysis?
}
\description{
/path/param.var_value/T-file-output.h5
}
