% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\docType{data}
\name{sep.char}
\alias{sep.char}
\title{Separation character for builidng variables names}
\format{An object of class \code{character} of length 1.}
\usage{
sep.char
}
\description{
Currently hard-coded to "", but could also be " ", "_" or "." (for example)
This should clearly be implemented in a more useful way...
}
\keyword{datasets}
\keyword{internal}
