% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome-schemes.R
\name{SimpleAdgvm2BiomeRules}
\alias{SimpleAdgvm2BiomeRules}
\title{Rules to classify coarse tropical biomes}
\usage{
SimpleAdgvm2BiomeRules(x)
}
\arguments{
\item{x}{Numerical vector of vegetation over values for a particular location. 
Certain fractions and quantities should have been pre-calculated.}
}
\value{
Biomes code (1-5, baren, C4 grassland, C3 grassland, woodland, forest)
}
\description{
No reference yet
}
\author{
Simon Scheiter \email{Simon.Scheiter@senckenberg.de}
}
\keyword{internal}
