% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome-schemes.R
\docType{data}
\name{SimpleHeightAdgvm2Biomes.scheme}
\alias{SimpleHeightAdgvm2Biomes.scheme}
\title{Meta-data describing a simple scheme for aDGVM2 output.}
\format{An object of class \code{BiomeScheme} of length 1.}
\usage{
SimpleHeightAdgvm2Biomes.scheme
}
\description{
Meta-data describing a simple scheme for aDGVM2 output.
}
\author{
Simon Scheiter \email{Simon.Scheiter@senckenberg.de}
}
\keyword{internal}
