% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome-schemes.R
\name{Smith2014BiomeRules}
\alias{Smith2014BiomeRules}
\title{Rules to classify biomes as per Smith et al. 2014}
\usage{
Smith2014BiomeRules(x)
}
\arguments{
\item{x}{Numerical vector of LAI values for a particular location. 
Certain fractions and quantities should have been pre-calculated.}
}
\value{
Biomes code (1-17, ordering as in Smith et al 2014 Figure)
}
\description{
Based on LAI only, see paper for details.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
