% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000rvc-classes.R
\docType{class}
\name{SpatialDataset-class}
\alias{SpatialDataset-class}
\title{Class to hold the data and metadata about spatial dataset}
\description{
Gathers up the meta-data and a raster layer for one particular layer of spatial environmental data.  Not intended for model output, rather for a dataset like a above ground biomass, or MODIS tree cover.
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular spatial dataset.  Recommended to be alphanumeric because it is used to construct file names.}

\item{\code{name}}{A character string to provide a fuller description of this dataset}

\item{\code{temporal.extent}}{A \code{TemporalExtent} oject to describe when this environmental data was measured.}

\item{\code{data}}{The actual data.  The slot type is currently "ANY" but perhaps it should be restricted to a \code{RasterLayer}.  That is what the benchmarking code certainly expects}

\item{\code{veg.quant}}{A \code{Quantity} object to describe exactly what this dataset is}

\item{\code{units}}{A character string defining the units in which this data is measured.  Probably obselete because of the units slot in the veg.quant slot}

\item{\code{correction.raster}}{A \code{RasterLayer} to apply multiplicatively to correct model data before comparing to this dataset.  For example a land-use correction.}
}}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

