% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-classes.R
\docType{class}
\name{SpatialExtent-class}
\alias{SpatialExtent-class}
\title{Spatial extent for defining the spatial area covered by model output or other spatial data.}
\description{
A simple S4 class to define a spatial extent.
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular spatial extent.  Recommended to be alphanumeric because it is used to construct file names.}

\item{\code{name}}{A character string to describe the spatial extent. Used for building plot labels, not file names, so doesn't need to be alphanumeric and can so can be prettier.}

\item{\code{extent}}{Currently can be a raster \code{Extent} object, but should be expanded to include a raster mask (to define an irregularly shaped spatial extent), or a c(Lon,Lat) coordinate to define a site.}
}}

\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
