% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-misc.R
\docType{data}
\name{standard.continental.extents}
\alias{standard.continental.extents}
\title{Continental extents}
\format{An object of class \code{list} of length 14.}
\usage{
standard.continental.extents
}
\description{
These were just defined by the author for studying different regions of the world.  Maybe also be handy for other people.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{datasets}

