% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Statistics-class}
\alias{Statistics-class}
\title{Comparison statistics}
\description{
This class stores the statistics resulting from the comparison of two layers from DGVMTools::Field objects.
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular raster compariosn.  Recommended to be alphanumeric because it is used to construct file names.}

\item{\code{R2}}{The R^2 between the two datasets}

\item{\code{R2.eff}}{The model efficiency of the two datasets}

\item{\code{P.cor}}{The Pearsons product moment correlation between the two datasets}

\item{\code{ME}}{The Mean Error (mean of residuals) between two datasets}

\item{\code{NME}}{The Normalised Mean Error (mean of residuals after the residuals have been divided the value of the data point) between two datasets}

\item{\code{NMSE}}{The Normalised Square Mean Error (mean of the square of the residuals after the residuals have been divided the value of the data point) two datasets}

\item{\code{RMSE}}{The Root Mean Squared Error between the two datasets}

\item{\code{sd.diff}}{The standard deviation of the difference between the two datasets (all gridcells)}

\item{\code{MM}}{The Manhattan Metric (for comparisons of relative proportions)}

\item{\code{SCD}}{The Manhattan Metric (for comparisons of relative proportions)}

\item{\code{Kappa}}{The overall Cohen's Kappa obtained when comparing categorical variables.}

\item{\code{individual.Kappas}}{The individual Cohen's Kappas for each category when comparing categorical data}
}}

\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
