% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-misc.R
\name{subsetGridlist}
\alias{subsetGridlist}
\title{Select a subset of a gridlist}
\usage{
subsetGridlist(gridlist.file, subset.extent, file.name = NULL,
  header = TRUE, offset = c(0.25, 0.25))
}
\arguments{
\item{gridlist.file}{A character string giving the location of the original gridlist}

\item{subset.extent}{A raster Extent object specifying the geograpical sub-domain required}

\item{file.name}{A character string specifying a path to write the new gridlist (can be ignored to write no file)}

\item{header}{Logical, whether or not the original file has a header}

\item{offset}{A two-member numeric vector specifying the longitude-latitude offset from the given coordinates 
to the gridcell centre, ie. should be c(0.25,0.25) for old LPj-GUESS gridlists, and c(0.0,0.0) for most other more sensible scenarios}
}
\value{
The new gridlist as a data.frame
}
\description{
Crop a gridlist (taken as a file in the form of a two-columned table to a smaller subset provided by an Extent object
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
