% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-benchmarking.R
\name{summariseRasterComparisons}
\alias{summariseRasterComparisons}
\title{Compare many VegRuns to a spatial dataset}
\usage{
summariseRasterComparisons(runs, dataset, label = NULL, diff.cuts = NULL,
  perc.diff.cuts = seq(-100, 200, 5), spatial.extent = NULL,
  showR2 = TRUE, layout.objs = NULL, ...)
}
\arguments{
\item{runs}{A list of VegRun objects (each with the RasterComparison object already created)}

\item{dataset}{The spatial dataset to which we are comparing (represented as a SpatialDataset object)}

\item{label}{A string with which to tag the resulting plots to specify the analysis (to differentiate them from other plots),
for example "ForcingDatasetComparison" or "SoilDepths.  Or whatever you want.}

\item{diff.cuts}{Numeric vector of cuts for plotting the absolute difference.  If not specified it is derived as the maximum
possible range based on the VegQuant object in the dataset.}

\item{perc.diff.cuts}{Numerice vector of cuts for plotting the percentage differences.  If not specified defaults to -100\% to +200\%
in 5\% intervals.}

\item{spatial.extent}{If specified (as a raster::Extent object), the spatial extent to plot.  Note that the statistics will *not* reflect
this extent}

\item{showR2}{A logical, if TRUE, put the R squared values on the plots.}

\item{layout.objs}{A list of layout objects (see the \code{sp} package) for putting on the spatial plots.  For example,
continent outlines}

\item{...}{Arguments to be passed to plotVegMaps when making the spatial plots.}
}
\value{
No return, just makes plots.
}
\description{
To use this function it is *required* that \code{compareRunToSpatialDataset()} has been performed on each of the runs with 
the dataset to which we are comparing.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

