% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000rvc-tools.R
\docType{methods}
\name{Summary-methods}
\alias{Print}
\alias{Summary-methods}
\alias{VegRun}
\alias{a}
\alias{object}
\alias{of}
\alias{summary}
\alias{summary,VegRun-method}
\title{Summary methods}
\usage{
summary(object, ...)

\S4method{summary}{VegRun}(object, ...)
}
\arguments{
\item{object}{a RVCTools object}

\item{...}{Other arguments, not currently used}
}
\value{
Not sure
}
\description{
Print easy to read summaries of RVCTool objects
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

