% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-classes.R
\docType{data}
\name{support.veg.models}
\alias{support.veg.models}
\title{Supported Veg Models}
\format{A list of character strings for each of the supported vegetation models}
\usage{
support.veg.models
}
\description{
List of supported vegetatiopn models
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{datasets}

