% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-classes.R
\docType{data}
\name{supported.models}
\alias{supported.models}
\title{Supported Veg Models}
\format{A list of character strings for each of the supported vegetation models}
\usage{
supported.models
}
\description{
List of supported vegetation models
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
