% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000rvc-classes.R
\docType{class}
\name{TemporalDataset-class}
\alias{TemporalDataset-class}
\title{Class to hold the data and metadata about temporal dataset}
\description{
Gathers up the meta-data and data for one particular layer of temporal environmental data, for eample a burnt area times series a flux tower measurement.  Not intended for model output, rather for a dataset like a above ground biomass, or MODIS tree cover.
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular spatial dataset.  Recommended to be alphanumeric because it is used to construct file names.}

\item{\code{name}}{A character string to provide a fuller description of this dataset}

\item{\code{temporal.extent}}{A \code{TemporalExtent} object to describe the length of the data times series}

\item{\code{data}}{The actual data.  The slot type is currently "ANY" but perhaps it should be restricted to a \code{RasterLayer}.  That is what the benchmarking code certainly expects}

\item{\code{extent}}{A \code{SpatialExtent} object define the spatial region over which this time series data applies.}

\item{\code{veg.quant}}{A \code{Quantity} object to describe exactly what this dataset is}

\item{\code{units}}{A character string defining the units in which this data is measured.  Probably obselete because of the units slot in the veg.quant slot}
}}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

