% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-classes.R
\docType{class}
\name{TemporalExtent-class}
\alias{TemporalExtent-class}
\title{Temporal extent for defining the duration of a model run and selecting time periods}
\description{
A simple S4 class to define a time period ie 1961-1990, but called here \code{TemporalExtent} to be analagous to \code{SpatialExtent} and distinct from the \code{Period} class (which contains months and seasons)
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular time period.  Recommended to be alphanumeric because it is used to construct file names.}

\item{\code{name}}{A character string to describe the time period. Used for building plot labels, not file names, so doesn't need to be alphanumeric and can so can be prettier.}

\item{\code{start}}{First year of the temporal extent (a year, eg 1961)}

\item{\code{end}}{Last year of the temporal extent (a year, eg 1990)}
}}

\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
