% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridarea.R
\name{thornthwaite}
\alias{thornthwaite}
\title{calculate monthly PET, AET, WC, WD based on Thornthwaite}
\usage{
thornthwaite(run = NULL, T = NULL, P = NULL, variable = "PET",
  as = "ModelObject")
}
\arguments{
\item{run}{the ModelRun object, if desired (see 'as', below).}

\item{T}{monthly temperature (degree Celcius )ModelObject, data.table or data.frame}

\item{P}{monthly precipitation in mm. Needed for everything without PET.}

\item{variable}{any of PET (potential evapotranspiration), AET (actual evapotranspiration), WC (water content), WD (water deficit)}

\item{as}{desired output type 'data.table', 'data.frame', anything else will try to return a ModelObject, which need 'run' to be defined.}
}
\value{
the requested variable in the requested ('as') container
}
\description{
calculate monthly PET, AET, WC, WD based on Thornthwaite
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}
\references{
Bonan, G. (2002): Ecological Climatology. Chapter 5 (Hydrological Cycle), Cambridge University Press, 690p. (http://www.cgd.ucar.edu/tss/aboutus/staff/bonan/ecoclim/index-2002.htm)
}

