% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-lpj.R
\name{translateLonLatAndCFIndex}
\alias{translateLonLatAndCFIndex}
\title{Returns}
\usage{
translateLonLatAndCFIndex(lon.lat = NULL, cf.index = NULL, lon.res = 0.5,
  lat.res = 0.5, lon.start = -180, lat.start = -90, southwest = FALSE)
}
\description{
Returns
}
