% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-framework-functions.R
\name{trimFieldsForPlottingXY}
\alias{trimFieldsForPlottingXY}
\title{Subsets data from Field for XY plotting}
\usage{
trimFieldsForPlottingXY(
  fields,
  layers,
  years = NULL,
  days = NULL,
  months = NULL,
  seasons = NULL,
  gridcells = NULL,
  dropEmpty = FALSE
)
}
\arguments{
\item{fields}{The list of Fields to be plotted (should have been check by santiseFieldsForPlotting first)}

\item{layers}{A character vector of the layers to be plotted}

\item{years}{The years to be extracted (as a numeric vector), if NULL all years are used}

\item{days}{The days to be extracted (as a numeric vector), if NULL all days are used}

\item{months}{The months to be extracted (as a numeric vector), if NULL all months are used}

\item{seasons}{The months to be extracted (as a character vector), if NULL all seasons are used}

\item{gridcells}{The months to be extracted (as a character vector), if NULL all seasons are used}

\item{dropEmpty}{Logical, if TRUE drop layers consisting only of zeros}
}
\value{
Returns a list of Fields
}
\description{
This is an internal helper function which pulls out the data needed to make a plot from a bunch of Fields, and returns 
a list of the Field with only the required layers and points in space and time included
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
