% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-lpj.R
\name{trimLPJFilename}
\alias{trimLPJFilename}
\title{Helper function to trim the ".out" from an LPJ-GUESS filename}
\usage{
trimLPJFilename(var.filename)
}
\arguments{
\item{var.filename}{The string of the filename to be trimmed}
}
\value{
A string less the last four characters.
}
\description{
Very simple, just removes the last four characters. 
No error checking to see if the last four characters are really ".out"!
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}

