% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000rvc-classes.R
\docType{class}
\name{VegObject-class}
\alias{VegObject-class}
\title{Contains one aspect of the model output, eg. LAI}
\description{
A key class of the package.  A \code{VegObject} stores the data and metadata for one quantity that comes from a vegetation model run (including information about the run iself).
For example LAI (Leaf Area Index), or monthly evapotranspiration.  The data can be averaged spatially or temporally, or neither ot both, and manipulated and plotted by mayn funtions in this package.
}
\details{
Generally these are not created directly by the user, but rather by functions like \code{getVegObject}.
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular vegetation object.  Recommended to be alphanumeric because it is used to construct file names.}

\item{\code{data}}{A data.table object.  This is used because is it very much faster for calculations that data.frame or raster layers.}

\item{\code{quant}}{A VegQuant object to define what output this VegObject contains}

\item{\code{spatial.extent}}{A SpatialExtent object which describes the area covered by this VegObject.  Particularly useful if the data has been spatially averaged.}

\item{\code{temporal.extent}}{A TemporalExtent object which describes the time periog covered by this VegObject.  Particularly useful if the data has been temporally averaged.}

\item{\code{is.site}}{Set to TRUE is this VegObject describes a single site}

\item{\code{is.spatially.averaged}}{Set to TRUE is this VegObject has been spatially averaged}

\item{\code{is.temporally.averaged}}{Set to TRUE is this VegObject has been temporally averaged}

\item{\code{run}}{A VegRunInfo object which contains the metadata about the run which this VegObject belongs too.}
}}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

