% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000rvc-classes.R
\docType{class}
\name{VegRun-class}
\alias{VegRun-class}
\title{An S4 class to contain metadata and, optionally, data and benchmarks for a single vegetation run}
\description{
A \code{VegRun} object contains the metadata concerning the an inherited \code{VegRunInfo} object
 and the actual model data run as \code{VegObjects} in a list in slot \code{objects} and comparisions to datasets 
 as \code{BiomeComparison} and \code{RasterComparison} in a list in slot \code{benchmarks}.
Such objects can be built by calls to \code{getVegObject()}, \code{getVegSpatial()}, \code{getVegTemporal()}, \code{calcNewVegObj}, \code{compareRunToSpatialDataset()}
 and \code{compareBiomes()}, and saved to the \code{VegRun} using \code{addToVegRun()}. 
 
Slots can be accessed by user directly, but more easily and usefully by functions \code{XXXX}
}
\section{Slots}{

\describe{
\item{\code{objects}}{List of \code{VegObjects} saved in this run}

\item{\code{benchmarks}}{List of benchmarks (\code{BiomeComparisons} and \code{RasterComparisons}) performed and saved for this run.}
}}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

