% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000rvc-classes.R
\docType{class}
\name{VegRunInfo-class}
\alias{VegRunInfo-class}
\title{Class to hold the metadata for a vegetation model run}
\description{
This class describes a vegetation run, including its location on disk, the model used, the PFT set used, an unique id and a description, offsets to apply to the longitudes and latitudes to make the co-rordinates gridcell centered and so on.
It is not primarily intended to be used by itself. Instead it is inherited by \code{VegRun} object (due to this inheritance the slots can be accessed directly)
and included in a \code{VegObject} in the \code{run} slot (not inherited, so needs to be access be \code{@run}).
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular model un.  Recommended to be alphanumeric because it is used to construct file names. (Mandatory)}

\item{\code{model}}{A character string to identify what model produced this run.  Can currently be "LPJ-GUESS", "LPJ-GUESS-SPITFIRE" or "aDGVM". (Mandatory)}

\item{\code{pft.set}}{A list of PFT objects which includes all the PFTs used is this model run (Mandatory)}

\item{\code{description}}{A character string describing this run, ie. "LPJ-GUESS v3.1"}

\item{\code{run.dir}}{The location of this run on the file system (Mandatory)}

\item{\code{driving.data}}{A character string identifying the climate or other data used to produce this model run}

\item{\code{map.overlay}}{A character defining which map overlay to plot as standard on for this model run.  Can be:
 \itemize{
  \item \code{hires} for a full high resolution outline of all countires of the world 
  \item \code{lowres} for a low resolution outline of all countires of the world  (much faster to plot that \code{hires}, generally sufficient for global plot)
  \item \code{hires-continents} for a full high resolution outline of the land masses only
  \item \code{lowres-continents} for a low resolution outline of the land masses only
}
These data come from the \code{mapdata} package}

\item{\code{lonlat.offset}}{A numeric of length 1 or 2 to define the offsets to Lon and Lat to centre the modelled localities.}

\item{\code{year.offset}}{A numeric of length 1 to match be added to the simulation years to convert them to calendar years}

\item{\code{tolerance}}{The tolerance arguement when converting uneven spaced grids to regular rasters for plotting}

\item{\code{london.centre}}{If TRUE, ensure that the longitudes are (-180,180) instead of (0,360)}

\item{\code{fill.col}}{A string to define an R colour used when plotting this run as a histogram or scatter plot or so}

\item{\code{line.col}}{A string to define an R colour used when plotting this runs as a line graph}

\item{\code{line.width}}{A numeric to define the width of a line representing this model run}

\item{\code{line.type}}{A numeric to define the line style representing this model run}

\item{\code{landuseSimulated}}{If TRUE it can be assumed that land use has been simulated for this run and so no correction for land use need be applied before benchmarking.}
}}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

