% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichLayers.R
\name{whichLayers}
\alias{whichLayers}
\title{Select PFTs}
\usage{
whichLayers(x, criteria = NULL, return.ids = TRUE)
}
\arguments{
\item{x}{Either a Field, or a list of PFT objects}

\item{criteria}{A character string to use to select PFTs.  This compared to every the value of every slot of a PFT object.  If one matches, the PFT is inlucded in the return.}

\item{return.ids}{A logical, if TRUE (default) then return the the id of the PFT, if FALSE return the entire PFT object}
}
\value{
Either a vector of characters (the ids of the PFTs) or a list of PFT object (depending on argument return.ids)
}
\description{
Uses PFT metadata to return all PFTs with a particular properties.  For example only tree PFTs, or only evergreen PFTs.
}
\examples{

## List PFTs from a list of PFT objects

# make a list of PFTs for selecting from 
PFT.list <- GUESS@default.pfts
print(PFT.list)

# return with no criteria (trivial)
whichLayers(PFT.list)

# return as the full PFT objects
whichLayers(PFT.list, return.ids = FALSE)

# return Tree, Grass, Evergreen and Tropical PFTs
whichLayers(PFT.list, criteria = "Tree")
whichLayers(PFT.list, criteria = "Grass")
whichLayers(PFT.list, criteria = "Evergreen")
whichLayers(PFT.list, criteria = "Tropical")

# return Tropical PFTs as a list of PFTs objects (not just the ids)
whichLayers(PFT.list, criteria = "Tropical", return.ids = FALSE)

\donttest{
 
## List PFTs from a Field object

# Load a per-PFT Field from the example data
run.dir <- system.file("extdata", "LPJ-GUESS_Runs", "CentralEurope", package = "DGVMTools")
test.Source <- defineSource(id = "LPJ-GUESS_Example", dir = run.dir,  format = GUESS)
test.Field.perPFT <- getField(source = test.Source, var = "lai") 

# Perform the same examples as above but with the Field instead of a list of PFTs objects

# return with no criteria (trivial)#' 
whichLayers(test.Field.perPFT )

# return as the full PFT objects
whichLayers(test.Field.perPFT, return.ids = FALSE)

# return Tree, Grass, Evergreen and Tropical PFTs
whichLayers(test.Field.perPFT, criteria = "Tree")
whichLayers(test.Field.perPFT, criteria = "Grass")
whichLayers(test.Field.perPFT, criteria = "Evergreen")
whichLayers(test.Field.perPFT, criteria = "Tropical")

# return Tropical PFTs as a list of PFTs objects (not just the ids)
whichLayers(test.Field.perPFT, criteria = "Tropical", return.ids = FALSE)

}

}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
