% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-netcdf_joerg.R
\name{write.nc}
\alias{write.nc}
\title{Create a NetCDF4 file from a \code{\linkS4class{Field}}.}
\usage{
write.nc(filename = NA, mo = NA, columns = NA, as.flux = FALSE,
  fill.value = FALSE, invert.lat = FALSE, globalAttr = NULL,
  compress = NA, chunks = NA, reduce = FALSE, leap = FALSE,
  start.year = NA, time.unit = "days", verbose = FALSE)
}
\arguments{
\item{filename}{output file name}

\item{mo}{the \code{\linkS4class{Field}}.}

\item{columns}{which colums to write as variable (TODO: option as dimension)}

\item{as.flux}{Converts the data to fluxes by dividing by the respective number of days, if TRUE or "day[s]". If set to "^s*", the divided by 86400 additionally.}

\item{fill.value}{value to use for _FillValue variable attribute.}

\item{invert.lat}{latitude dimension starts in the north.}

\item{globalAttr}{a named vector of additional global attributes.}

\item{compress}{should NetCDF4 compression be turned on. Either TRUE (compression level 4) of an integer between 1 and 9.}

\item{chunks}{chunk size for faster readability. If set make sure its length agrees with the data dimensions.}

\item{reduce}{reduced grid, only valid points are written. Spatial dimensions are reduced to one landID (TODO, currently not implemented).}

\item{leap}{turn leap years on (default off; TODO, currently not implemented).}

\item{start.year}{start year for units attribute of time axis.}

\item{time.unit}{unit for time axis units attribute, so far "days", "months" and "years" are valid. Default: "days".}

\item{verbose}{print some information}
}
\description{
Create a NetCDF4 file from a \code{\linkS4class{Field}}.
}
