% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-netcdf.R
\name{writeNetCDF}
\alias{writeNetCDF}
\title{Write a netCDF file}
\usage{
writeNetCDF(x, filename, start.date = NULL, monthly = FALSE,
  verbose = FALSE, ...)
}
\arguments{
\item{filename}{A character string (including the path) specifiying the name of the netCDF file.}

\item{data.in}{The data as a Raster* object.  This should be broadened to also include data.frames and data.tables for writing 
time series from a particular point.}

\item{var.name}{A character string containing the name of the variable as used in the netCDF file produced.}

\item{var.units}{A character string containing the units of the variable as used in the netCDF file produced.}

\item{time.resolution}{A character string denoting the time resolution of the data. Currently can be "monthly" or "annual".}

\item{time.units.string}{A string to represent the time units.  CHECK THIS.}

\item{time.vals}{A list of strings providing the values for the time axis.}

\item{long.name}{A charcter string for the "long_name" of the netCDF file.}

\item{ordering}{A character string specifying the ordering of the dimension in the resulting netCDF file, can be "standard" for normal ordering 
or "lpj" for ordering for fast LPJ-GUESS reading.}

\item{missing.value}{A numeric value for the "missing_value" of the netCDF file.}
}
\value{
Noting, writes a netCDF file to disk
}
\description{
This function gives more flexibility (for example it can write more meta-data and can re-order the dimensions for efficient LPJ-GUESS reading) 
than the similar raster::\code{writeRaster} function.  It is also intended that this function can write single sites/time series functionality 
is not included yet.  Also might need some work to be CF compliant on the time dimension.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
