% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridarea.R
\name{addArea}
\alias{addArea}
\title{Adds the gridcell area to a spatial Field or data.table/data.frame}
\usage{
addArea(input, unit = "m^2", ellipse = FALSE, verbose = TRUE, digits)
}
\arguments{
\item{input}{a spatial Field or a data.frame/data.table with at least the columns Lon and Lat.}

\item{unit}{area unit. Default m^2, if something else is spefified udunits2 must be installed. If udunits2 is not installed this option is ignored.}

\item{ellipse}{If the eath should be assumed to be a ellipsoid instead of a sphere.}

\item{verbose}{print some information.}

\item{digits}{Numeric, number of digits to which to truncate the coordinates when merging the area data.table with the input. This is a technical detail,
you only need to use it if you have troubles because of coordinates with a few decimal places.}
}
\value{
same class as input
}
\description{
Adds the gridcell area to a spatial Field or data.table/data.frame.
For unit conversion it makes use of \code{\link[udunits2]{ud.convert}}, if installed. If udunits2 is not installed the option is ignored and the returned unit is m^2.
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}
