% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvc-gridarea.R
\name{addArea}
\alias{addArea}
\title{addArea: Adds the gridcell area to a spatial ModelObject or data.table/data.frame}
\usage{
addArea(input, unit = "m^2", ellipse = FALSE, verbose = TRUE)
}
\arguments{
\item{input}{a spatial ModelObject or a data.frame/data.table with at least the columns Lon and Lat.}

\item{unit}{area unit. Default m^2, if something else is spefified udunits2 must be installed. If udunits2 is not installed this option is ignored.}

\item{ellipse}{If the eath should be assumed to be a ellipsoid instead of a sphere.}

\item{verbose}{print some information.}
}
\value{
same class as input
}
\description{
Adds the gridcell area to a spatial ModelObject or data.table/data.frame
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

