% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addStandardSpatialAttributes.R
\name{addStandardSpatialAttributes}
\alias{addStandardSpatialAttributes}
\title{Adds standard spatial attributes to a netCDF file}
\usage{
addStandardSpatialAttributes(nc.file)
}
\arguments{
\item{nc.file}{The netcdf object to which the attributes should be added}
}
\value{
A netcdf object
}
\description{
A very simple convenience function for filling the the standard Lon/Lat info when preparing a netCDF file
}
