% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTo.R
\name{addTo}
\alias{addTo}
\title{Add \linkS4class{Layer}/\linkS4class{Quantity} to a \linkS4class{Format}/\linkS4class{Source}/\linkS4class{Field}}
\usage{
addTo(x, add.to)
}
\arguments{
\item{x}{The \linkS4class{Layer}/\linkS4class{Quantity} to be added}

\item{add.to}{The \linkS4class{Format}/\linkS4class{Source}/\linkS4class{Field}}
}
\value{
The returned oobject will be or an updated \linkS4class{Format}/\linkS4class{Source}/\linkS4class{Field} 
object if "add.to" was specified.
}
\description{
Adds metadata encapsulated in a \linkS4class{Layer} or \linkS4class{Quantity}
}
\seealso{
\linkS4class{Layer}, \linkS4class{Quantity}, \link{defineQuantity}, \link{defineLayer}
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
