% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{addToModelRun}
\alias{addToModelRun}
\title{Add an object (either a \code{ModelObject} or a  \code{SpatialComparison}) to a \code{ModelRun} object to be use later.}
\usage{
addToModelRun(object, run)
}
\arguments{
\item{object}{Object to add to the \code{ModelRun},   The parameters are the slots of a ModelRunInfo object. Note that that \code{id} and \code{run.dir} are compulsory, the rest will be filled with dummy/default values if left blank.
compulsory.  Take care with \code{lon.lat.offset} and \code{year.offset} which are initialised to 0 which is unsuitable for some LPJ-GUESS confgurations.}

\item{run}{The \code{ModelRun} object to which the object argument should be added}
}
\value{
A ModelRun object the the object argument added
}
\description{
Stores an object in its run for later calculations, plotting, comparisons.
}
\seealso{
ModelRun, ModelRunInfo
}

