% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateSpatial.R
\name{aggregateSpatial.uncompiled}
\alias{aggregateSpatial.uncompiled}
\title{Spatially aggregate a Field}
\usage{
aggregateSpatial.uncompiled(x, method = "mean", verbose = FALSE, ...)
}
\arguments{
\item{x}{Field (or data.table) to be averaged}

\item{method}{The method with which to spatially aggregate.  Can be "weighted.mean", "w.mean", "mean", 
"weighted.sum", "w.sum", "sum", "mode", "median", "max", "min", "sd", "var" and "cv" (= coefficient of variation: sd/mean).}

\item{verbose}{If TRUE give some progress update about the averaging.}

\item{...}{Extra arguments passed to addArea function if a weighted method is being used. Note in particular the lon_centres and lat_centres arguments
if you are using a regular but sparsely populated grid.}
}
\value{
A Field or data.table depending on the input object
}
\description{
Spatially aggragte all gridcells of a Field (or a data.table assuming it has columns "Lon" and "Lat"). 
Can use area-weighting to take into account the different areas of gridcells (even though they are constant in Lon, Lat)
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
