% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateSpatial.R
\name{aggregateSpatial.uncompiled}
\alias{aggregateSpatial.uncompiled}
\title{Spatially average a Field, DataObject or data.table}
\usage{
aggregateSpatial.uncompiled(input.obj, method = "mean",
  verbose = FALSE, ...)
}
\arguments{
\item{input.obj}{Field, DataObject or data.table to be averaged}

\item{method}{The method with which to spatially aggregate.  Can be "weighted.mean", "w.mean", "mean", 
"weighted.sum", "w.sum", "sum", "max", "min", "sd" or "var".}

\item{verbose}{If TRUE give some progress update about the averaging.}

\item{...}{Extra arguments passed to addArea function if a weighted method is being used.}
}
\value{
A Field, DataObject or data.table depending on the input object
}
\description{
Spatially average all gridcells of a Field, DataObject or data.table (assuming the data.table has columns "Lon" and "Lat"). 
Can use area-weighting to take into account the difference different areas of gridcells (even though they are constant in Lon,Lat)
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
