% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateSubannual.R
\name{aggregateSubannual}
\alias{aggregateSubannual}
\title{Sub-annual aggregation}
\usage{
aggregateSubannual(input.obj, method = "mean", target = "Year",
  verbose = FALSE)
}
\arguments{
\item{input.obj}{data.table or Field}

\item{method}{A character string describing the method by which to aggregate the data.  Can currently be "mean", "sum", "max", "min", "sd" and "var".
For technical reasons these need to be implemented in the package in the code however it should be easy to implement more, please just contact the author!}

\item{target}{A character string defining the subannual period to which the data should be aggregate. Can be "Month", "Season" or "Year" (also "Annual" is valid)  (default)}

\item{verbose}{If TRUE give some progress update about the averaging.}
}
\value{
A Field or data.table depending on the input object
}
\description{
Aggregates data with sub-annual time resultion to a coarser time resolution.  For example, going from monthly to annual.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
