% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateYears.R
\name{aggregateYears.uncompiled}
\alias{aggregateYears.uncompiled}
\title{Aggregate years}
\usage{
aggregateYears.uncompiled(input.obj, method = "mean", verbose = FALSE)
}
\arguments{
\item{input.obj}{A Field or data.table (with a "Year" column)}

\item{method}{A character string describing the method by which to aggregate the data.  Can currently be "mean", "sum", "max", "min", "sd" and "var".
For technical reasons these need to be implemented in the package in the code however it should be easy to implement more, please just contact the author!}

\item{verbose}{If TRUE give some progress update about the averaging.}
}
\value{
A Field or data.table depending on the input object
}
\description{
Aggregated all available years in a Field or a data.table object (with years denoted by column "Years").
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
