% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-aDGVM.R
\name{availableQuantities_aDGVM}
\alias{availableQuantities_aDGVM}
\title{List aDGVM quantities avialable}
\usage{
availableQuantities_aDGVM(source)
}
\arguments{
\item{source}{A path to a directory on the file system containing some .out files}
}
\value{
A list of all the .out files present, with the ".out" removed. 

Needs to be implemented by an aDGVMer
}
\description{
Simply lists all quantitied aDGVM  output variables
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
