% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageFields.R
\name{averageFields}
\alias{averageFields}
\title{Average Fields provided as a list}
\usage{
averageFields(list.of.fields, run = NULL, method = mean)
}
\arguments{
\item{list.of.fields}{The Fields that you want to average, stored in a list.}

\item{run}{Either a Source object which provides the metadata about the newly created "run" (representing, say, the ensemble mean) or \code{NULL}, 
in which case the function just returns a data.table with the average values.}

\item{method}{The function which you want to use.  Is \code{mean} by default, but it can be anything that operates on a vector of numbers.
Useful options might be \code{max},  \code{min} and \code{sd}.}
}
\value{
Either a Field (if a Source has been supplied to provide the necessary meta-data), or a data.table (if no meta-data has been supplies in the form of a Source)
}
\description{
Useful for averaging model ensembles, replicate simulations etc.  
Note that since you can supply the function, you don't actually need to average, you can also do, for example, \code{max}, \code{min}, \code{sd}, etc.  
Fails with an error message if the supplied Fields don't have the same fields
}
\details{
The function currently does no checks that the spatial extent, yearly period and Quantity are consistent between runs.  The resulting Field simple takes these things
from the first object from list.of.vegruns.  It is therefore up to the user not do anything silly and average quantites, or spatial or temporal extents that don't makes sense.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
