% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-benchmarking.R
\name{benchmarkSpatial}
\alias{benchmarkSpatial}
\title{Benchmark VegRuns against a spatial dataset}
\usage{
benchmarkSpatial(runs, layer.name, dataset, tag = NULL, diff.cuts = NULL,
  perc.diff.cuts = seq(-100, 200, 5), spatial.extent = NULL,
  showR2 = TRUE, layout.objs = NULL, histo.plot.range = NULL,
  ignore.raster = NULL, summary.plot.dir = NULL, canvas.options = list(dpi
  = 72, type = "png", width = 1800, height = 1000, bg = "transparent"), ...)
}
\arguments{
\item{runs}{A list of VegRun objects (each with the RasterComparison object already created)}

\item{layer.name}{Layer in the vegetation model output to compare to the data}

\item{dataset}{The spatial dataset to which we are comparing (represented as a SpatialDataset object)}

\item{tag}{A string with which to tag the resulting plots to specify the analysis (to differentiate them from other plots),
for example "ForcingDatasetComparison" or "SoilDepths.  Or whatever you want.}

\item{diff.cuts}{Numeric vector of cuts for plotting the absolute difference.  If not specified it is derived as the maximum
possible range based on the VegQuant object in the dataset.}

\item{perc.diff.cuts}{Numerice vector of cuts for plotting the percentage differences.  If not specified defaults to -100\% to +200\%
in 5\% intervals.}

\item{spatial.extent}{If specified (as a raster::Extent object), the spatial extent to plot.  Note that the statistics will *not* reflect
this extent}

\item{showR2}{A logical, if TRUE, put the R squared values on the plots.}

\item{layout.objs}{A list of layout objects (see the \code{sp} package) for putting on the spatial plots.  For example,
continent outlines}

\item{histo.plot.range}{A two-value numeric vector defining the range for the histo plots}

\item{ignore.raster}{A raster used to mask out gridcells wich are not to be compared (ie any gridcells set to NA in the ignore.raster
are set to NA in thr model and data rasters)}

\item{summary.plot.dir}{A directory (full path as a character string) so save the plots which compare many runs}

\item{canvas.options}{A list of options (to be given to the \code{Cairo}) function to define the canvas. See the \code{Cairo} dicumentation}

\item{...}{Arguments to be passed to plotVegMaps when making the spatial plots.}
}
\value{
No return, just makes plots.
}
\description{
To use this function it is *required* that \code{compareRunToSpatialDataset()} has been performed on each of the runs with 
the dataset to which we are comparing.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

