% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification-schemes.R
\docType{data}
\name{Smith2014BiomeScheme}
\alias{Smith2014BiomeScheme}
\alias{Hickler2012BiomeScheme}
\alias{Forrest2015BiomeScheme}
\alias{MeditBiomeScheme}
\alias{DevMegaBiomeScheme}
\alias{FireMIPBiomeScheme}
\alias{FPCMegaBiomeScheme}
\alias{aDGVMBiomeScheme}
\alias{SimpleAdgvm2BiomeScheme}
\alias{SimpleHeightAdgvm2BiomeScheme}
\alias{GrowthFormAdgvm2BiomeScheme}
\alias{PhenologyAdgvm2BiomeScheme}
\title{Smith et al 2014
 
\strong{Smith2014BiomeScheme} Classification of global output as per Smith et al. 2014. Designed to be compared to Haxeltine and Prentice (1996) 
map of potential natural biomes.}
\format{An object of class \code{Scheme} of length 1.}
\usage{
Smith2014BiomeScheme

Hickler2012BiomeScheme

Forrest2015BiomeScheme

MeditBiomeScheme

DevMegaBiomeScheme

FireMIPBiomeScheme

FPCMegaBiomeScheme

aDGVMBiomeScheme

SimpleAdgvm2BiomeScheme

SimpleHeightAdgvm2BiomeScheme

GrowthFormAdgvm2BiomeScheme

PhenologyAdgvm2BiomeScheme
}
\description{
\strong{Hickler2012BiomeScheme}  Classification of european output as per Hickler et al. 2012.

\strong{Forrest2015BiomeScheme}  Classification of global output into coarser mega biomes as per Forrest et al. 2015 (see appendix), 
 which is really just an aggregation of biomes following Harrison and Prentice 2006.  Can be compared to aggregated Haxeltine and Prentice (1996) map.

\strong{MeditBiomeScheme} A biome classification scheme used for Turkey.

\strong{DevMegaBiomeScheme}  Unpublished for potentially useful classification of global biomes which is slightly simpler than Smith et al. 2014.

\strong{FireMIPBiomeScheme}  Simple biomes based on a common PFT set for the FireMIP models

\strong{FPCMegaBiome} an unpublished scheme based on fractional cover developed by Joerg Steinkamp

\strong{SimpleAdgvm2BiomeScheme} SS document here

\strong{SimpleAdgvm2BiomeScheme} SS document here

\strong{SimpleHeightAdgvm2BiomeScheme} SS document here

\strong{GrowthFormAdgvm2BiomeSchemes} SS document here

\strong{PhenologyAdgvm2BiomeScheme} SS document here
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}

Glenn Moncrief \email{glenn@saeon.ac.za}

Simon Scheiter \email{Simon.Scheiter@senckenberg.de}
}
\keyword{datasets}
