% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-classes.R
\docType{class}
\name{BiomeScheme-class}
\alias{BiomeScheme-class}
\title{A biome scheme}
\description{
This class stores the information about a biome scheme.  It describes what how the model output (in the form of a data.table) must be prepared, and then the rules which are used to classify the biomes.
}
\details{
This is ultimately not flexible enough for all conceivable biome schemes from all models and will need to be somehow generalised and expanded at some point.  This is most certainly a challenge for another day!
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular biome scheme.  Recommended to be alphanumeric because it is used to construct file names.}

\item{\code{name}}{A character string that can be more descriptive.}

\item{\code{rules}}{A function which is applied to every row of the data.table and describes the biome classification rules.}

\item{\code{combineShadeTolerance}}{If TRUE, call combineShadeTolerance before doing the biome classification}

\item{\code{totals.needed}}{List of vegetation totals needed to calculate biomes, for example c("Tree", "Grass")}

\item{\code{max.needed}}{List of maximum values needed to calculate biomes, for example c("PFT", "Tree")}

\item{\code{fraction.of.total}}{List of vegetation fraction of totals needed to calculate biomes}

\item{\code{fraction.of.tree}}{List of vegetation fraction of tree total needed to calculate biomes}

\item{\code{fraction.of.woody}}{List of vegetation fraction of woody total needed to calculate biomes}

\item{\code{needGDD5}}{If TRUE the biome rules require GDD5 for the classification}

\item{\code{strings}}{List of character strings containing the names of the biomes}

\item{\code{cols}}{List of character strings containing the names of R colous for the biomes (ordering matching the string slot)}

\item{\code{data.reference}}{Character string giving a reference where the data for this biome scheme comes from}

\item{\code{published.reference}}{Character string giving a reference where this model output classification scheme was published}
}}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

