% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/biome-schemes.R
\docType{class}
\name{BiomeScheme-class}
\alias{BiomeScheme-class}
\alias{Smith2014BiomeScheme}
\alias{Hickler2012BiomeScheme}
\alias{Forrest2015BiomeScheme}
\alias{MeditBiomeScheme}
\alias{DevMegaBiomeScheme}
\alias{FireMIPBiomeScheme}
\alias{FPCMegaBiomeScheme}
\alias{SimpleAdgvm2BiomeScheme}
\alias{SimpleHeightAdgvm2BiomeScheme}
\alias{GrowthFormAdgvm2BiomeScheme}
\alias{PhenologyAdgvm2BiomeScheme}
\title{A biome scheme}
\format{An object of class \code{BiomeScheme} of length 1.}
\usage{
Smith2014BiomeScheme

Hickler2012BiomeScheme

Forrest2015BiomeScheme

MeditBiomeScheme

DevMegaBiomeScheme

FireMIPBiomeScheme

FPCMegaBiomeScheme

SimpleAdgvm2BiomeScheme

SimpleHeightAdgvm2BiomeScheme

GrowthFormAdgvm2BiomeScheme

PhenologyAdgvm2BiomeScheme
}
\description{
This class stores the information about a biome scheme.  It describes what how the model output (in the form of a data.table) must be prepared, and then the rules which are used to classify the biomes.

\strong{Hickler2012BiomeScheme}  Classification of european output as per Hickler et al. 2012.

\strong{Forrest2015BiomeScheme}  Classification of global output into coarser mega biomes as per Forrest et al. 2015 (see appendix), 
 which is really just an aggregation of biomes following Harrison and Prentice 2006.  Can be compared to aggregated Haxeltine and Prentice (1996) map.

\strong{MeditBiomeScheme} A biome classification scheme used for Turkey.

\strong{DevMegaBiomeScheme}  Unpublished for potentially useful classification of global biomes which is slightly simpler than Smith et al. 2014.

\strong{FireMIPBiomeScheme}  Simple biomes based on a common PFT set for the FireMIP models

\strong{FPCMegaBiome} an unpublished scheme based on fractional cover developed by Joerg Steinkamp

\strong{SimpleAdgvm2BiomeScheme} SS document here

\strong{SimpleHeightAdgvm2BiomeScheme} SS document here

\strong{GrowthFormAdgvm2BiomeSchemes} SS document here

\strong{PhenologyAdgvm2BiomeScheme} SS document here
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular biome scheme.  Recommended to be alphanumeric because it is used to construct file names. (Inherited from Quantity via "contains")}

\item{\code{name}}{A character string that can be more descriptive of the biome scheme. (Inherited from Quantity via "contains")}

\item{\code{units}}{A list of character strings giving the names of categories (biomes). (Inherited from Quantity via "contains")}

\item{\code{colours}}{A function that returns the colour scale for this BiomeScheme. (Inherited from Quantity via "contains")}

\item{\code{format}}{Either a the string "Standard" to denote that this is a standard quantity to be compared across all model and data, the id of the Format object with which this Quantity is associated.}

\item{\code{rules}}{A function which is applied to every row of the data.table and describes the biome classification rules.}

\item{\code{layers.needed}}{List of vegetation layers needed to calculate biomes and the name of the new layer, to be interpreted by layerOp(), specifified as a list of three- or four-item list 
whose elements first element id the id of a DGVMTools::Quantity and whose other elements are passed as arguements to the layerOp() function.  
For example one element could be \code{Woody = list(quantity = "LAI_std", operator = "+", layers = c(".Tree", ".Shrubs"), new.layer = "Woody")}, 
which would make a layer called "Woody" which would be the sum of all LAI trees and shrubs.}

\item{\code{data.reference}}{Character string giving a reference where the data for this biome scheme comes from}

\item{\code{published.reference}}{Character string giving a reference where this model output classification scheme was published}
}}

\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}

Simon Scheiter \email{Simon.Scheiter@senckenberg.de}
}
\keyword{datasets}
