% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNewModelObj.R
\name{calcNewModelObj}
\alias{calcNewModelObj}
\title{Simple calculation of new ModelObjects}
\usage{
calcNewModelObj(x, y, op, x.col = NULL, y.col = NULL, quant = NULL,
  verbose = TRUE)
}
\arguments{
\item{x}{The first \code{\linkS4class{ModelObject}}}

\item{y}{The second \code{\linkS4class{ModelObject}}}

\item{op}{which arithmetic should be performed: addition ('+'), substraction ('-'), multiplication ('*') or division ('/').}

\item{x.col}{the column of the first ModelObect. If empty or NULL all columns are used.}

\item{y.col}{the column of the second ModelObect. If empty or NULL all columns are used.}

\item{quant}{new Quantity definition to use, if NULL it will be guessed.}

\item{verbose}{print some messages.}
}
\value{
hopefully a new ModelObject.
}
\description{
Basic calculation of new ModelObjects by operation '+', '-', '*' or '/'
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}
