% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{calcNME}
\alias{calcNME}
\title{Calculate Normalised Mean Error}
\usage{
calcNME(mod, obs)
}
\arguments{
\item{mod}{A numeric vector of modelled values (same size as obs)}

\item{obs}{A numeric vector of observed values}
}
\value{
A numeric
}
\description{
Calculates NME between two datasets (represented as two equally sized numeric vectors)
}
\details{
No check currently done on vector lengths
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
