% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{categoricalComparison}
\alias{categoricalComparison}
\title{Comparison between two datasets of categorical variables}
\usage{
categoricalComparison(vector1, vector2, name1, name2, verbose = TRUE)
}
\arguments{
\item{vector1}{The first of the datasets to be compared to each other, as a vector of factors (categorical data)}

\item{vector2}{The second of the datasets to be compared to each other, as a vector of factors (categorical data)}

\item{name1}{A character string giving the name of the first dataset (for making a useful id)}

\item{name2}{A character string giving the name of the second dataset (for making a useful id)}

\item{verbose}{A logical, if TRUE print out all the Kappa scores

Note that there are many other slots in a Statistics object which will not be filled in the resulting object because they are for continuous as opposed to categorical data}
}
\value{
A spatial comparison object
}
\description{
Calculates a Statistics object (which contains the Cohen's Kappa scores) given a stack containing two maps of categorical data (eg. biomes, land cover classes).
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
