% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{combineShadeTolerance}
\alias{combineShadeTolerance}
\title{Combine shade-intolerant PFTs with their shade-tolerant cousions}
\usage{
combineShadeTolerance(input)
}
\arguments{
\item{input}{The Field which is to have the shade tolerance classes combined.}
}
\value{
A Field with the data for the shade-intolerant PFTs set to zero but their values added to the shade-tolerant versions
}
\description{
The effects of this depend on the shade-tolerant cousin PFTs being defined in the PFT list.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
