% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineShadeTolerance.R
\name{combineShadeTolerance}
\alias{combineShadeTolerance}
\title{Combine shade-intolerant PFTs with their shade-tolerant cousions}
\usage{
combineShadeTolerance(input, verbose = FALSE)
}
\arguments{
\item{input}{The Field which is to have the shade tolerance classes combined.}

\item{verbose}{Logical, if TRUE print some progress updates}
}
\value{
Not necessary anything since the Field is modified, but it does return Field with the data for the shade-intolerant PFTs set to zero but their values added to the shade-tolerant versions
}
\description{
The effects of this depend on the shade-tolerant cousin PFTs being defined in the PFT list.  Note that this function modifies the original Field in place, 
so make a copy of the Field first if you want to keep the old one.  Also it means that it doesn't need to be re-asigned with the "<-" syntax, you can actually just call the function.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
