% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-benchmarking.R
\name{compareBiomes}
\alias{compareBiomes}
\title{Compare a run to a biome map}
\usage{
compareBiomes(runs, variable, period, scheme, biome.dataset, plot = TRUE,
  show.stats = TRUE, summary.plot.dir = NULL, tag = "",
  canvas.options = canvas.options, ...)
}
\arguments{
\item{runs}{List of VegRuns to be compared to the biome data}

\item{variable}{The VegQuant (or name) of the variable from which the biomes are to be derived. This is not flexible enough, 
and should be folded into the scheme.}

\item{period}{The temporal extent over which the model output should be averaged for calculating the biomes}

\item{scheme}{The biome scheme we are comparing. Also not flexible enough, here should provide a raster (or SpatialDataset?) 
containing the data}

\item{biome.dataset}{The biomes to which you wish to comapre, as a SpatialDataset.}

\item{plot}{Logical, if true make a biome plot.}

\item{show.stats}{A logical, if TRUE, put the Kappa values on the plots.}

\item{summary.plot.dir}{A directory (full path as a character string) so save the plots which compare many runs}

\item{tag}{A character string (no spaces) used in labels and titles to differentiate these plots from similar ones.
For example "Corrected" or "EuropeOnly"}

\item{canvas.options}{A list of options (to be given to the \code{Cairo}) function to define the canvas. See the \code{Cairo} documentation}

\item{...}{Additional parameters supplied to the plotVegMaps() plotting function.}
}
\value{
A Biome comparison object
}
\description{
Takes a run, a time period and a variable from which to calculate a modelled biome map, and biome scheme, 
and compares the model to the data
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

