% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareLayers.R
\name{compareLayers}
\alias{compareLayers}
\title{Compare single layers to each other}
\usage{
compareLayers(field1, field2, layers1, layers2 = layers1,
  keepall1 = FALSE, keepall2 = FALSE, override.quantity = FALSE,
  verbose = FALSE, match.NAs = FALSE, show.stats = TRUE,
  dec.places = NULL)
}
\arguments{
\item{field1}{A Field or DataObject from which to get the first layer for comparison}

\item{field2}{A Field or DataObject from which to get the second layer for comparison}

\item{layers1}{The name of the first layer (character string)}

\item{layers2}{The name of the first layer (character string).  If not defined taken to the be the same as layers1}

\item{keepall1}{Boolean, keep all data points in layers1 even if there is not corresponding data in layers2}

\item{keepall2}{Boolean, keep all data points in layers2 even if there is not corresponding data in layers2}

\item{override.quantity}{Boolean, if TRUE ignore situation where field1 and field2 have non-identical Quantities and use the Quantity from field1 for the returned Comparison}

\item{verbose}{Boolean, if TRUE print some informative output

The returned Comparison object can be plotted using \code{plotSpatail} (to give the absolute difference, original values side-by-side and percentage difference, also the NME spatial - to be implemented)
The stats slot (which contains a Statistics object) holds information such as RSME, NME, Nash-Sutcliffe Model Efficiency, etc. between the datasets in the case of a continous data.
In the case of comparing categorical data (ie layers of factors) it contains Cohen's Kappa.}

\item{match.NAs}{Boolean, if TRUE copy NAs from one layer to the other.  This is mostly to make sure that when you later plot the data in the final Comparison side-by-side,
that the both have 'no data' (typically grey areas) plotted on both maps.}

\item{show.stats}{Boolean, if TRUE print the summary statistics}

\item{dec.places}{Numeric, passed to copyLayers. Defines to how many decimal places to round the coordinates in order to get a match.  Default is no rounding (value is NULL) and if dine for most regularing spaced grids.}
}
\value{
A Comparison object
}
\description{
Compare two layers (each from a Field or DataObject) to calculated various statistic metric and also the error (at every spatial/temporal locality) which is returned as a Comparison object.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
