% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-benchmarking.R
\name{compareManyRunsToBiomes}
\alias{compareManyRunsToBiomes}
\title{Compare many runs to a biome map}
\usage{
compareManyRunsToBiomes(runs, biome.dataset, plot.data = TRUE, ...)
}
\arguments{
\item{runs}{List of VegRun objects, each run should have already had \code{compareBiomes()} for the relevant biome dataset 
run on it.}

\item{biome.dataset}{A SpatialDataset object holding the biomes to which we are comparing.}

\item{plot.data}{A logical, if TRUE put the data on the plot as well as the model runs}

\item{...}{Further arguments passed to the plotting function plotVegMaps(). Note in particular that "tag" is a useful option to differentiate the resulting plot from other plots of the same type in the same directory}
}
\description{
Takes a list of VegRuns, looks up the relevant BiomeComparison object (which should have been pre-calculated)
and plots all the biome maps on one plot, including the data
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

