% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Comparison-class}
\alias{Comparison-class}
\title{Contains a comparison between two layers}
\description{
This object is produced as the result of a call to the functions \code{compareLayers()} and shouldn't be directly created by a user.
 
It contains a data slot with each of the compared layers and the difference between them (layer 1 - layer 2), a Comparisons object with the calculated values of 
various statistical metric, meta-data about the source of the two compared layers and spatio-temporal meta-data describing where the comparison is valid.
}
\details{
It can be plotted by functions like   
   
Generally these are not created directly by the user, but rather by functions like \code{getField}.
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular vegetation object.  Recommended to be alphanumeric because it is used to construct file names.}

\item{\code{name}}{A character string describing this comaprison layer, is automatically generated}

\item{\code{data}}{A data.table object.  This is used because is it very much faster for calculations that data.frame or raster layers.}

\item{\code{quant1}}{A Quantity object to define what quantity the data from first field represents}

\item{\code{quant2}}{A Quantity object to define what quantity the data from second field represents}

\item{\code{layers1}}{A character string (or vector thereof) of the layer(s) from the first field that were compared}

\item{\code{layers2}}{A character string (or vector thereof) of the layer(s) from the second field that were compared}

\item{\code{stats}}{An SpatialComaprison object giving ther statistical comparison metric between these two layers}

\item{\code{source1}}{A Source object describing the source of the first layer in the comparison}

\item{\code{source2}}{A Source object describing the source of the second layer in the comparison}

\item{\code{sta.info1}}{The STAInfo object for the first field}

\item{\code{sta.info2}}{The STAInfo object for the second field}
}}

\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
