% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{continuousComparison}
\alias{continuousComparison}
\title{Compare continuous data}
\usage{
continuousComparison(vector1, vector2, name1, name2, verbose = TRUE)
}
\arguments{
\item{vector1}{The first of the datasets to be compared to each other, as a vector of numerics (categorical data)}

\item{vector2}{The second of the datasets to be compared to each other, as a vector of numerics(categorical data)}

\item{name1}{A character string giving the name of the first dataset (for making a useful id)}

\item{name2}{A character string giving the name of the second dataset (for making a useful id)}

\item{verbose}{A logical, if TRUE print out all the Kappa scores

Note that there are many other slots in a Statistics object which will not be filled in the resulting object because they are not for continuous data.}
}
\value{
A spatial comparison object
}
\description{
Compares two datasets of continuous data. Specifically calculates and returns a Statistics object (which contains many metrics) given numeric two vectors of equal size
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
