% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-import.R
\name{daylength}
\alias{daylength}
\title{daylength calculation}
\usage{
daylength(lat, doy, leap = FALSE)
}
\arguments{
\item{lat}{latitude (vector)}

\item{doy}{day of the year (vector)}

\item{leap}{use February 29th or not}
}
\value{
daylength in hours as vector or matrix, depending on input shape
}
\description{
daylength calculation depending on latitude and day of year.
taken from function daylengthinsoleet (driver.cpp) LPJ-GUESS v2.1
}
\references{
This code is based on the original LPJ-GUESS C++ source code (\href{http://iis4.nateko.lu.se/lpj-guess/}{LPJ-GUESS homepage}).
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}
