% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineModelRun.R
\name{defineModelRun}
\alias{defineModelRun}
\title{Define a Source object that represents a model run, setting up all the required metadata (only).}
\usage{
defineModelRun(id, name, dir, model, pft.set, lonlat.offset = c(0, 0),
  year.offset = 0, london.centre = TRUE, driving.data = "Not specified",
  land.use.included = FALSE, contact = "Not specified",
  institute = "Not specified")
}
\arguments{
\item{id}{A unique character string to identify this particular model un.  Recommended to be alphanumeric because it is used to construct file names. (Mandatory)}

\item{name}{A character string describing this run, ie. "LPJ-GUESS v3.1"}

\item{model}{A character string to identify what model produced this run.  Can currently be "LPJ-GUESS", "LPJ-GUESS-SPITFIRE" or "aDGVM". (Mandatory)}

\item{pft.set}{A list of PFT objects which includes all the PFTs used is this model run (Mandatory)}

\item{lonlat.offset}{A numeric of length 1 or 2 to define the offsets to Lon and Lat to centre the modelled localities.}

\item{year.offset}{A numeric of length 1 to match be added to the simulation years to convert them to calendar years}

\item{london.centre}{If TRUE, ensure that the longitudes are (-180,180) instead of (0,360)}

\item{driving.data}{A character string identifying the climate or other data used to produce this model run}

\item{land.use.included}{If TRUE it can be assumed that land use has been simulated for this run and so no correction for land use need be applied before benchmarking.}

\item{contact}{Name and email address of responsible person (default to OS username).}

\item{institute}{Name of the institute (default "none").

The parameters are the slots of a SourceInfo object. Note that that \code{id}, \code{model}, \code{pft.set} and \code{dir} are compulsory, the rest will be filled with dummy/default values if left blank.
Take care with \code{lon.lat.offset} and \code{year.offset} which are initialised to 0 which is unsuitable for some LPJ-GUESS configurations.}

\item{run}{The location of this run on the file system (Mandatory)}

\item{fill.col}{A string to define an R colour used when plotting this run as a histogram or scatter plot or so}

\item{line.col}{A string to define an R colour used when plotting this runs as a line graph}

\item{line.width}{A numeric to define the width of a line representing this model run}

\item{line.type}{A numeric to define the line style representing this model run}
}
\value{
A Source object, with metadata defined by empty data slots
}
\description{
This function is preferred to a \code{new("SourceInfo",...)} initialisation followed by  \code{new("Source",...)} initialisation because it does both intialisations in one step and also performs some extra checks and initialisations of defaults.
}
\details{
Note that initialliy, no actual data from the run is stored, only metadata. The data, stored as Fields, can be added to a Source object, are added built with later with other commands and added to \code{Source} command using \code{addToSource}
}
\seealso{
Source, SourceInfo
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
