% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{divideLayers}
\alias{divideLayers}
\title{Calculate fractions from the layers of a Field with respect to other layers}
\usage{
divideLayers(input, layers, denominators = list("Total"),
  aggregate.method = "sum")
}
\arguments{
\item{input}{The Field for which to calculate the new fractional layers}

\item{layers}{The layers to be divided ie. the numerators (will be calculated by \code{getVegTotals} if the don't exist)}

\item{denominators}{The denominator layers (will be calculated by \code{getVegTotals} if the don't exist) (defaults to just "Total")}

\item{aggregate.method}{If the denominators need to be made, which method should be used to make them, "sum" or "mean", defaults to "sum".
parame}
}
\value{
A Field (or data.table) with the new layers added
}
\description{
This is very useful and important function.  It is fully flexible  -both the numerator and denominator can be specified (although the denominator defaults to "Total").
Note that if a layer doesn't exist it will be created if possible.
}
\details{
Division is safe with respect to a zero denominator, the results of dividing by zero is, in this case, zero.

For convenience, both \code{layers} and \code{denominators} will be expanded using \code{expandLayers}.
This allows all lifeforms fractions in a Field to be calculated using a simple call such as

 \code{veg.obj <- divideLayers(veg.obj, c("lifeforms")}
 
See documention of \code{expandLayers} for details.
}
\seealso{
expandslayers getVegTotals
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
