% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-runtools.R
\name{doSpatialAverage}
\alias{doSpatialAverage}
\title{Spatially average a data.table}
\usage{
doSpatialAverage(input.dt, verbose = FALSE, area.weighted = TRUE)
}
\arguments{
\item{input.dt}{data.table}

\item{verbose}{If TRUE give some progress update about the averaging.}

\item{area.weighted}{If TRUE area-weight the gridcells}
}
\value{
A data.table
}
\description{
Spatially average all gridcells (identified by "Lat" and "Lon" columns) of a data.table object. 
Can use area-weighting to take into account the difference different areas of gridcells (even though they are constant in Lon,Lat)
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

