% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-processing.R
\name{expandTargets}
\alias{expandTargets}
\title{Expand plotting/processing targets from convenient short hand codes.}
\usage{
expandTargets(targets, PFTs, include.woody = TRUE)
}
\arguments{
\item{targets}{Character vector of targets to expand}

\item{PFTs}{List of all PFTs present (can be determined beforehand with a call to \code{getPFTs})}

\item{include.woody}{If TRUE and "lifeform" is included in the target list "Woody" is also returned}
}
\value{
A list of targets
}
\description{
Expands characters strings with particular meaning (eg. "lifefoms", "pfts", "seasons", etc...).  
This allows the user to simply specify "lifeforms" to a plotting command instead of c("Tree","Grass","Shrub") when plotting or processing.
}
\details{
Supported short-hand strings are:
\itemize{
  \item{"seasons"}{ expands to c("DJF","MAM","JJA","SON")}
  \item{"PFTs"}{ expands to a vector of each of the ids of PFTs supplied in the \code{PFTs} argument.}
  \item{"lifeforms"}{ expands all the different PFT lifeforms present as determined by the lifeform slots of the \code{PFTs} argument (and optionally "Woody").  For example c("Tree", "Grass", "Woody")}
  \item{"zones"}{ expands all the different PFT climactic zones present as determined by the zone slots of the \code{PFTs} argument.  For example c("Temperature", "Boreal", "Tropical")}
  \item{"leafforms"}{ expands all the different PFT leaf forms present as determined by the leafforms slot of the \code{PFTs} argument.  For example c("Needleleved", "Broadleaved")}
  \item{"phenologies"}{ expands all the different PFT leaf forms present as determined by the phenology slot of the \code{PFTs} argument.  For example c("Evergreen", "Summergreen", "Raingreen", "GrassPhenology)}
  \item{"all"}{ expands to c("pfts", "lifeforms", "leafforms", "zones", "phenologies"), and then these are further expanded.}
}
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

