% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-methods.R
\docType{methods}
\name{export-methods}
\alias{export-methods}
\alias{as.data.frame.Field}
\alias{as.data.frame.Comparison}
\alias{as.data.table.Field}
\alias{as.data.table.Comparison}
\alias{as.Raster}
\alias{as.Raster,Field-method}
\alias{as.Raster,Comparison-method}
\alias{as.array}
\alias{as.array,Field-method}
\title{Coerce from Fields and DataObjects}
\usage{
\method{as.data.frame}{Field}(x, row.names, optional, ...)

\method{as.data.frame}{Comparison}(x, row.names, optional, ...)

\method{as.data.table}{Field}(x, keep.rownames, ...)

\method{as.data.table}{Comparison}(x, keep.rownames, ...)

as.Raster(x)

\S4method{as.Raster}{Field}(x)

\S4method{as.Raster}{Comparison}(x)

as.array(x, ...)

\S4method{as.array}{Field}(x, ...)
}
\arguments{
\item{x}{A Field or a DataObject}

\item{row.names}{NULL or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{logical. If TRUE, setting row names and converting column names (to syntactic names: see make.names) is optional. 
Note that all of R's base package as.data.frame() methods use optional only for column names treatment, 
basically with the meaning of data.frame(*, check.names = !optional).}

\item{...}{Just as ... in data.frame. Usual recycling rules are applied to vectors of different lengths to create a list of equal length vectors.}

\item{keep.rownames}{If ... is a matrix or data.frame, TRUE will retain the rownames of that object in a column named rn.}
}
\description{
Functions to coerce Fields and DataObjects into other objects (data.frame, data.table, RasterBrick, SpatialPixelsDataFrame)
}
\details{
Note that for coercing to a Raster* object (RasterLayer or RasterBrick) the function is called "as.Raster" (capital "R") to avoid conflict with  
another function in the raster package called "as.raster"
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}, Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}
