% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent-methods.R
\docType{methods}
\name{extent-methods}
\alias{extent-methods}
\alias{extent}
\alias{extent,data.table-method}
\alias{extent,ModelObject-method}
\alias{extent,DataObject-method}
\alias{extent,ComparisonLayer-method}
\alias{extent,SpatialExtent-method}
\title{Extent methods}
\usage{
\S4method{extent}{data.table}(x)

\S4method{extent}{ModelObject}(x)

\S4method{extent}{DataObject}(x)

\S4method{extent}{ComparisonLayer}(x)

\S4method{extent}{SpatialExtent}(x)
}
\arguments{
\item{x}{A ModelObject, DataObject or a data.table.}

\item{...}{Other arguments, not currently used}
}
\value{
A raster::extent
}
\description{
Methods for returning a raster::extent object from a ModelObject, DataObject or a data.table.  The data.table, or the data.table in the "data" slot of the ModelObject and DataObject, 
must have columns named "Lon" and "Lat" otherwise these methods fail.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
