% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Field-class}
\alias{Field-class}
\title{Field-class contains data}
\description{
Field is a key class of the package as it actually holds the data (most other classes are for metadata).  A \code{\linkS4class{Field}} stores the data and metadata for one quantity
that comes from a dataset or vegetation model run (including information about the run itself). For example LAI (Leaf Area Index), or evapotranspiration. 
The data can be aggregated across space, and across and within years, and manipulated and plotted by many functions in this package.
}
\details{
Generally these are not created directly by the user, but rather by functions like \code{\link{getField}}.
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular vegetation object.  Recommended to be alphanumeric because it is used to construct file names.}

\item{\code{data}}{A data.table object.  This is used because is it very much faster for calculations that data.frame or raster layers.}

\item{\code{quant}}{A Quantity object to define what output this Field contains}

\item{\code{first.year}}{The first year of data}

\item{\code{last.year}}{The last year of data}

\item{\code{year.aggregate.method}}{A character string describing the method by which the years have been aggregated.  Will be the zero character ("character(0)") if data it not 
yearly aggregated}

\item{\code{spatial.extent}}{An object which can be used to crop or select gridcells.  Can be anything from which a raster::extent can be derived (in which case raster::crop is 
used) or a list of gridcells used by \code{\link{selectGridcells}} (see that documentation for how to format the gridcell list).}

\item{\code{spatial.extent.id}}{A character id to handily record this spatial domain if some spatial subselection has been called, for example "Europe" or "Duke_Forest" or whatever}

\item{\code{spatial.aggregate.method}}{Set to TRUE is this.Field has been spatially averaged}

\item{\code{subannual.aggregate.method}}{Method by which this Field has been subannually aggregated}

\item{\code{subannual.original}}{Original subannual resolution of this field}

\item{\code{source}}{A Source object which contains the metadata about the run which this Field belongs too.}
}}

\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
