% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-methods.R
\name{FieldToArray}
\alias{FieldToArray}
\title{Convert a Field to Arrays}
\usage{
FieldToArray(
  x,
  cname = FALSE,
  invertlat = FALSE,
  fill.gaps = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{x}{the data.table of a \code{\linkS4class{Field}}}

\item{cname}{the column name to convert, if not set a list is returned}

\item{invertlat}{start in the north}

\item{fill.gaps}{logical, if TRUE fill longitude and latitude gaps (must be regularly spaced grid)}

\item{verbose}{print some information}
}
\value{
a array or a list or arrays
}
\description{
Convert a Field (or the data.table from a Field) into a list of multi-deminsional arrays (indexed Lon, Lat and Time),
with one array in the list per Layer in the input Field.
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}, Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
