% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/Format-GUESS.R, R/Format-NetCDF.R,
%   R/Format-aDGVM.R, R/Format-aDGVM2.R
\docType{class}
\name{Format-class}
\alias{Format-class}
\alias{GUESS}
\alias{NetCDF}
\alias{aDGVM}
\alias{aDGVM2}
\title{Format class}
\format{
A \code{\linkS4class{Format}} object is an S4 class.

A \code{\linkS4class{Format}} object is an S4 class.

A \code{\linkS4class{Format}} object is an S4 class.

A \code{\linkS4class{Format}} object is an S4 class.
}
\usage{
GUESS

NetCDF

aDGVM

aDGVM2
}
\description{
This class encapsulates all the specific details of a supported model or dataset.  This comprises functions to read the data off the disk, 
and metadata describing certain pre-defined \code{\linkS4class{Quantity}} and \code{\linkS4class{Layer}} objects. From a user-perspective, these are used as the \code{\linkS4class{Format}} argument to
\code{\link{defineSource}} function.

\code{GUESS} - a Format for reading standard LPJ-GUESS(-SPITFIRE) model output

\code{NetCDF} - a Format object defined here for reading NetCDF files.
It is essentially CF-compliant netCDF with a couple of extra attributes defined.

\code{aDGVM} - a Format for reading standard aDGVM model output

\code{aDGVM2} - a Format for reading aDGVM2 model output
}
\details{
Format objects which are included in the package are listed below:

For DGVMTools to support a particular model or dataset, this is the object that needs to be defined.  But normally a user won't need to deal with this 
class since it defines model/dataset specific metadata and functions which should be defined once and then 'just work' (haha) in the future. 
If someone wants their model to be supported by DGVMTools then this is the object that needs to be defined correctly.

Note that the 'predefined.layers' and  'quantities' arguments are just default values, it is easy to add new ones.  Equally they don't all need
to be available for a particular run or dataset.  You can define your own for your own data format if you, for example, want to include a new model type.
}
\section{Slots}{

\describe{
\item{\code{id}}{Simple character string to gave an uniquely identify this format}

\item{\code{predefined.layers}}{'Standard' Layer type that this format uses, as a list of \code{\linkS4class{Layer}} objects.
This is most likely applicable to formats describing DGVM output, but also for some data sets.  
This is just a default Layer set available for convenience, can be easily over-ridden when defining a Source object (see \code{\link{defineSource}}).}

\item{\code{quantities}}{'Standard' quantities (as a list of \code{\linkS4class{Quantity}} objects) which might be availably from the model output or dataset.}

\item{\code{availableQuantities}}{A function to determine which quantities \emph{are actually available} from the model run or dataset.}

\item{\code{getField}}{A function to retrieve actually data from the model output or dataset.  This is likely to be a fairly complex function, and really depends on the specifics 
and idiosynchrasies of the model output format or dataset.}
}}

\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{datasets}
