% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome-schemes.R
\name{FPCMegaBiomeRules}
\alias{FPCMegaBiomeRules}
\title{Rules for biome scheme based on FPC}
\usage{
FPCMegaBiomeRules(x)
}
\arguments{
\item{fpc}{Vector of FPC values}
}
\description{
lalala
}
\keyword{internal}
