% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-import.R
\name{getAnnualClimate}
\alias{getAnnualClimate}
\title{Get Annual values of climate drivers}
\usage{
getAnnualClimate(run = NA, file = NA, operation = "mean",
  data.name = NA, lon.name = NA, lat.name = NA, time.name = NA,
  period = NA, forceReCalculation = FALSE, write = TRUE, verbose = TRUE,
  ...)
}
\arguments{
\item{run}{a VegRun object, if a VegSpatial object should be returned, otherwise NA}

\item{file}{path to the netCDF file}

\item{operation}{Any of sum, mean, gdd5}

\item{data.name}{Name of the data field in the netcdf file}

\item{lon.name}{Name of the longitude vector in the netcdf file}

\item{lat.name}{Name of the latitude vector in the netcdf file}

\item{time.name}{Name of the time vector in the netcdf file}

\item{period}{Either a TemporalExtent object, a vector of years (min/max are choosen) or 'NA' (whole netCDF timespan is used)}

\item{forceReCalculation}{Recalculate the desired values (default: FALSE)}

\item{write}{write the calculated values for fture speedup (default: TRUE)}

\item{verbose}{print some messages}
}
\value{
data.table if no VegRun was given, otherwise a VegSpatial
}
\description{
Get Annual values of climate drivers from netCDF files, files must be CF-compliant
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

