% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-aDGVM.R
\name{getDailyField_aDGVM}
\alias{getDailyField_aDGVM}
\title{Get a 'daily' aDGVM Field}
\usage{
getDailyField_aDGVM(run, quant, target.sta, file.name = file.name,
  adgvm.file.type, verbose = FALSE, data.table.only = FALSE,
  adgvm.fire = 1, adgvm.climate = 0, adgvm.header, ...)
}
\arguments{
\item{run}{A \code{Source} containing the meta-data about the aDGVM run}

\item{quant}{A Quant to define what output file from the aDGVM run to open, 
can also be a simple string defining the aDGVM output file if the \code{return.data.table} argument is TRUE}

\item{file.name}{An optional character string (or a list of character strings) holding the name of the file(s)
This can be left blank, in which case the file name is automatically generated.}

\item{verbose}{A logical, set to true to give progress/debug information}

\item{data.table.only}{A logical, if TRUE return a data.table and not a Field}

\item{first.year}{The first year (as a numeric) of the data to be return}

\item{last.year}{The last year (as a numeric) of the data to be return}
}
\value{
a data.table (with the correct tear offset and lon-lat offsets applied)
}
\description{
\code{getYearlyField_aDGVM} returns a data.table object given a string defining a vegetation quantity 
from the run (eg. "lai", to read the file "lai.out") and  \code{Source} object which defines where the run is on disk and the offsets to apply
}
\details{
Note that the files can be gzipped on UNIX systems, but this might fail on windows systems.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}

Glenn Moncrief \email{glenn@saeon.ac.za}
}
\keyword{internal}
