% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-aDGVM.R
\name{getField_aDGVM}
\alias{getField_aDGVM}
\title{Get a Field for aDGVM}
\usage{
getField_aDGVM(source, quant, target.STAInfo, file.name, verbose,
  adgvm.file.type = "Yearly", ...)
}
\arguments{
\item{source}{A \code{Source} containing the meta-data about the aDGVM run}

\item{quant}{A string the define what output file from the aDGVM run to open, for example "anpp" opens and read the "anpp.out" file}

\item{target.STAInfo}{The spatial-temporal target domain}

\item{file.name}{An optional character string (or a list of character strings) holding the name of the file(s)
This can be left blank, in which case the file name is automatically generated.}

\item{verbose}{A logical, set to true to give progress/debug information}
}
\value{
A list containing firstly the data.tabel containing the data, and secondly the STA.info
}
\description{
An internal function that reads data from an aDGVM run.  It actually call one of three other functions depending on the type of quantity specified.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}

Glenn Moncrief \email{glenn@saeon.ac.za}
}
\keyword{internal}
