% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-aDGVM.R
\name{getField_aDGVM}
\alias{getField_aDGVM}
\title{Get a Field for aDGVM}
\usage{
getField_aDGVM(source, quant, target.STAInfo, verbose, adgvm.scheme,
  adgvm.daily)
}
\arguments{
\item{quant}{A string the define what quantity from the aDGVM2 run to extract}

\item{verbose}{A logical, set to true to give progress/debug information}

\item{adgvm.scheme}{A number that defines if pop-files (=1) or trait-files (=2) are used.}

\item{adgvm.daily}{A logical, set to true to read daily data (only for \code{adgvm.scheme=1} and if daily data are provided in pop file)}

\item{run}{A \code{source} containing the meta-data about the aDGVM2 run}

\item{first.year}{The first year (as a numeric) of the data to be return}

\item{last.year}{The last year (as a numeric) of the data to be return}
}
\value{
A list containing firstly the data.table containing the data, and secondly the STA.info
}
\description{
An internal function that reads data from an aDGVM2 run. It actually calls one of two other functions depending on the type of quantity specified.
}
\seealso{
\code{\link{getQuantity_aDGVM_Scheme1}, \link{getQuantity_aDGVM_Scheme2}}
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
