% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-NetCDF.R
\name{getField_NetCDF}
\alias{getField_NetCDF}
\title{Get a Field for NetCDF}
\usage{
getField_NetCDF(
  source,
  quant,
  target.STAInfo,
  file.name,
  verbose = FALSE,
  nc.verbose = FALSE
)
}
\arguments{
\item{source}{A \code{Source} containing the meta-data about the NetCDF source}

\item{quant}{A Quantity object to specify what quantity should be opened.}

\item{file.name}{Character string holding the name of the file.  This can be left blank, in which case the file name is automatically generated}

\item{verbose}{A logical, set to true to give progress/debug information}

\item{nc.verbose}{A logical, set to true to give progress/debug information from the ncdf4 package functions.  This can be a lot, 
so it is handy to control that separately.}

\item{target.sta.info}{An STAInfo object defining the spatial-temporal-annual extent over which we want the data}
}
\value{
A list containing firstly the data.table containing the data, and secondly the STAInfo for the data that we have
}
\description{
An internal function that reads data from an NetCDF .nc file.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
