% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{getModelObject}
\alias{getModelObject}
\title{Get a \code{ModelObject}, optionally for spatial/temporal averaging/cropping}
\usage{
getModelObject(run, var, temporal.extent = NULL, temporally.average = FALSE,
  spatial.extent = NULL, spatially.average = FALSE, area.weighted = TRUE,
  write = FALSE, read.full = TRUE, verbose = TRUE,
  store.internally = FALSE, store.full = FALSE, adgvm.scheme = 1)
}
\arguments{
\item{run}{The \code{ModelRun} object for which the spatially-averaged \code{ModelObject} should be built (eg. "lai")}

\item{var}{The quantity (either a \code{Quantity} or a string containing its \code{id})}

\item{temporal.extent}{The temporal extent (as a \code{TemporalExtent} object over which the data is to be averaged)}

\item{temporally.average}{Whether or not to temporally average (logical)}

\item{spatial.extent}{The spatial extent (as a \code{SpatialExtent} object over which the data is to be averaged)}

\item{spatially.average}{Whether or not to spatially average (logical)}

\item{area.weighted}{If TRUE, perform weighting by gridcell area when doing spatial averaging}

\item{write}{If TRUE, write the data of the \code{ModelObject} to disk as text file.}

\item{read.full}{If TRUE ignore any pre-averaged file on disk, if FALSE use one if it is there (can save a lot of time if averaged file is already saved on disk)}

\item{verbose}{If TRUE give a lot of information for debugging/checking.}

\item{store.internally}{If TRUE store the resulting \code{ModelObject} in the \code{ModelRun} for using later}

\item{store.full}{If TRUE save the full temporal and spatial output in memory (if it is read) to save time if making more \code{ModelObjects} from the variable later.  However, use with caution, saving too many full variables can easily fill up your system's RAM memory!}

\item{adgvm.scheme}{In the case of analysing an aDGVM run, select the PFT classification scheme for when post-hoc assigning the individuals into PFTS.}
}
\value{
A spatially averaged\code{ModelObject} but with no temporal averaging or cropping (ie. include complete original simulation duration). In other words, a time-series!
}
\description{
Given a \code{ModelRun} object a \code{Quantity} object, return an appropriate spatially-averaged \code{ModelObject} oject for that run and quantity. Arguments can also be provided for averaging over different spatial or temporal extents (very useful) or optionall just cropping to those extents
}
\details{
Note that because there are two types of averaging available, the resulting \code{ModelRun} obhect can either be full spatial-temporal dataset, a spatial-only dataset (map), a temporal only datasey a time-series) or an average across both space and time, i.e. a single number.
Also not that the data is stored internal as a data.table object, but this is mostly not important to the user.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\seealso{
\code{getModelObject}, \code{getVegspatial}
}

