% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODISTreeCover.R
\name{getMODISTreeCover}
\alias{getMODISTreeCover}
\title{Read MODIS TreeCover map}
\usage{
getMODISTreeCover(location = "/data/forrest/TreeCover/MOD44B/Collection5",
  resolution = "original")
}
\arguments{
\item{location}{A character string specifying the the location of the original data files}

\item{resolution}{A character string specifying the resolution, currently supported
are "1km-Turkey"

Should be moved to DGVMDatasets, not distributed with DGVMTools}
}
\value{
A \code{SpatialDataset} object.
}
\description{
Blah
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

