% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-adgvm.R
\name{getQuantity_aDGVM_Scheme1}
\alias{getQuantity_aDGVM_Scheme1}
\title{Simple classification of yearly output for aDGVM}
\usage{
getQuantity_aDGVM_Scheme1(run, period, variable)
}
\arguments{
\item{run}{A \code{ModelRun}}

\item{period}{A \code{TemporalExtent} object to define the period over which to average.}

\item{variable}{A character string specifying which variable to get, can be "lai" or "agb"}
}
\description{
Original from Simon Scheiter, this version modified by M. Forrest to fit into the DGVM framework. 
Simple scheme, classifies only as tree or grass based on only the pop file, doesn't open the trait file.
}
\author{
Simon Scheiter \email{simon.scheiter@senckenberg.de}, Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\seealso{
\code{getQuantity_aDGVM_Scheme2}
}
\keyword{internal}

