% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-aDGVM.R
\name{getQuantity_aDGVM_Scheme1}
\alias{getQuantity_aDGVM_Scheme1}
\title{Read aDGVM2 quantities from pop file}
\usage{
getQuantity_aDGVM_Scheme1(run, variable, first.year, last.year,
  adgvm.daily)
}
\arguments{
\item{run}{A \code{source} containing the meta-data about the aDGVM2 run}

\item{variable}{A character string specifying which variable/quantity to get, can be "agb“, "aGPP_std“, "basalarea“, "bgb“, "canopyheight_std“, "LAI_std“, meanheight“, "nind“, "pind“, "vegC_std“, "vegcover_std"}

\item{first.year}{First year of data to read}

\item{last.year}{Last year of data to read}
}
\description{
An internal function to read quantities from aDGVM2 pop files. Quantities are provided for trees, C4 grasses and C3 grasses. Trait files are not opened
}
\seealso{
\code{\link{getQuantity_aDGVM_Scheme2}}
}
\author{
Simon Scheiter \email{simon.scheiter@senckenberg.de}, Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
