% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-adgvm.R
\name{getQuantity_aDGVM_Scheme2}
\alias{getQuantity_aDGVM_Scheme2}
\title{Slightty more complex classification of yearly output for aDGVM}
\usage{
getQuantity_aDGVM_Scheme2(run, period, variable)
}
\arguments{
\item{run}{A \code{ModelRun}}

\item{period}{A \code{TemporalExtent} object to define the period over which to average.}

\item{variable}{A character string specifying which variable to get, only "agb" currently supported}
}
\description{
Original from Simon Scheiter, this version modified by M. Forrest to fit into the DGVM framework. 
Classifies as evergreen tree, deciduous tree or grass based on the pop file and the trait file.
}
\seealso{
\code{getQuantity_aDGVM_Scheme1}
}
\author{
Simon Scheiter \email{simon.scheiter@senckenberg.de}, Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
