% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-aDGVM.R
\name{getQuantity_aDGVM_Scheme2}
\alias{getQuantity_aDGVM_Scheme2}
\title{An internal function to read quantities from aDGVM2 trait files. Quantities are currently provided for raingreen trees, evergreen trees, raingreen shrubs, evergreen shrubs, C4 grasses and C3 grasses. Pop files are not opened.}
\usage{
getQuantity_aDGVM_Scheme2(run, variable, first.year, last.year)
}
\arguments{
\item{run}{A \code{source}}

\item{variable}{A character string specifying which variable/quantity to get, only "agb" currently supported}

\item{first.year}{First year of data to read}

\item{last.year}{Last year of data to read}
}
\description{
An internal function to read quantities from aDGVM2 trait files. Quantities are currently provided for raingreen trees, evergreen trees, raingreen shrubs, evergreen shrubs, C4 grasses and C3 grasses. Pop files are not opened.
}
\seealso{
\code{getQuantity_aDGVM_Scheme1}
}
\author{
Simon Scheiter \email{simon.scheiter@senckenberg.de}, Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
