% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-GUESS.R
\name{getStandardQuantity_LPJ}
\alias{getStandardQuantity_LPJ}
\title{Returns the data from one LPJ-GUESS output variable as a \code{data.table}.}
\usage{
getStandardQuantity_LPJ(run, quant, first.year, last.year,
  verbose = FALSE)
}
\arguments{
\item{run}{A \code{Source} containing the meta-data about the LPJ-GUESS run from which the data is to be read.  Most importantly it must contain the run.dara nd the offsets.}

\item{quant}{A string the define what output file from the LPJ-GUESS run to open, for example "anpp" opens and read the "anpp.out" file}

\item{first.year}{The first year (as a numeric) of the data to be return}

\item{last.year}{The last year (as a numeric) of the data to be return}

\item{verbose}{A logical, set to true to give progress/debug information}
}
\value{
a data.table (with the correct tear offset and lon-lat offsets applied)
}
\description{
This funtion can retrieve a 'Standard' vegetation quantity (returned as a data.table) with standard definition and units
to compare to other models and to data.  This must be implemented for each and every Standard quantity 
for each and every model to to ensure completeness.
}
\details{
output variable.  Normally it will read the file from disk, but if that has already been done, and the \code{data.table} has been saved to the 
\code{Source} object, it will return that to save time.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
