% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{getVegSpatial}
\alias{getVegSpatial}
\title{Get a temporally-averaged \code{ModelObject}}
\usage{
getVegSpatial(run, var, period, ...)
}
\arguments{
\item{run}{The \code{ModelRun} object for which the temporally-averaged \code{ModelObject} should be built (eg. "lai")}

\item{var}{The quantity (either a \code{Quantity} or a string containing its \code{id})}

\item{period}{The time period (as a \code{TemporalExtent} over which the data is to be averaged)}

\item{...}{additional arguments passed to \code{getModelObject}, for example \code{read.full}, and \code{store.internally}.}
}
\value{
A temporally averaged\code{ModelObject}  but with no-spatial averaging or cropping (ie. include complete original simulation domain). In other words, a map!
}
\description{
Given a \code{ModelRun} object, a \code{Quantity} object and a \code{TemporalExtent} object, return an appropriate temporally-averaged \code{ModelObject} oject for that run, quantity and time period over the whole simulation domain.
}
\seealso{
\code{getModelObject}, \code{getVegTemporal}
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
