% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-import.R
\name{import.raster}
\alias{import.raster}
\title{Import external data}
\usage{
import.raster(file, scale = 1, nodata = NA, nodata.limit = "eq",
  method = "bilinear", to = raster(nrows = 360, ncols = 720, xmn = -180, xmx
  = 180, ymn = -90, ymx = 90, crs = CRS("+proj=longlat +ellps=WGS84")),
  mem.only = FALSE)
}
\arguments{
\item{file}{path to the raster file.}

\item{scale}{scaling factor for the data.}

\item{nodata}{value or threshold for invalid pixels.}

\item{nodata.limit}{Any of "eq", "lt", "le", "ge" or "gt".}

\item{method}{raster interpolation method.}

\item{to}{target rater grid.}

\item{mem.only}{should the rasterOptions be set for operation in memory only.}
}
\value{
data.frame with columns c("Lon", "Lat", "value").
}
\description{
Import external raster data and return them as data.frame
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

