% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000rvc-tools.R
\docType{methods}
\name{is.equal}
\alias{is.equal}
\alias{is.equal,SpatialExtent,SpatialExtent-method}
\alias{is.equal,TemporalExtent,TemporalExtent-method}
\alias{is.equal,VegQuant,VegQuant-method}
\title{Checks two RVCTools metadata objects for equality}
\usage{
is.equal(a, b)

\S4method{is.equal}{VegQuant,VegQuant}(a, b)

\S4method{is.equal}{TemporalExtent,TemporalExtent}(a, b)

\S4method{is.equal}{SpatialExtent,SpatialExtent}(a, b)
}
\arguments{
\item{a}{a RVCTools metadata object}

\item{b}{another RVCTools metadata object of the same type}
}
\value{
logical
}
\description{
checks if two objects have the same quantity, spatial or temporal extent
}
\section{Methods (by class)}{
\itemize{
\item \code{a = VegQuant,b = VegQuant}: Checks two VegQuant objects for equality

\item \code{a = TemporalExtent,b = TemporalExtent}: Checks two TemoralExtent objects for equality

\item \code{a = SpatialExtent,b = SpatialExtent}: Checks two SpatialExtent objects for equality
}}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

