% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-methods.R
\name{is.object-methods}
\alias{is.object-methods}
\alias{is.DataObject}
\alias{is.ModelObject}
\alias{is.ComparisonLayer}
\alias{is.ModelRun}
\alias{is.Quantity}
\alias{is.SpatialComparison}
\title{Check if an object is a particular type of DGVMTools object.}
\usage{
is.DataObject(input, spatial = FALSE, temporal = FALSE, site = FALSE)

is.ModelObject(input, spatial = FALSE, temporal = FALSE, site = FALSE)

is.ComparisonLayer(input)

is.ModelRun(input)

is.Quantity(input)

is.SpatialComparison(input)
}
\arguments{
\item{input}{Any R object to be checked}

\item{spatial}{check if input is a spatial object}

\item{temporal}{check if input is a temporal object}

\item{site}{check if input is a site object}
}
\value{
logical
}
\description{
Returns TRUE if an object of the particular object type and FALSE otherwise
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}

Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}
