% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000rvc-tools.R
\name{is.VegObject}
\alias{is.VegObject}
\title{Check is an object is a \code{VegObject}.}
\usage{
is.VegObject(input)
}
\arguments{
\item{input}{Any R object to bec checked}
}
\value{
logical
}
\description{
Returns TRUE if an object is a \code{VegObject}, and FALSE otherwise
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

