% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPFTs.R
\name{listPFTs}
\alias{listPFTs}
\title{Select PFTs}
\usage{
listPFTs(x, criteria = NULL, return.ids = TRUE)
}
\arguments{
\item{x}{Either a Field, or a list of PFT objects}

\item{criteria}{A character string to use to select PFTs.  This compared to every the value of every slot of a PFT object.  If one matches, the PFT is inlucded in the return.}

\item{return.ids}{A logical, if TRUE (default) then return the the id of the PFT, if FALSE return the entire PFT object}
}
\value{
Either a list of characters (the ids of the PFTs) or a list of PFT object (depending on argument return.ids)
}
\description{
Uses PFT metadata to return all PFTs with a particular properties.  For example only tree PFTs, or only evergreen PFTs.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
