% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-utility-functions.R
\name{LondonCentre}
\alias{LondonCentre}
\title{London centre a gridcell}
\usage{
LondonCentre(lon)
}
\arguments{
\item{lon}{A vector longitude value to transform}
}
\value{
The transformed longitude values
}
\description{
Transform a vector of longitude values to lie in the range (-180,180) instead of (0,36)
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
