% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantities.R
\name{lookupQuantity}
\alias{lookupQuantity}
\title{Retrieve a Quantity from the standard list.}
\usage{
lookupQuantity(quant.id, model.str = "Standard", verbose = TRUE)
}
\arguments{
\item{quant.id}{String holding the id of the \code{Quantity} you want}

\item{model.str}{String holding the name of the model or "Standard"}

\item{verbose}{Logical, if TRUE (default) give some extra information}
}
\value{
A \code{Quantity} object if it finds the right one.  Otherwise the code halts.
}
\description{
Function to get an Quantity based on an ID string *and* the scope (i.e. a particular model or the 'Standard' properties) from the standard \code{dgvm.quantities} list
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
