% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpj-quantities.R
\docType{data}
\name{lpj.quantities}
\alias{lpj.quantities}
\title{Global scope list of all LPJ-GUESS (and SPITFIRE) quantities defined as \code{VegQuant} object 
from the default LPJ-GUESS model config and a few extra.}
\format{A list of \code{VegQuant} objects that store meta-data for output variable commonly used from LPJ-GUESS (and LPJ-GUESS-SPITFIRE)}
\usage{
lpj.quantities
}
\description{
These quantities have sensible colour schemes, long names and cut ranges for standard global runs,
but they will likely need to be modified and new ones aded for a specific analysis.
Once the package is loaded this is just a standard R list so can be modified and extended as you wish.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\seealso{
\code{VegQuant}-class
}
\keyword{datasets}

