% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tools.R
\name{makeDatasetID}
\alias{makeDatasetID}
\title{Make an ID string for a \code{ModelObject}}
\usage{
makeDatasetID(data.string, units = NULL, temporal.resolution = NULL,
  spatial.resolution = NULL, temporal.extent = NULL,
  spatial.extent = NULL, temporally.averaged = FALSE,
  spatially.averaged = FALSE)
}
\arguments{
\item{data.string}{Character string to describe the dataset}

\item{units}{Character string to describe the units in which the data are measured}

\item{temporal.resolution}{Character string to describe the temporal resolution of this data, eg. "Annual" or "Daily"}

\item{spatial.resolution}{Character string to describe the spatial resolution of this data eg "HD" (for half degree) or "QD" (for quarter degree)}

\item{temporal.extent}{The temporal extent of this object if it has been cropped from the orginal duration, otherwise NULL}

\item{spatial.extent}{The spatial extent of this object if it has been cropped from the orginal simulation extent, otherwise NULL}

\item{temporally.averaged}{Logical, should be TRUE if the data represent a snapshot or data averaged over a time period}

\item{spatially.averaged}{Logical, should be TRUE if the data have been averaged over spatial domain}
}
\value{
A character string
}
\description{
Given a string for the quantity and temporal and spatial extents and averaging flags, build an appropriate (and unique) ID string
for use in the \code{id} slot of a \code{ModelObject} and for filenames etc.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

