% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helper.R
\name{makeOverlay}
\alias{makeOverlay}
\title{Returns a SpatialLines object (of, for example, country or continent outlines) for overlaying on an spplot with (for example) plotSpatial}
\usage{
makeOverlay(which, interior.lines = TRUE, lwd = 0.5, col = NULL)
}
\arguments{
\item{which}{A string specifiying the name of one of the standard maps in from the map packages the packages directly (see packages maps and mapdata for documentation) or a special (eg paleo maps or other particular cases, currently undefined).}

\item{interior.lines}{A boolean specifying whether on not to plot the interior lines (countires or provences or other administraive regions) if available}

\item{lwd}{Width to plot the lines}

\item{col}{Colour to plot the lines}
}
\value{
a SpatialLines object
}
\description{
Returns a SpatialLines object (of, for example, country or continent outlines) for overlaying on an spplot with (for example) plotSpatial
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}, Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

