% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-helper.R
\name{makePlotTitle}
\alias{makePlotTitle}
\title{Make a plot title}
\usage{
makePlotTitle(quantity.str, layer = NULL, source = NULL, period = NULL,
  extent = NULL)
}
\arguments{
\item{quantity.str}{Character string for the quantity plotted}

\item{layer}{The names of the layer (or another identifier)}

\item{source}{The ModelObject, DataObject or ComparisonLayer that is being plotted}

\item{period}{The time period plotted as TemporalExtent (optional)}

\item{extent}{The spatial extent plotted as SpatialExtent (optional)}
}
\value{
A character string for use as a plot title
}
\description{
Build an appropriate plot title from some possibly relevant variables.  
It will use a string to represent the quantity (obligatory), and optionally a period and an ID.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
