% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePlotTitle.R
\name{makePlotTitle}
\alias{makePlotTitle}
\title{Make a plot title}
\usage{
makePlotTitle(fields)
}
\arguments{
\item{quantity.str}{Character string for the quantity plotted}

\item{layer}{The names of the layer (or another identifier)}

\item{source}{The Field or Comparison that is being plotted}

\item{first.year}{The first year of the data plotted}

\item{last.year}{The last year of the data plotted}

\item{extent.str}{The spatial extent plotted as described by a character string (optional)}
}
\value{
A character string for use as a plot title
}
\description{
Build an appropriate plot title from some possibly relevant variables.  
It will use a string to represent the quantity (obligatory), and optionally a period and an ID.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
