% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome-schemes.R
\docType{data}
\name{Megabiomes_dev.scheme}
\alias{Megabiomes_dev.scheme}
\title{Meta-data describing an unpublished "mega biome" scheme for LPJ-GUESS output.}
\format{An object of class \code{BiomeScheme} of length 1.}
\usage{
Megabiomes_dev.scheme
}
\description{
Note 'substitution' is probably a redundant slot, this is a part of the data, 
not the classication of the model output.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}

