% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names-methods.R
\docType{methods}
\name{names-methods}
\alias{names-methods}
\alias{names,Field-method}
\alias{names,Comparison-method}
\title{Names methods}
\usage{
\S4method{names}{Field}(x)

\S4method{names}{Comparison}(x)
}
\arguments{
\item{x}{A Field or DataObject}

\item{...}{Other arguments, not currently used}
}
\value{
A vector of character strings of the names of the layers
}
\description{
Methods for returning the names of the layers in a Field or a DataObject.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
