% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names-methods.R
\docType{methods}
\name{layer-names-methods}
\alias{layer-names-methods}
\alias{names,Field-method}
\alias{names,Comparison-method}
\alias{layers}
\alias{layers,Field-method}
\alias{layers,Comparison-method}
\alias{layers,data.table-method}
\title{Layer name methods}
\usage{
\S4method{names}{Field}(x)

\S4method{names}{Comparison}(x)

\S4method{layers}{Field}(x)

\S4method{layers}{Comparison}(x)

\S4method{layers}{data.table}(x)
}
\arguments{
\item{x}{A Field or Comparison.}
}
\value{
A vector of character strings of the names of the layers
}
\description{
Methods for returning the names of the layers in a Field or a Comparison.  Note, for changing the names of layers see \code{renameLayers}.
}
\details{
For convenience both \code{names()} and \code{layers()} are defined, but they both give the same result.  There is a logical equivalency here since 
layer names in a field are actually the column names in underlying data.table.  
AS such,  \code{layers()} is more consistent with the terminology and framework of DGVMTools, whereas \code{names()} is more consistent with the data.table approach.  
But, in the end, both of these functions return exactly the same thing via exactly the same method.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
