% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-GUESS.R
\name{openLPJOutputFile}
\alias{openLPJOutputFile}
\title{Open an LPJ-GUESS .out file}
\usage{
openLPJOutputFile(run, quant, target.sta, verbose = FALSE,
  data.table.only = FALSE, ...)
}
\arguments{
\item{run}{A \code{Source} containing the meta-data about the LPJ-GUESS run}

\item{quant}{A Quant to define what output file from the LPJ-GUESS run to open, 
can also be a simple string defining the LPJ-GUESS output file if the \code{return.data.table} argument is TRUE}

\item{verbose}{A logical, set to true to give progress/debug information}

\item{data.table.only}{A logical, if TRUE return a data.table and not a Field}

\item{first.year}{The first year (as a numeric) of the data to be return}

\item{last.year}{The last year (as a numeric) of the data to be return}
}
\value{
a data.table (with the correct tear offset and lon-lat offsets applied)
}
\description{
\code{openLPJOutputFile} returns a data.table object given a string defining a vegetation quantity 
from the run (eg. "lai", to read the file "lai.out") and  \code{Source} object which defines where the run is on disk and the offsets to apply
}
\details{
Note that the files can be gzipped on UNIX systems, but this might fail on windows systems.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
