% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-GUESS.R
\name{openLPJOutputFile_FireMIP}
\alias{openLPJOutputFile_FireMIP}
\title{Open an LPJ-GUESS .out file}
\usage{
openLPJOutputFile_FireMIP(run, variable, first.year, last.year,
  verbose = FALSE, soil_water_capacities = "none")
}
\arguments{
\item{run}{A \code{Source} containing the meta-data about the LPJ-GUESS run}

\item{variable}{A string the define what output file from the LPJ-GUESS run to open, for example "anpp" opens and read the "anpp.out" file}

\item{first.year}{The first year (as a numeric) of the data to be return}

\item{last.year}{The last year (as a numeric) of the data to be return}

\item{verbose}{A logical, set to true to give progress/debug information}
}
\value{
a data.table (with the correct tear offset and lon-lat offsets applied)
}
\description{
\code{openLPJOutputFile} returns a data.table object given a string defining a vegetation quantity 
from the run (eg. "lai", to read the file "lai.out") and  \code{Source} object which defines where the run is on disk and the offsets to apply
}
\details{
Note that the files can be gzipped on UNIX systems, but this might fail on windows systems.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
