% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-classes.R
\docType{class}
\name{PFT-class}
\alias{PFT-class}
\title{Class to hold the metadata for a Plant Functional Type (PFT)}
\description{
These will are defined in lists for the default PFTs (or common post-processing schemes in the case of aDGVM) for default model setups, but the user may well need to define their own
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular PFT.  Recommended to be alphanumeric because it is used to construct file names.}

\item{\code{name}}{A character string to describe the PFT. Used for building plot labels, not file names, so doesn't need to be alphanumeric and can so can be prettier.}

\item{\code{lifeform}}{A string defining the lifeform of the PFT, typically either "Tree", "Grass" or "Shrub"}

\item{\code{leafform}}{A string defining the leafform of the PFT, typically either "Broadleaved" or "Needleleaved"}

\item{\code{phenology}}{A string defining the phenology of the PFT, typically "Evergreen", "Summergreen", "Raingreen" or "GrassPhenology"}

\item{\code{zone}}{A string defining the climate zone of a PFT, typically "Boreal", "Temperate" or "Tropical" (could go crazy and also have "Mediterranean", for example)}

\item{\code{colour}}{A string defining a preferred R colour to plot this PFT (for line graphs etc)}

\item{\code{combine}}{A string defining the \code{id} of a PFT that this PFT should be combined with when combining PFT with similar functioning but different shade tolerance.  Used primarily with LPJ-GUESS.}
}}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

