% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/Format-GUESS.R, R/Format-aDGVM.R
\docType{class}
\name{PFT-class}
\alias{PFT-class}
\alias{GUESS.PFTs}
\alias{aDGVM.PFTs}
\title{Class to hold the metadata for a Plant Functional Type (PFT)}
\format{An S4 class object with the slots as defined below.}
\usage{
GUESS.PFTs

aDGVM.PFTs
}
\description{
This is a class to hold meta-data about PFTs.  As detailed in the 'Slots' section below, this includes an id (should be unique) and a name, 
as well as their growth form, phenology, leaftype, climate zone, shade tolerance and a default plot colour.
These are defined in lists for the default PFTs for supported model format (see'Usage' below).
}
\details{
The \code{PFT-class} is only meta-data but is very important as it allows users to calculate, for example, the total Tree LAI with a simple command. 
The standard PFTs for some models are included (see above), but if you have other PFTs then you simply need define them.  You then combine the PFTs into an R list
and then provide them as the 'pft.set' argument to the \code{defineSource} call and bingo! you are using your custom PFTs.
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular PFT.  Recommended to be alphanumeric because it is used to construct file names.}

\item{\code{name}}{A character string to describe the PFT. Used for building plot labels, not file names, so doesn't need to be alphanumeric and can so can be prettier.}

\item{\code{growth.form}}{A string defining the growth.form of the PFT, typically either "Tree", "Grass" or "Shrub"}

\item{\code{leaf.form}}{A string defining the leaf.form of the PFT, typically either "Broadleaved" or "Needleleaved"}

\item{\code{phenology}}{A string defining the phenology of the PFT, typically "Evergreen", "Summergreen", "Raingreen" or "GrassPhenology"}

\item{\code{climate.zone}}{A string defining the climate climate.zone of a PFT, typically "Boreal", "Temperate" or "Tropical" (could go crazy and also have "Mediterranean", for example)}

\item{\code{shade.tolerance}}{A string defining the \code{id} of a shade tolerance characteristric of a PFT.}

\item{\code{colour}}{A string defining a preferred R colour to plot this PFT (for line graphs etc)}
}}

\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{datasets}
