% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-ggplot2.R
\name{plotGGCategorialAggregated}
\alias{plotGGCategorialAggregated}
\title{Plot a categorically aggregated summary.}
\usage{
plotGGCategorialAggregated(x, cat, x.col.name = "value",
  cat.col.name = "category", name.map = NA, area.weighted = TRUE,
  vertical = FALSE, bar = FALSE, plot = TRUE, ...)
}
\arguments{
\item{x}{A Field or a list of them.}

\item{cat}{A Field containing the categories to summarize 'x' or a list of them.}

\item{x.col.name}{The column name of 'x' data.table to use. Default: 'value'.}

\item{cat.col.name}{The column name of the grouping variable. Should be either of type character, factor or integer.}

\item{name.map}{a named vector, to translate numerical category-data into human understandable names.}

\item{area.weighted}{weight the mean by the gridcell area (default: TRUE) or not.}

\item{vertical}{boolean: values on the y-axis, categories on the x-axis}

\item{bar}{boolean: use bars instead of points.}

\item{plot}{If FALSE only the data is returned, without drawing the map.}

\item{...}{Ignored further parameters.}
}
\value{
A ggplot object, which can either be printed directly or further modified, or a data.table if plot is FALSE.
}
\description{
Plots vertical lines as average over latitudinal bands.
}
\examples{
message("See templates/Example.ggplot.R")
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}
