% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvc-ggplot2.R
\name{plotGGCategorialAggregated}
\alias{plotGGCategorialAggregated}
\title{Plot a categorically aggregated summary.}
\usage{
plotGGCategorialAggregated(input, targets = NULL, name.map = NA,
  area.weighted = TRUE, long.title = TRUE, vertical = FALSE,
  bar = FALSE, plot = TRUE, ...)
}
\arguments{
\item{input}{A VegSpatial or a list of VegSpatial objects.}

\item{targets}{data.frame/list, either specifying x/y columns as c('VegSpatial name', 'column name') or slot/column columns as c('x', 'y').}

\item{name.map}{a named vector, to translate numerical y-data into human understandable names}

\item{area.weighted}{weight the mean by the gridcell area (default: TRUE) or not.}

\item{long.title}{If the description (default) should be used as titles or the shorter id.}

\item{vertical}{boolean: values on the y-axis, categories on the x-axis}

\item{bar}{boolean: use bars instead of points}

\item{plot}{If FALSE only the data is returned, without drawing the map.}

\item{...}{Ignored further parameters}
}
\value{
A ggplot object, which can either be printed directly or further modified, or a data.table if plot is FALSE.
}
\description{
Plots vertical lines as average over latitudinal bands.
}
\examples{
paste("See templates/Example.ggplot.R")
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

