% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvc-ggplot2.R
\name{plotGGMeridional}
\alias{plotGGMeridional}
\title{Plot a meridional mean with ggplot2}
\usage{
plotGGMeridional(input, column = "value", what = list(center = "mn", var =
  "sd"), alpha = c(0.8, 0.1), colors = NA, long.title = TRUE,
  plot = TRUE, ...)
}
\arguments{
\item{input}{A VegSpatial or a list of VegSpatial objects.}

\item{column}{Name of the column(s) in the data slot of input, which should be plotted. Several columns can only be supplied with a single VegSpatial. Default: 'value'.}

\item{what}{A list with the elements 'center' and 'var'. 'center' can be one of 'mn', 'mean', 'md', 'median' for the line. 'var' is for the variability polygon drawn in the background. 'var' can be 'sd' (sandard deviation), 'se' (standard error), 'mm'/'minmax', a numeric value (0-1) of quantile to be shown arround the central line or a vector specifying the upper and lower quantile.}

\item{alpha}{Vector of two elements: Transparency values for the line (first) and background polygon (second).}

\item{colors}{Colors for the diffent VegSpatial objects.}

\item{long.title}{If the description (default) should be used as titles or the shorter id.}

\item{plot}{If FALSE only the data is returned, without drawing the map.}

\item{...}{Ignored further parameters}
}
\value{
A ggplot object, which can either be printed directly or further modified, or a data.table if plot is FALSE.
}
\description{
Plots vertical lines as average over latitudinal bands.
}
\examples{
message("See templates/Example.ggplot.R")
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

