% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-ggplot2.R
\name{plotGGSpatial}
\alias{plotGGSpatial}
\title{Spatial plot (map) with ggplot2}
\usage{
plotGGSpatial(input, column = "value", colors = NA, sym.col = FALSE,
  wrap = 1, terr.bg = NA, map.overlay = NA, long.title = TRUE,
  plot = TRUE, ...)
}
\arguments{
\item{input}{A VegSpatial or a list of VegSpatial objects.}

\item{column}{Name of the column(s) in the data slot of input, which should be plotted. Several columns can only be supplied with a single VegSpatial. Default: 'value'.}

\item{colors}{A (named) vector of colors or a data.frame with the columns 'color' and ('name' or 'value'), to be used for plotting.}

\item{sym.col}{boolean, if the colors should be distributed symetrically around 0.}

\item{wrap}{a single number of facet_wrap columns or a vector/list with the run, VegSpatial name, column and optionally ncol (number of columns), which is used to split the data in different panels. Only valid when a vector of column names or a list of VegSpatial was given as input. Otherwise it is ignored.}

\item{terr.bg}{which colour should be used for missing terrestial pixels (e.g. Greenland)}

\item{map.overlay}{if set to 'lowres' or 'highres' national borders are included (uses package maps). It can also be a data.frame as returned by \code{ggplot2::fortify}.}

\item{long.title}{If the description (default) should be used as titles or the shorter id.}

\item{plot}{If FALSE only the data is returned, without drawing the map.}

\item{...}{Additional parameters, which are ignored so far.}
}
\value{
A ggplot object, which can either be printed directly or further modified, or a data.table if plot is FALSE.
}
\description{
Plots a map (or a set of maps) using the geom_raster from the ggplot2 library
}
\examples{
message("See templates/Example.ggplot.R")
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}

