% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rvc-ggplot2.R
\name{plotGGTemporal}
\alias{plotGGTemporal}
\title{Plot a timeseries}
\usage{
plotGGTemporal(input, columns = "value", scale = 1, colors = NA,
  type = "line", wrap = NA, long.title = TRUE, lty = "sens",
  alpha = NA, plot = TRUE, ...)
}
\arguments{
\item{input}{a VegTemporal object or a list of several}

\item{colors}{Colors for the diffent VegSpatial objects or columns.}

\item{type}{"line", overlayed "area" or "stack"ed area}

\item{wrap}{if input is a list and columns has more than one element use "sens" or "column" fro wrapping intp several panels.}

\item{long.title}{If the description (default) should be used as titles or the shorter id.}

\item{lty}{Only used if type="line" and if input is a list and columns has more than one element and no wrapping is desired, use either "sens" or "column" for different line types.}

\item{alpha}{alpha value for plots}

\item{plot}{If FALSE only the data is returned, without drawing the map.}

\item{...}{Ignored further parameters
Plot a timeseries 

Plot a timeseries with ggplot. Either as line default, overlaying area,
applicable p.e. for height or stacked polygons (e.g. for succession).}

\item{column}{the column(s) to display. Several columns can only be supplied with a single VegSpatial. Default: 'value'.}

\item{input}{a VegTemporal object or a list of several.}

\item{column}{The column(s) to display. Default: 'value'.}

\item{colors}{Colors for the diffent VegSpatial objects or columns. Must have the same length otherwise colors are choosing automatically.}

\item{long.title}{If the description (default) should be used as titles or the shorter id.}

\item{plot}{If FALSE only the data is returned, without drawing the map.}

\item{...}{Ignored further parameters}
}
\description{
Plot a timeseries with ggplot. Years are on the x-axis
}

