% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{plotResidualsHisto}
\alias{plotResidualsHisto}
\title{Plot the residuals from a DataObject as histogram}
\usage{
plotResidualsHisto(data.obj, cols = NULL, fills = NULL, types = NULL,
  labels = NULL, title = NULL, xlab = NULL, ylab = NULL, alpha = NULL,
  reverse = FALSE, bins = NULL, limit = TRUE)
}
\arguments{
\item{data.obj}{The DataObject for which to plot the residuals}

\item{cols, types, fills}{Vector of line colours, line types and fill colours respectively. Each can be left empty but if defined they must have one entry for each set of residuals you are plotting.}

\item{labels}{Vector of character for label the histos. Can be left empty (defaults to the run id) but if defined it must have one entry for each set of residuals you are plotting.}

\item{title}{Character for plot title (optional)}

\item{xlab}{Character for x-axis label (optional)}

\item{ylab}{Character for y-axis label (optional)}

\item{alpha}{Numerical value [0,1] to specify opacity of histo fill colours (optional)}

\item{reverse}{Logical if TRUE reverse the layering of the histograms (optional, default = FALSE)}

\item{bins}{Numerical vector (evenly spaced) defining the histogram bins (optional)}

\item{limit}{Logical if TRUE (and argument bins provided) limit x-axis to the bins given.  If FALSE (and argument bins provided) then bins defines the bin width, but the plot is not limited}
}
\value{
A ggplot2 plot.
}
\description{
Function will overlay histograms if more than one ModelRun has been compared to the the DataObject.  
Line colours, line types and fill colours can all be specified but have sensible defaults.
}
\details{
This function should be called after a call to \code{benckmarkSpatial} for a DataObject.  It plots the residuals for each model run to which it was compared.  
It is called automatically by \code{benckmarkSpatial}, but can be called again for better flexibility.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

