% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScatter.R
\name{plotScatter}
\alias{plotScatter}
\title{Make A scatter plot}
\usage{
plotScatter(x, y = x, layer.x, layer.y = layer.x, alpha = 1,
  text.multiplier)
}
\arguments{
\item{x}{The first Field, Comparison or data., object from which the data to be plotted should be taken.}

\item{y}{The second DGVMTools::Field object from which the data to be plotted should be taken.
Default value is x.}

\item{layer.x}{The first layer to be plotted (taken from x)}

\item{layer.y}{The second layer to be plotted (taken from y).  Defaults to layer.x.}

\item{alpha}{Numeric between 0 and 1 specifing the transparency of the points.  Default is 1 (= fully opaque).}

\item{text.multiplier}{A number specifying an overall multiplier for the text on the plot.
Make it bigger if the text is too small on large plots and vice-versa.}
}
\value{
A ggplot2 object
}
\description{
This simple function makes (and returns it as an object, it doesn't print it) a simple scatter plot (using ggplot2). 
The data are two layers from either one or two Field objects.  Data points which appear in one Field but not the other are excluded
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
