% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScatter.R
\name{plotScatter}
\alias{plotScatter}
\title{Make A scatter plot}
\usage{
plotScatter(field1, field2 = field1, layer1, layer2 = layer1)
}
\arguments{
\item{field1}{The first DGVMTools::Field object from which the data to be plotted should be taken.}

\item{field2}{The second DGVMTools::Field object from which the data to be plotted should be taken.
Default value is field1.}

\item{layer1}{The first layer to be plotted (taken from field1)}

\item{layer2}{The second layer to be plotted (taken from field2).  Defaults to layer1.}
}
\value{
A ggplot2 object
}
\description{
This simple function makes (and returns it as an object, it doesn't print it) a simple scatter plot (using ggplot2). 
The data are two layers from either one or two Field objects.  Data points which appear in one Field but not the other are excluded
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
