% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-plotting.R
\name{plotScatterComparison}
\alias{plotScatterComparison}
\title{Make a scatter plot by comparing two rasters}
\usage{
plotScatterComparison(model, data, run, data.name, quant, stat.results = NULL)
}
\arguments{
\item{model}{Raster of the model}

\item{data}{Raster of the data}

\item{run}{The \code{VegRun} object for the run plotted (optional)}

\item{data.name}{Character string for the data}

\item{quant}{The quantity plotted (as \code{VegQuant} object)}

\item{stat.results}{The \code{RasterComparion} object if it has already been calculated

The plot is saved to the run directory of the run object}
}
\description{
Takes either two rasters or a \code{RasterComparion} object and makes a scatter plot.
}
\details{
Probably should be made a method of \code{RasterComparion}, 
with further additions to that class necessary.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

