% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScatterComparison.R
\name{plotScatterComparison}
\alias{plotScatterComparison}
\title{Plot scatter of model vs data}
\usage{
plotScatterComparison(input.CLayers, run.ids = NULL, run.labels = NULL,
  wrap = TRUE, facet.labels = NULL, xlim = NULL, ylim = NULL,
  xlab = NULL, ylab = NULL, text.multiplier = NULL,
  showFitLine = TRUE, showStats = TRUE, alpha = 0.05,
  text.size = 6)
}
\arguments{
\item{input.CLayers}{The DataObject for which to plot the residual}

\item{run.ids}{The character vector of run ids of the runs to scatter against the data (must be a vector, not a list).  Leave blank to compare all runs that have been previouslty compared to this dataset.}

\item{run.labels}{A vector of more descriptive strings from each run (each the run@names)}

\item{wrap}{An optional string identifying an additional column by which to subdivide (facet) that data}

\item{facet.labels}{An option vector of strings giving descriptive labels for facets, if provided it *must* be a named vector, 
where each element has a name corresponding to the value in the facet column.}

\item{xlim}{An optional vector of two numerics to define the x plot range}

\item{ylim}{An optional vector of two numerics to define the y plot range}

\item{xlab}{An optional character string for the x-axis}

\item{ylab}{An optional character string for the y-axis}

\item{text.multiplier}{A number specifying an overall multiplier for the text on the plot.
Make it bigger if the text is too small on large plots and vice-versa.}

\item{showFitLine}{Boolean, if TRUE shows a linear fit and the fit equation}

\item{showStats}{Boolean, if TRUE show some stats do quantify how well the model fits the data}

\item{alpha}{Numeric for the transparency of tre points (range (0,1], default = 0.05))}

\item{text.size}{Numeric to scale the text size on the plot (default = 6)}

\item{labels}{Character vector of labels (one for each run).  If not provided uses the run ids}
}
\value{
A ggplot2 plot.
}
\description{
Function will makes multiple scatter plots on one page if more than one Source has been compared to the the DataObject.
}
\details{
This function should be called after a call to \code{benckmarkSpatial} for a DataObject.  It plots the scatters for each model run to which it was compared .  
It is called automatically by \code{benckmarkSpatial}, but can be called again for better flexibility.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
