% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-spatial.R
\name{plotSpatial}
\alias{plotSpatial}
\title{Plot maps from a temporally-averaged \code{ModelObject}, \code{DataObject} or \code{ComaprisonLayer} (and lists thereof)}
\usage{
plotSpatial(data, layers = NULL, title = NULL, layout.objs = NULL,
  facet.labels = NULL, facet.order = NULL, plot.bg.col = "white",
  useLongnames = FALSE, text.multiplier = NULL, xlim = NULL,
  ylim = NULL, limits = NULL, override.cols = NULL,
  override.cuts = NULL, map.overlay = NULL, dont.grid = FALSE,
  return.data = FALSE, tile = TRUE, interpolate = FALSE,
  interior.lines = TRUE)
}
\arguments{
\item{data}{The data to plot. Can be a ModelObject, data.table, a SpatialPixelsDataFrame or a Raster* object.}

\item{layers}{A list of strings specifying which layers to plot.  Defaults to all layers.}

\item{title}{A character string to override the default title.}

\item{layout.objs}{A character string name of a map overlay from the \code{maps} or \code{mapdata} packages. For example, "world" or "worldHires".
Note that using these, especially "worldHires", can add quite a bit off time.}

\item{facet.labels}{List of character strings to be used as panel labels for summary plots and titles for the individual plots.
Sensible titles will be constructed if this is not specified.}

\item{facet.order}{A vector of the characters that, if supplied, control the order of the facets.  To see what these values are you can call this funtion with "return.data=TRUE"
and check the values of the XXXX column.  But generally they will be the values of the @names slots of the Data/ModelObjects and/or the layers (as layers plotted as defined by the layers arguments 
in this function).}

\item{plot.bg.col}{Colour string for the plot background.}

\item{useLongnames}{Boolean, if TRUE replace PFT IDs with the PFT's full names on the plots. NOT CURRENLTLY IMPLEMENTED!!}

\item{text.multiplier}{A number specifying an overall multiplier for the text on the plot.
Make it bigger if the text is too small on large plots and vice-versa.}

\item{xlim}{An optional vector of two numerics to specify the x/longitude range of the plot.}

\item{ylim}{An optional vector of two numerics to specify the y/latitude range of the plot.}

\item{limits}{A numeric vector with two members (lower and upper limit) to limit the plotted values.}

\item{override.cols}{A colour palette function to override the defaults.}

\item{override.cuts}{Cut ranges (a numeric vector) to override the defaults.}

\item{map.overlay}{A character string specifying which map overlay (from the maps and mapdata packages) should be overlain.}

\item{dont.grid}{Boolean, if TRUE then don't use facet_grid() to order the panels in a grid.  Instead use facet_wrap().
Useful when not all combinations of Sources x Layers exist which would leave blank panels.}

\item{return.data}{Return a data.table with the final data instead of the ggplot object.  This can be useful for inspecting the structure of the facetting columns, amongst other things.}

\item{tile}{Boolean If true use the ggplot function geom_tile() instead of the function geom_raster().}

\item{interpolate}{Boolean, whether or not to use interpolation in the case of a call to geom_raster().  Probably not a good idea, results tend to hurt my eyes.}

\item{interior.lines}{Boolean, if TRUE plot country lines with the continent outlines of the the requested map.overlay
Other things can be overlain on the resulting plot with further ggplot2 commands.}
}
\value{
Returns a plot object (from spplot)
}
\description{
This is a heavy lifting function for plotting maps from ModelObjects, DataObjects, and ComparisonLayers (and lists of those things) with flexibility, but also with a high degree of automation. 
As a consequence, iIt has a really large amount of parameters for a huge amount of flexibility.  However they are all set to sensible defaults.  In principle you can supply only the objext and it will plot.
It is basically a complex wrapper for the ggplot2 function geom_raster() and it returns a ggplot object, which will need to be displayed using a \code{print()} command.  Note that this object can be firther modified 
using further ggplot2 commands.
}
\details{
This function is heavily used by the benchmarking functions and can be very useful to the user for making quick plots
in standard benchmarking and post-processing.  It is also highly customisable for final results plots for papers and so on.
However, the \code{plotGGSpatial} function makes pretty plots with a simpler syntax, but with less flexibility.

The function works best for \code{ModelObjects} (which contain a lot of useful metadata).
}
\seealso{
\code{plotGGSpatial}, \code{expandLayers}, \code{sp::spplot}, \code{latice::levelplot}
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
