% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatial.R
\name{plotSpatial}
\alias{plotSpatial}
\title{Plot maps from a \code{\linkS4class{Field}} or a lists of \code{\linkS4class{Field}}}
\usage{
plotSpatial(
  fields,
  layers = NULL,
  title = character(0),
  subtitle = character(0),
  legend.title = character(0),
  facet.order = NULL,
  plot.bg.col = "white",
  panel.bg.col = "white",
  useLongNames = FALSE,
  text.multiplier = NULL,
  xlim = NULL,
  ylim = NULL,
  years = NULL,
  days = NULL,
  months = NULL,
  seasons = NULL,
  limits = NULL,
  cols = NULL,
  cuts = NULL,
  drop.cuts = TRUE,
  map.overlay = NULL,
  plot = TRUE,
  tile = FALSE,
  pixel.size = NULL,
  ...
)
}
\arguments{
\item{fields}{The data to plot. Can be a Field or a list of Fields.}

\item{layers}{A list of strings specifying which layers to plot.  Defaults to all layers.}

\item{title}{A character string to override the default title.  Set to NULL for no title.}

\item{subtitle}{A character string to override the default subtitle. Set to NULL for no subtitle.}

\item{legend.title}{A character string or expression to override the default legend title. Set to NULL for no legend title.  The default legend title is the \code{units}
of the \code{\linkS4class{Quantity}} of the first \code{\linkS4class{Field}} provided in the \code{field} argument.  This argument allows general flexibility, but it is particularly handy
to facilitate expressions for nicely marked up subscript and superscript.}

\item{facet.order}{A vector of the characters that, if supplied, control the order of the facets.  To see what these values are you can call this funtion with "plot=FALSE"
and check the values of the Facet column.  But generally they will be the values of the @names slots of the Data/Fields and/or the layers (as layers plotted as defined by the layers arguments 
in this function).}

\item{plot.bg.col}{Colour string for the plot background, default "white".}

\item{panel.bg.col}{Colour string for the panel background, default "white".}

\item{useLongNames}{Boolean, if TRUE replace Layer IDs with the Layer's full names on the plots.}

\item{text.multiplier}{A number specifying an overall multiplier for the text on the plot.
Make it bigger if the text is too small on large plots and vice-versa.}

\item{xlim}{An optional vector of two numerics to specify the x/longitude range of the plot.}

\item{ylim}{An optional vector of two numerics to specify the y/latitude range of the plot.}

\item{years}{An optional numeric vector specifying which years to plot (take care, this defaults to all the years in the input Fields which can be a lot!)}

\item{days}{An optional numeric vector specifying which days to plot (take care, this defaults to all the days in the input Fields which can be a lot!)}

\item{months}{An optional numeric vector specifying which months to plot(defaults to all the days in the input Fields)}

\item{seasons}{An optional character vector specifying which seasons to plot (any or all of "DJF", "MAM, "JJA", "SON", defaults to all the seasons in the input Fields)}

\item{limits}{A numeric vector with two members (lower and upper limit) to limit the plotted values.}

\item{cols}{A colour palette function to override the defaults.}

\item{cuts}{Cut ranges (a numeric vector) to override the default colour delimitation,  discretise the data into discrete colour bands}

\item{drop.cuts}{Logical, if TRUE then drop cut at each end which do not have any data in them in order more to fully use the colour scale.
Default is TRUE.  Ignored if 'cuts' argument is not used.}

\item{map.overlay}{An optional character string specifying either "coastlines" (alternatively "ne_coastlines") or "countries" (alternatively "ne_countries" or,
for backwards compatibility with previous versions "world") to plot either the coastlines or coastlines and country borders from the ranturalearth(data) packages.
Other things can be overlain on the resulting plot with further ggplot2 commands, and check the rnaturalearth package family for higher res versions, lakes, etc.}

\item{plot}{Boolean, if FALSE return a data.table with the final data instead of the ggplot object.  This can be useful for inspecting the structure of the facetting columns, amongst other things.}

\item{tile}{Logical, if TRUE use \code{geom_tile} instead of \code{geom_raster}.  The advantage is that plots made with \code{geom_tile} are more malleable and can, 
for example, be plotted on polar coordinates.  However \code{geom_tile} is much slower than \code{geom_raster}.}

\item{pixel.size}{Numeric, allows you to alter the plotted pixel size (height and width simultaneously using the same value).  This is useful 
if you are plotting a collection of individual sites which do not have regular spacing.  Note the "tile = TRUE" (see above) will automatically set if you haven't done it manually.}

\item{...}{Arguments passed to \code{ggplot2::facet_wrap()}.  See the ggplot2 documentation for full details but the following are particularly useful.
\itemize{
 \item{"nrow"}{The number of rows of facets}
 \item{"ncol"}{The number of columns of facets}
 \item{"scales"}{Whether the scales (ie. x and y ranges) should be fixed for all facets.  Options are "fixed" (same scales on all facets, default)
 "free" (all facets can their x and y ranges), "free_x" and "free_y"  (only x and y ranges can vary, respectively).}
 \item{"labeller"}{A function to define the labels for the facets.  This is a little tricky, please look to the ggplot2 documentation.
 But basically what you want is to define a named character vector, the names are the previous facet names and the values are the new names.  
 Then make this into a function by passing it to the ggplot function "as.labeller", and then that becomes your 'labeller' argument.} 
}}
}
\value{
Returns a ggplot object
}
\description{
This is a heavy lifting function for plotting maps from Fields with flexibility, but also with a high degree of automation. 
As a consequence, it has a really large amount of parameters for a huge amount of flexibility.  However they are all set to sensible defaults.  
In principle you can supply only the Fields and it will plot something sensible. It extracts the relevant data from the Fields, bashes it into
a data.table then calls ggplot2 function geom_raster() and it returns a ggplot object, which will need to be displayed using a \code{print()} command.  Note that this object can be firther modified 
using further ggplot2 commands.
}
\details{
This function is the main spatial plotting functions and can be very useful to the user for making quick plots
in standard benchmarking and post-processing.  It is also highly customisable for final results plots for papers and so on.
}
\seealso{
\link{plotTemporal}, \link{plotSpatialComparison}
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
