% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatialComparison.R
\name{plotSpatialComparison}
\alias{plotSpatialComparison}
\title{Plot a comparison between two spatial layers}
\usage{
plotSpatialComparison(
  comparisons,
  type = c("difference", "percentage.difference", "values", "nme"),
  limits = NULL,
  legend.title,
  panel.bg.col = "white",
  override.cols = NULL,
  symmetric.scale = TRUE,
  do.phase = FALSE,
  ...
)
}
\arguments{
\item{comparisons}{The data to plot, must be a Comparison or a list of Comparisons}

\item{type}{A character specifying what type of plot to make. Can be "difference" (default, for a difference plot), "percentage.difference", "values" 
(actual values, side-by-side) or "nme" (for the Normalised Mean Error, not yet implemented)}

\item{limits}{A numeric vector with two members (lower and upper limit) to limit the plotted values.}

\item{legend.title}{A character string or expression to override the default legend title. Set to NULL for no legend title.  The default legend title is the \code{units}
of the \linkS4class{Quantity} of the first \linkS4class{Comparison} provided in the \code{comparisions} argument.  This argument allows general flexibility, but it is particularly handy
to facilitate expressions for nicely marked up subscript and superscript.}

\item{panel.bg.col}{Colour string for the panel background, default to "white" for absolute values plots, and a grey for difference plots.}

\item{override.cols}{A colour palette function to override the defaults.}

\item{symmetric.scale}{If plotting a differences, make the scale symmetric around zero (default is TRUE)}

\item{do.phase}{Logical, only applies to plotting Comparison objects of type "seasonal".
If TRUE plot the the seasonal phase, if FALSE (the default), plot the seasonal concentration.}

\item{...}{Parameters passed to \link{plotSpatial}}
}
\value{
Returns a ggplot object
}
\description{
This function is for plotting maps from Comparison objects (or a list of those Comparisons).  Three types of comparisons plots are supported: 'difference' - 
a difference map; "values" - the absolute values plotted in panels and "percentage.difference" - the percentage differences.
}
\details{
A wrapper for around \link{plotSpatial} to plot the spatial Comparisons as maps.  Extra arguments to \link{plotSpatial} can also be specified.
}
\seealso{
\link{plotSpatial},  \link{compareLayers}
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
