% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatialComparison.R
\name{plotSpatialComparison}
\alias{plotSpatialComparison}
\title{Plot a comparion between two layers}
\usage{
plotSpatialComparison(comparisons, type = c("difference",
  "percentage.difference", "values", "nme"), limits = NULL,
  panel.bg.col = "white", override.cols = NULL,
  symmetric.scale = TRUE, percentage.difference.limit = 300,
  do.phase = FALSE, ...)
}
\arguments{
\item{comparisons}{The data to plot, must be a Comparison or a list of Comparisons}

\item{type}{A character specifying what type of plot to make. Can be "difference" (default, for a difference plot), "percentage.difference", "values" 
(actual values, side-by-side) or "nme" (for the Normalised Mean Error, not yet implemented)}

\item{limits}{A numeric vector with two members (lower and upper limit) to limit the plotted values.}

\item{panel.bg.col}{Colour string for the panel background, default to "white" for absolute values plots, and a sort of blue grey for difference plots.}

\item{override.cols}{A colour palette function to override the defaults.}

\item{symmetric.scale}{If plotting a differences, make the scale symmetric around zero (default is TRUE)}

\item{percentage.difference.limit}{If percentage difference to be plotted, what to limit the scale to.  Default is 300.}

\item{do.phase}{Logical, only applies to plotting Comparison objects of type "seasonal".
If TRUE plot the the seasonal phase, if FALSE (the default), plot the seasonal concentration.}

\item{...}{Parameters passed to \code{plotSpatial()}}
}
\value{
Returns a ggplot object
}
\description{
This function is for plotting maps from Comparison objects (or a list of those Comparisons).  Three types of comparisons plots are supported: 'difference' - 
a difference map; "values" - the absolute values plotted in panels and "percentage.difference" - the percentage differences.
}
\details{
A wrapper for around \code{plotSpatial()} to plot the spatial Comparisons as maps.  Extra arguments to \code{plotSpatial} can also be specified.
}
\seealso{
\code{plotSpatial},  \code{compareLayers}
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
