% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatialComparison.R
\name{plotSpatialComparison}
\alias{plotSpatialComparison}
\title{Plot a comparion between two layers}
\usage{
plotSpatialComparison(sources, type = c("difference",
  "percentage.difference", "values", "nme"), limits = NULL,
  panel.bg.col = "white", override.cols = NULL,
  symmetric.scale = TRUE, percentage.difference.limit = 300, ...)
}
\arguments{
\item{sources}{The data to plot, must be a Comparison or a list of Comparisons}

\item{type}{A character specifying what type of plot to make. Can be "difference" (default, for a difference plot), "percentage.difference", "values" 
(actual values, side-by-side) or "nme" (for the Normalised Mean Error, not yet implemented)}

\item{limits}{A numeric vector with two members (lower and upper limit) to limit the plotted values.}

\item{panel.bg.col}{Colour string for the panel background, default to "white" for absolute values plots, and a sort of blue grey for difference plots.}

\item{override.cols}{A colour palette function to override the defaults.}

\item{symmetric.scale}{If plotting a differences, make the scale symmetric around zero (default is TRUE)}

\item{percentage.difference.limit}{If precentage difference to be plotted, what to limit the scale to.}

\item{...}{Parameters passed to \code{plotSpatial()}}
}
\value{
Returns a ggplot object
}
\description{
This is a heavy lifting function for plotting maps from Fields, DataObjects, and Comparisons (and lists of those things) with flexibility, but also with a high degree of automation. 
As a consequence, it has a really large amount of parameters for a huge amount of flexibility.  However they are all set to sensible defaults.  In principle you can supply only the objext and it will plot.
It is basically a complex wrapper for the ggplot2 function geom_raster() and it returns a ggplot object, which will need to be displayed using a \code{print()} command.  Note that this object can be firther modified 
using further ggplot2 commands.
}
\details{
A wrapper for around \code{plotSpatial()} to plot the spatial Comparisons as maps.
}
\seealso{
\code{plotGGSpatial}, \code{expandLayers}, \code{sp::spplot}, \code{latice::levelplot}
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
