% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTemporal.R
\name{plotTemporal}
\alias{plotTemporal}
\title{Plot temporal data}
\usage{
plotTemporal(fields, layers = NULL, gridcells = NULL,
  title = character(0), subtitle = character(0), cols = NULL,
  col.by = "Layer", col.labels = waiver(), types = NULL,
  type.by = NULL, type.labels = waiver(), sizes = NULL,
  size.by = NULL, size.labels = waiver(), alphas = NULL,
  alpha.by = NULL, alpha.labels = waiver(), y.label = NULL,
  y.lim = NULL, x.label = NULL, x.lim = NULL,
  legend.position = "bottom", text.multiplier = NULL, plot = TRUE,
  ...)
}
\arguments{
\item{fields}{The data to be plotted, either as a Field, DataObject or a list of Model/DataObjects.}

\item{layers}{A list of strings specifying which layers to plot.  Defaults to all layers.}

\item{gridcells}{A list of gridcells to be plotted in different panels, for formatting of this argument see \code{selectGridcells}.
Leave empty or NULL to plot all gridcells (but note that if this involves too many gridcells the code will stop)}

\item{title}{A character string to override the default title.  Set to NULL for no title.}

\item{subtitle}{A character string to override the default subtitle. Set to NULL for no subtitle.}

\item{col.by, type.by, size.by, alpha.by}{Character strings defining the aspects of the data which which should be used to set the colour, line type, line size (width) and alpha (transparency).
Can meaningfully take the values "Layer", "Source" or "Quantity". By default \code{col.by} is set to "Layer" and all others set to NULL, which means the different aspects are 
distinguished by different facet panels.  Thus the standard behaviour the that different Layers are distinguished by different colours, but everything is seperated into different panels.}

\item{x.label, y.label}{Character strings for the x and y axes (optional)}

\item{x.lim, y.lim}{Limits for the x and y axes (each a two-element numeric, optional)}

\item{legend.position}{Position of the legend, in the ggplot2 style.  Passed to the ggplot function \code{theme()}. Can be "none", "top", "bottom", "left" or "right" or two-element numeric vector}

\item{text.multiplier}{A number specifying an overall multiplier for the text on the plot.}

\item{plot}{Logical, if FALSE return the data.table of data instead of the plot}

\item{...}{Arguments passed to \code{ggplot2::facet_wrap()}.  See the ggplot2 documentation for full details but the following are particularly useful.
\itemize{
 \item{"nrow"}{The number of rows of facets}
 \item{"ncol"}{The number of columns of facets}
 \item{"scales"}{Whether the scales (ie. x and y ranges) should be fixed for all facets.  Options are "fixed" (same scales on all facets, default)
 "free" (all facets can their x and y ranges), "free_x" and "free_y"  (only x and y ranges can vary, respectively).}
 \item{"labeller"}{A function to define the labels for the facets.  This is a little tricky, please look to the ggplot2 documentation} 
}}

\item{labels}{A list of character strings which are used as the labels for the lines.  Must have the same length as the layers argument (after expansion if necessary)}
}
\value{
A ggplot
}
\description{
Makes a line plot graphing the temporal evolution of data (using ggplot2).  Full functionality not implemented, or even defined...
}
\details{
This function is WORK IN PROGRESS!!  For questions about functionality or feature requests contact the author

Note that like all \code{DGVMTools} plotting functions, \code{plotTemporal} splits the data into separate panels using the \code{ggplot2::facet_wrap()}.  If you want to 'grid' the facets
using \code{ggplot2::facet_grid()} you can do so afterwards. 'gridding the facets' implies the each column and row of facets vary by one specific aspect.
For example you might have one column for each Source, and one row for each "Quantity".
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
