% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-temporal.R
\name{plotTemporal}
\alias{plotTemporal}
\title{cheese}
\usage{
plotTemporal(input.data, layers = NULL, expand.layers = TRUE,
  title = NULL, quant = NULL, cols = NULL, types = NULL,
  labels = NULL, y.label = NULL, y.lim = NULL, x.label = NULL,
  x.lim = NULL, wrap = NULL, wrap.scales = "fixed")
}
\arguments{
\item{input.data}{The data to be plotted, either as a ModelObject, DataObject or a data.table.
If it is a data.table (this is the way to compare runs or runs and data) then it must have been "pre-melted" 
so that the run/data identified are values in a column (name can specified using the "by" argument).}
}
\description{
biscuits
}

