% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTemporal.R
\name{plotTemporal}
\alias{plotTemporal}
\title{Plot temporal data}
\usage{
plotTemporal(fields, layers = NULL, gridcells = NULL,
  title = character(0), subtitle = character(0), cols = NULL,
  col.by = "Layer", col.labels = waiver(), types = NULL,
  type.by = NULL, type.labels = waiver(), sizes = NULL,
  size.by = NULL, size.labels = waiver(), alphas = NULL,
  alpha.by = NULL, alpha.labels = waiver(), y.label = NULL,
  y.lim = NULL, x.label = NULL, x.lim = NULL, facet = TRUE,
  facet.scales = "fixed", legend.position = "bottom",
  text.multiplier = NULL, plot = TRUE)
}
\arguments{
\item{fields}{The data to be plotted, either as a Field, DataObject or a list of Model/DataObjects.}

\item{layers}{A list of strings specifying which layers to plot.  Defaults to all layers.}

\item{gridcells}{A list of gridcells to be plotted in different panels, for formatting of this argument see \code{selectGridcells}.
Leave empty or NULL to plot all gridcells (but note that if this involves too many gridcells the code will stop)}

\item{title}{A character string to override the default title.  Set to NULL for no title.}

\item{subtitle}{A character string to override the default subtitle. Set to NULL for no subtitle.}

\item{col.by, type.by, size.by, alpha.by}{Character strings defining the aspects of the data which which should be used to set the colour, line type, line size (width) and alpha (transparency).
Can meaningfully take the values "Layer", "Source" or "Quantity". By default \code{col.by} is set to "Layer" and all others set to NULL, which means the different aspects are 
distinguished by different facet panels.  Thus the standard behaviour the that different Layers are distinguished by different colours, but everything is seperated into different panels.}

\item{x.label, y.label}{Character strings for the x and y axes (optional)}

\item{x.lim, y.lim}{Limits for the x and y axes (each a two-element numeric, optional)}

\item{facet}{Logical, if TRUE split the plot into panels by source.  If false, plots all data in a single panel.}

\item{facet.scales}{Character string.  If faceting (see above) use "fixed" to specify same scales on each ribbon (default), or "free"/"free_x"/"free_y" for tailored scales}

\item{legend.position}{Position of the legend, in the ggplot2 style.  Passed to the ggplot function \code{theme()}. Can be "none", "top", "bottom", "left" or "right" or two-element numeric vector}

\item{text.multiplier}{A number specifying an overall multiplier for the text on the plot.}

\item{plot}{Logical, if FALSE return the data.table of data instead of the plot
Make it bigger if the text is too small on large plots and vice-versa.}

\item{labels}{A list of character strings which are used as the labels for the lines.  Must have the same length as the layers argument (after expansion if necessary)}
}
\value{
A ggplot
}
\description{
Makes a line plot graphing the temporal evolution of data (using ggplot2).  Full functionality not implemented, or even defined...
}
\details{
This function is WORK IN PROGRESS!!  For questions about functionality or feature requests contact the author
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
