% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-temporal.R
\name{plotTemporal}
\alias{plotTemporal}
\title{Plot temporal data}
\usage{
plotTemporal(input.data, layers = NULL, expand.layers = TRUE,
  title = NULL, quant = NULL, cols = NULL, types = NULL,
  labels = NULL, y.label = NULL, y.lim = NULL, x.label = NULL,
  x.lim = NULL, facet = NULL, facet.scales = "fixed",
  legend.position = "bottom", text.multiplier = NULL)
}
\arguments{
\item{input.data}{The data to be plotted, either as a ModelObject, DataObject or a list of Model/DataObjects.}

\item{layers}{A list of strings specifying which layers to plot.  Defaults to all layers.}

\item{expand.layers}{A boolean, determines wether to expand the layers arguement.  See documentation for \code{expandLayers} for details.}

\item{title}{Main plot title (character string)}

\item{quant}{A Quantity object to provide meta-data about how to make this plot}

\item{cols, types}{Colour and types for the lines.  They do not each necessarily need to be specified, but if they are then the they need to be 
the same length as the labels arguments}

\item{labels}{A list of character strings which are used as the labels for the lines.  Must have the same length as the layers argument (after expansion if necessary)}

\item{x.label, y.label}{Character strings for the x and y axes (optional)}

\item{x.lim, y.lim}{Limits for the x and y axes (each a two-element numeric, optional)}

\item{facet}{Character string. If specified, split the data (ie melt) the data by the column specified in the argument, and then split the plot into ribbons accordingly.}

\item{facet.scales}{Character string.  If faceting (see above) use "fixed" to specify same scales on each ribbon (default), or "free"/"free_x"/"free_y" for tailored scales}

\item{legend.position}{Position of the legend, in the ggplot2 style.  Passed to the ggplot function \code{theme()}. Can be "none", "top", "bottom", "left" or "right" or two-element numeric vector}

\item{text.multiplier}{A number specifying an overall multiplier for the text on the plot.
Make it bigger if the text is too small on large plots and vice-versa.}
}
\value{
A ggplot
}
\description{
Makes a line plot graphing the temporal evolution of data (using ggplot2).  Full functionality not implemented, or even defined...
}
\details{
This function is WORK IN PROGRESS!!  For questions about functionality or feature requests contact the author
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
