% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTemporalComparison.R
\name{plotTemporalComparison}
\alias{plotTemporalComparison}
\title{Plot a comparison between two Temporal layers}
\usage{
plotTemporalComparison(
  comparisons,
  type = c("difference", "percentage.difference", "values", "nme"),
  limits = NULL,
  symmetric.scale = TRUE,
  percentage.difference.limit = 300,
  do.phase = FALSE,
  plot.zero.line = TRUE,
  ...
)
}
\arguments{
\item{comparisons}{The data to plot, must be a \code{\linkS4class{Comparison}} or a list of Comparisons}

\item{type}{A character specifying what type of plot to make. Can be "difference" (default, for a difference plot), "
percentage.difference", "values" (actual values, side-by-side).}

\item{limits}{A numeric vector with two members (lower and upper limit) to limit the plotted values.}

\item{symmetric.scale}{If plotting a differences, make the scale symmetric around zero (default is TRUE)}

\item{percentage.difference.limit}{If percentage difference to be plotted, what to limit the scale to.  Default is 300.}

\item{do.phase}{Logical, only applies to plotting Comparison objects of type "seasonal".}

\item{plot.zero.line}{Logical, if TRUE (default) plot a black line at y=0 in difference plots to better guide the eye for 
If TRUE plot the the seasonal phase, if FALSE (the default), plot the seasonal concentration.}

\item{...}{Parameters passed to \code{\link{plotTemporal}}}
}
\value{
Returns a ggplot object
}
\description{
This function is for plotting maps from \code{\linkS4class{Comparison}} objects (or a list of those Comparisons).  Three types of comparisons plots are supported: 'difference' - 
a difference map; "values" - the absolute values plotted in panels and "percentage.difference" - the percentage differences.
}
\details{
A wrapper for around \code{\link{plotTemporal}} to plot the temporal Comparisons as maps.  Extra arguments to \code{\link{plotTemporal}} can also be specified.
}
\seealso{
\code{\link{plotTemporal}},  \code{\link{compareLayers}}
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
