% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg-plotting.R
\name{plotVegMaps}
\alias{plotVegMaps}
\title{Plots a map from a temporally-averaged \code{VegObject}, a data.table, a Raster* or a SpatialPixelsDataFrame.}
\usage{
plotVegMaps(data, layers = NULL, expand.layers = TRUE, quant = NULL,
  period = NULL, run = NULL, PFT.set = global.PFTs, plot.dir = NULL,
  title = NULL, tag = NULL, layout.objs = NULL, plot.labels = NULL,
  plot.bg.col = "transparent", useLongnames = FALSE, text.multiplier = 1,
  plot.extent = NULL, limit = FALSE, limits = NULL,
  override.cols = NULL, override.cuts = NULL, special = "none",
  maxpixels = 1e+06, biome.scheme = Smith2014.scheme, biome.data = NULL,
  kappa.list = NA, kappa.position = NULL, ...)
}
\arguments{
\item{data}{The data to plot. Can be a VegObject, data.table, a SpatialPixelsDataFrame or a Raster* object.}

\item{layers}{A list of strings specifying which layers to plot.  Defaults to all layers.}

\item{expand.layers}{A boolean, determines wether to expand the layers arguement.  See documentation for \code{expandLayers} for details.}

\item{quant}{A \code{VegQuantity} object describy the quantity to be plotted.  This provides an override \code{VegQuant} when plotting a \code{VegObject}
and is useful to specify metadata (colours, plot ranges, names, etc.) when plotting other objects.}

\item{period}{The time period (represented by a \code{TemporalExtent} object), used only for plot labels and filenames.
In the case of plotting a \code{VegRun} object this is taken automatically from he object, but this provides an override.}

\item{run}{A \code{VegRun} object from which to pull metadata.  Note that normally this information is stored in the \code{VegObject}.}

\item{PFT.set}{A PFT set, necessary for exapnding layers and plotting long names.  Normally taken from the \code{VegObject}.}

\item{plot.dir}{A character string given the location for the plot to be saved. Usually the \code{run.dir} of the \code{VegObject}, but this provides an override.
If not a \code{VegObject} and not specified, this defaults to the current directory}

\item{title}{A character string to override the deafualt title.}

\item{tag}{A string with which to tag the resulting plots to specify the analysis (to differentiate them from other plots),
for example "ForcingDatasetComparison" or "SoilDepths.  Or whatever you want.}

\item{layout.objs}{List of overlays (for example coastlines or rivers) or other objects to be plotted by \code{spplot} 
so see the there for how to build them.}

\item{plot.labels}{List of character strings to be used as panel labels for summary plots and titles for the individual plots.
Sensible titles will be constructed if this is not specified.}

\item{plot.bg.col}{Colour string for the plot background.}

\item{useLongnames}{Boolean, if TRUE replace PFT IDs with the PFT's full names on the plots.}

\item{text.multiplier}{A number specifying an overall multiplier for the text on the plot.
Make it bigger if the text is too small on large plots and vice-versa}

\item{plot.extent}{An extent object to limit the plot area.}

\item{limit}{Boolean, whether or not to limit the plotted values between a range, either the limits argument below,
or the range of the plot.}

\item{limits}{A numeric vector with two members (lower and upper limit) to limit the plotted values}

\item{override.cols}{A colour palette function to override the defaults}

\item{override.cuts}{Cut ranges (a numeric vector) to override the defaults.}

\item{special}{A character string to indicate certain "special modes" which modifies the behaviour of the function a bit.
Special modes are currectly "fraction", "difference", "percentage.difference" and "firert" (fire return time).
Biomes, dominant PFT and burnt area fraction should be added soon.}

\item{maxpixels}{Maximum number of pixels to plot (see \code{raster} version of \code{spplot})}

\item{biome.scheme}{A biome scheme of type BiomeScheme.  This control the defines the biomes types, names, colours for 
plotting etc (only used if argument "special" = "biomes")}

\item{biome.data}{A biome dataset, as a Raster or a SpatialDataset (can be left NULL to plot no data, 
only used if argument "special" = "biomes")}

\item{kappa.list}{A list of Kappa scores (to be coded in a more friendly style, 
only used if argument "special" = "biomes")}

\item{kappa.position}{A numeric vector with two elements to define the position (in Lon-Lat) of the overall 
Kappa score on the plot (only used if argument "special" = "biomes")}

\item{...}{Extra arguments to be be passed to \code{spplot} and therefore also to \code{lattice::levelplot}.

This function is heavily used by the benchmarking functions and can be very useful to the user for making quick plots
in standard benchmarking and post-processing.  It is also highly customisable for final results plots for papers and so on.
However, the \code{plotGGSpatial} function makes pretty plots with a simpler syntax, but with less flexibility.

The function works best for \code{VegObjects} (which contain a lot of useful metadata).}
}
\value{
Returns a plot object (from spplot)
}
\description{
This is a heavy lifting function for plotting models variables with flexibly, but with a high degree of automation. It's main use is to plot a map from a \code{VegObject}, although plotting a Raster* is also useful.
It has a really large amount of parameters for a huge amount of flexibility.  However they are all set to sensible defaults,
so that in the case of plotting a \code{VegObject} all you *need* to supply is the data itself, everything else is either set to a sensible default,
or provided by the \code{VegObject} itself.  It is basically a very complex wrapper for spplot, and can plot things like biomes, dominant PFTs, months of maximum values, 
burnt fraction on an approximately logarithic scale etc.  It returns a plot, which will need to be displayed using a \code{print()} command.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\seealso{
\code{plotGGSpatial}, \code{expandLayers}, \code{sp::spplot}, \code{latice::levelplot}
}

