% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{proportionsComparison}
\alias{proportionsComparison}
\title{Compare relative proportions data}
\usage{
proportionsComparison(dt1, dt2, name1, name2, verbose = TRUE)
}
\arguments{
\item{dt1}{The first of the datasets to be compared to each other, as a data.table with a column for the  elative proportion of each class (sum of classes must equal one for metric to be meaningful)}

\item{dt2}{The second of the datasets to be compared to each other, as a data.table with a column for the  elative proportion of each class (sum of classes must equal one for metric to be meaningful)}

\item{name1}{A character string giving the name of the first dataset (for making a useful id)}

\item{name2}{A character string giving the name of the second dataset (for making a useful id)}

\item{verbose}{A logical, if TRUE print out all the Kappa scores

Note that there are many other slots in a Statistics object which will not be filled in the resulting object because they are not for relative proportions data.}
}
\value{
A spatial comparison object
}
\description{
Compares two datasets of relative proportions of multiple classes (sum of classes equals one at each point) where the total for each  data where the totally value for each. 
Specifically calculates and returns a Statistics object (which contains many metrics) with the relevants slots are Manhattan Metric (MM) and Square Chord Distance (SCD).
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
