% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-classes.R, R/quantities.R
\docType{class}
\name{Quantity-class}
\alias{Quantity-class}
\alias{dgvm.quantities}
\title{Class to hold meta-data for a vegetation quantity}
\format{\code{dgvm.quantities} is list of \code{Quantity} objects that store meta-data for common output variables from supported models as well as some standard quantities
which should be derivable from all model output.}
\usage{
dgvm.quantities
}
\description{
This class hold meta-data for specific vegetation or land surface quantities like "lai" (LAI) or "mwcont_upper" (mean monthly water).  
Includes metadata about the quantity (the units it is measured in for example), values for default plot proprties (color scales, plot ranges) and the aggragating method - ie whether the quantity shoudd generally be summed (for example monthly burnt area or C mass per PFT) or averaged (for example monthly soil water content).
See the 'Slots' documentation below.
Note that the user will probably need to define their own, and modify these for their own analysis and plots

A global scope list of all many quantites for many models defined by the \code{dgvm.quantities} list below.
Once the package is loaded this is just a standard R list so can be modified and extended as you wish.  It can be conveniently accesed by the \code{lookupQuantity()} function.
}
\details{
These quantities have sensible colour schemes, long names and cut ranges for standard global runs,
but they will likely need to be modified and new ones aded for a specific analysis.
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular vegetation quantity, should match with the name of a particular model output variable.  Recommended to be alphanumeric because it is used to construct file names.}

\item{\code{name}}{A longer character string to provide a more complete description of this quantity}

\item{\code{type}}{A character string defining if this quantity is defined per PFT ("PFT"), per month ("monthly"), or something else.  The first two have a specific meaning to DGVMTools, but in principle the use can define anything.}

\item{\code{units}}{A character string defining the units this quantity is defined in.  Possibly formally link to udunits2?}

\item{\code{colours}}{A function that returns a colour scale suited for plotting this quantity.}

\item{\code{aggregate.method}}{A character string defining the default method for how to aggregate the quantity, either "sum" or "average"}

\item{\code{model}}{Either a the string "Standard" to denote that this is a  standard quantity to be compared across all model and data, or vector of model names to denote to which models this Quantity is applicable.}

\item{\code{cf.name}}{A character string for the "standard_name" or "long_name" attribute for a CF-compliant netCDF file.  Won't make sense for all variables (not all DGVM quantities have a CF defined variable),
but it makes sense to use this where possible.}
}}

\seealso{
\code{lookupQuantity}
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{datasets}
