\name{rvc.ggplot.theme}
\alias{rvc.ggplot.theme}
\title{Returns the predefined theme for ggplots}
\description{
  Returns the predefined theme for ggplots.
}
\usage{
    rvc.ggplot.theme(x)
}
\arguments{
  \item{x}{String, currently predefined themes available: "map", "scatter", "timeseries".}
}
\value{
  The requested ggplot theme list of FALSE, if an unknown string is given.
}
\author{
  Joerg Steinkamp <joerg.steinkamp@senckenberg.de>
}
