% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{seasonalComparison}
\alias{seasonalComparison}
\title{Compare seasonal data}
\usage{
seasonalComparison(x, layers1, layers2, additional, verbose = TRUE)
}
\arguments{
\item{x}{A data.table containing the spatial-temporal-annual columns and two columns containg the data to be compared}

\item{layers1}{A character string giving the first layer to compare (should be a column in x).  For the normalised metrics, this is the *modelled* values.}

\item{layers2}{A character string giving the second layer to compare (should be a column in x).  For the normalised metrics, this is the *observed* values.}

\item{additional}{A list of functions define additions metrics, see the custom.metrics argument of \code{compareLayers()}}

\item{verbose}{A logical, if TRUE print out all the metric scores}
}
\value{
A named list of metric statistics
}
\description{
Compares two datasets of monthly continuous data. Specifically calculates and returns a Statistics object (which contains many metrics) given numeric two vectors of equal size
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
