% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ext-import.R
\name{seasonalityIndex}
\alias{seasonalityIndex}
\title{Calculates the seasonality index}
\usage{
seasonalityIndex(x, wgt = c(31, 28.25, 31, 30, 31, 30, 31, 31, 30, 31, 30,
  31), add.origin = TRUE, variable = "strength", debug = FALSE)
}
\arguments{
\item{x}{values, e.g. monthly precipitation values}

\item{wgt}{weight for the angle width of each value, e.g. days per month (default)}

\item{add.origin}{add 0 at the begin of both input vectors.}

\item{variable}{what to return: 's' seasonality index, 'a' or 'di' for angle or direction between 0 and 2pi; 'day' or 'doy' angle as fraction between 0 and sum(wgt); 'both': index and angle; '^b.*day' index and prev. fraction.}

\item{debug}{return the all values as data.frame.}
}
\value{
the seasonality index, by default (see 'variable' for other options) following Markhan (1970)
}
\description{
Calculates a seasonality index based on a monthly vector by default, originally developed for precipitation. With specifying an additional wheighting vector different to months can be used.
}
\examples{
x <- matrix(c(c(31, 28.25, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31),
              c(135, 97, 77, 49, 22, 4, 2, 1, 10, 18, 58, 127),
              c(10, 18, 58, 127, 135, 97, 77, 49, 22, 4, 2, 1),
              c(1, 10, 18, 58, 127, 135, 97, 77, 49, 22, 4, 2),
              c(77, 49, 22, 4, 2, 1, 10, 18, 58, 127, 135, 97),
              c(2, 4, 49, 135, 77, 22, 1, 10, 18, 127, 97, 58),
              c(135, 77, 22, 1, 10, 18, 127, 97, 58, 2, 4, 49)),
            nrow=12)
titles <- c("circular", "early", "early/mid", "late/mid", "late", "dual 1", "dual 2")
for (i in 1:ncol(x)) {
  si <- seasonalityIndex(x[, i], variable="both")
  z  <- seasonalityIndex(x[, i], debug=TRUE)
  plot(cumsum(z$vx), cumsum(z$vy), type="b", lwd=3, main=titles[i], asp=1)
  lines(c(0, sum(z$x) * si[1] * sin(Arg(sum(z$c))+pi/2)),
        c(0, sum(z$x) * si[1] * cos(Arg(sum(z$c))-pi/2)), lty="dashed", col="red", lwd=3)
  print(paste0("Seasonality index (", titles[i], "): ", signif(si[1], 2)))
  lines(c(0, sin(si[2]) * si[1] * sum(x[, i])),
        c(0, cos(si[2]) * si[1] * sum(x[, i])), lty="dotted", col="blue")
}
}
\references{
Charles G. Markhan (1970) Seasonality of precipitation in the United States, Annals of the Association of American Geographers, 60:3, 593-597, DOI: \href{http://dx.doi.org/10.1111/j.1467-8306.1970.tb00743.x}{10.1111/j.1467-8306.1970.tb00743.x}
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}
