% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectGridcells.R
\name{selectGridcells}
\alias{selectGridcells}
\title{Select gridcells}
\usage{
selectGridcells(
  x,
  gridcells,
  spatial.extent.id = NULL,
  tolerance = NULL,
  decimal.places = NULL,
  cover.threshold = NULL,
  ...
)
}
\arguments{
\item{x}{The Field, data.table or data.frame from which the gridcells should be selected.  Note that a data.table or data.frame 
should have columns "Lon" and "Lat" included.}

\item{gridcells}{The gridcells to be extracted.  These can be specified as one of the following:
\itemize{
 \item{A simple two-element numeric to pull out one gridcell (ordering is lon, lat)} 
 \item{A data.frame or data.table in which the first two columns are assumed to be longitude and latitude.}
 \item{An sf:sf object.  By default points will be selected that lie under a polygon (feature), but see the argument \code{cover.threshold} to 
 select based on an overlap threshold.  Note also that is very easy to get sf objects representing countries from the rnaturalearth package,
 which, in combination with this function, provides a very easy way to subset by country.}
}}

\item{spatial.extent.id}{A character string to describe the gridcells selected.  When selecting gridcells from a Field you *must* specify 
a \code{spatial.extent.id} for meta-data consistency (you have free choice here, but avoid spaces).}

\item{tolerance}{A single numeric specifying how close a required gridcell in gridcells must be to one in \code{x}.
Doesn't currently work, non-exact matching is not implemented!  Contact the author if this is a critical feature for you.}

\item{decimal.places}{A single numeric specifying how many decimal place to which the coordinates should rounded to facilitate a match.
If not specified not rounding is done.}

\item{cover.threshold}{An optional numeric specifying what fraction of the gridcell must be covered by a feature in the case the \code{gricells} argument 
is a SpatialPolygonsDataFrame.  Note that is done using the \code{getCover} argument of \code{raster::rasterize()} which is only sensitive to about 1\% cover. 
fractions}

\item{...}{Further arguments.  Currently not used.}
}
\value{
A Field, data.frame or data.table depending on the class of the main input x.
}
\description{
Select gridcells from a Field (or a data.table or a data.frame) using either gridcell coordinates or polgyons (from, for example, a shapefile).
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
