% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectGridcells.R
\name{selectGridcells}
\alias{selectGridcells}
\title{Select gridcells from a Data/Field}
\usage{
selectGridcells(x, gridcells, tolerance = NULL,
  spatial.extent.id = NULL, decimal.places = NULL)
}
\arguments{
\item{x}{The Field, DataObject, data.table or data.frame from which the gridcells should be selected.  Note that a data.table or data.frame 
should have columns "Lon" and "Lat" included.}

\item{gridcells}{The gridcells to be extracted.  Can either be a simple two-element numeric to pull out one gridcell (ordering is lon, lat), or it can be a data.frame or data.table
in which the first two columns are assumed to be longitude and latitude.}

\item{tolerance}{A single numeric specifying how close a required gridcell in gridcells must be to one in x.  Doesn't currently work, non-exact matching is not implemented.}

\item{spatial.extent.id}{A character string to describe the gridcells selected, this is required for meta-data consistency.}

\item{decimal.places}{A single numeric specifying how many decimal place to which the coordinates should rounded to facilitate a match.}
}
\value{
A Field, DataObject, data.table or data.frame depending on the type of the input x.
}
\description{
Select gridcells from a Data/Field
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
