% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tools.R
\name{setKeyDGVM}
\alias{setKeyDGVM}
\title{Sets keys on data.table in order Lon, Lat, Year (if present)}
\usage{
setKeyDGVM(dt)
}
\arguments{
\item{dt}{The data.table for which to set the key}
}
\value{
Returns nothing because changes the original data.table by reference (this is the data.table way)
}
\description{
Keys should be set on all data.table object for sorts, joins, DGVMTool-defined operators etc.  
 This function should be called on a data.table stored in a ModelObject after it has been created,
 including if it was created by avergaing another data.table because it seems as keys are not conserved.
}

