% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome-schemes.R
\docType{data}
\name{Smith2014.scheme}
\alias{Smith2014.scheme}
\title{Meta-data describing the Smith 2014 et al. 2014 biome scheme for LPJ-GUESS output.}
\format{An object of class \code{BiomeScheme} of length 1.}
\usage{
Smith2014.scheme
}
\description{
Meta-data describing the Smith 2014 et al. 2014 biome scheme for LPJ-GUESS output.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}

