% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000rvc-classes.R
\docType{class}
\name{SpatialComparison-class}
\alias{SpatialComparison-class}
\title{Result of comparing a model raster to a data raster}
\description{
This class stores the rasters (model, data, difference and precentage difference) and the statistics (R^2 etc) resulting when two rasters are compared using \code{compareRunToSpatialDataset}
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular raster compariosn.  Recommended to be alphanumeric because it is used to construct file names.}

\item{\code{R.squ}}{The R squared between the model and the data}

\item{\code{P.cor}}{The Pearsons product moment correlation between the model and the the data,}

\item{\code{RMSE}}{The Root Mean Squared Error between the model and the the data}

\item{\code{mean.diff}}{The difference between the model and the the data (all gridcells)}

\item{\code{sd.diff}}{The standard deviation of the difference between the model and the the data (all gridcells)}
}}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

