% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-classes.R
\docType{class}
\name{SpatialComparison-class}
\alias{SpatialComparison-class}
\title{Result of comparing a model raster to a data raster}
\description{
This class stores the rasters (model, data, difference and precentage difference) and the statistics (R^2 etc) resulting when two rasters are compared using \code{compareRunToSpatialDataset}
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular raster compariosn.  Recommended to be alphanumeric because it is used to construct file names.}

\item{\code{R2}}{The R^2 between the model and the data}

\item{\code{R2.eff}}{The model efficiency of the model compared to the data}

\item{\code{P.cor}}{The Pearsons product moment correlation between the model and the the data,}

\item{\code{ME}}{The Mean Error (mean of residuals) between the model and the data}

\item{\code{NME}}{The Normalised Mean Error (mean of residuals after the residuals have been divided the value of the data point) between the model and the data}

\item{\code{NMSE}}{The Normalised Square Mean Error (mean of the square of the residuals after the residuals have been divided the value of the data point) between the model and the data}

\item{\code{RMSE}}{The Root Mean Squared Error between the model and the the data}

\item{\code{sd.diff}}{The standard deviation of the difference between the model and the the data (all gridcells)}

\item{\code{Kappa}}{The overall Cohen's Kappa obtained when comparing categorical variables.}

\item{\code{individual.Kappas}}{The individual Cohen's Kappas for each category when comparing categorical data}
}}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

