% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-classes.R
\docType{class}
\name{SpatialExtent-class}
\alias{SpatialExtent-class}
\title{Spatial extent for defining the spatial area covered by model output or other spatial data.}
\description{
A simple S4 class to define a spatial extent.  It contains a raster::Extent object
}
\section{Slots}{

\describe{
\item{\code{id}}{A unique character string to identify this particular spatial extent.  Recommended to be alphanumeric because it is used to construct file names.}

\item{\code{name}}{A character string to describe the spatial extent. Used for building plot labels, not file names, so doesn't need to be alphanumeric and can so can be prettier.}

\item{\code{xmin}}{Minimum x (longitude) of this extent (inherited from raster::Extent)}

\item{\code{xmax}}{Maximum x (longitude) of this extent (inherited from raster::Extent)}

\item{\code{ymin}}{Minimum y (latitude) of this extent (inherited from raster::Extent)}

\item{\code{ymax}}{Maximum y (latitude) of this extent (inherited from raster::Extent)}
}}

\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}

llal
}
