% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{STAInfo-class}
\alias{STAInfo-class}
\title{STAInfo class}
\arguments{
\item{first.year}{A single numeric, the first year.}

\item{last.year}{A single numeric, the last year.}

\item{year.aggregate.method}{A character specifying how the years have been aggregated, for example "mean", or "sum" or "var". See aggregateYears.
If no yearly aggregation has been applied it should be NULL.}

\item{spatial.extent}{This can be of any type that can be used but DGVMTools::crop, and stores the current spatial extent.
But default (and with no cropping) it is simple teh raster::Extent object of the whole domain.}

\item{spatial.aggregate.method}{A character method specifying how the spatial extent has been aggregated, for eample "mean" or "sum",
see aggregateSpatial.  If no spatial aggregation has been applied it should be NULL.}

\item{subannual.original}{A character string specifying the original sub-annual resolution of this data, eg. "Year", Month", "Day"}

\item{subannual.resolution}{A character string specifying the current sub-annual resolution of this data, eg. "Year", Month", "Day"}

\item{subannual.aggregate.method}{A character specifying how the subannual perods have been aggregated, for example "mean", "max", "sum" or "var". 
See aggregateSubannual(). If no sub-annual aggregation has been applied it should be NULL.}
}
\description{
This class encapsulations all the Spatial (S, longitude and latidude), Temporal (T, monthly, daily etc.) and Annual (A, years included) 
Information  (hence the name 'STAInfo') about a particular DGVMTools::Field.  Normally the user won't have to deal with this (it is mostly used internally),
but it can be handy, for example to extract the STAInfo from one Field, and use it extract another field with the same dimensions.
}
\details{
This is mostly a behind-the-scenes class which is partly just created to bundle together a lot of dimension information in a tidy form.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
