% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small-utility-functions.R
\name{stats_mode}
\alias{stats_mode}
\title{Statistical mode}
\usage{
stats_mode(x)
}
\arguments{
\item{x}{vector from which to find the most common value}
}
\value{
The mode, ie the most common value. In case of ties the return is first in the original vector
}
\description{
Calculate the statistical mode with is strangely missing from base R.  In a tie it returns the first of the most common values
encountered in the original vector.
}
\keyword{internal}
