% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-misc.R
\docType{methods}
\name{Summary-methods}
\alias{Summary-methods}
\alias{summary,ModelRun-method}
\alias{summary}
\title{Summary methods}
\usage{
summary(object, ...)

\S4method{summary}{ModelRun}(object, ...)
}
\arguments{
\item{object}{a DGVMTools object}

\item{...}{Other arguments, not currently used}
}
\value{
A list of strings
}
\description{
Print easy to read summaries of DGVMTools objects
}
\author{
Joerg Steinkamp \email{joerg.steinkamp@senckenberg.de}
}
