% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome-schemes.R
\docType{data}
\name{supported.biome.schemes}
\alias{supported.biome.schemes}
\title{Currently supported biome schemes}
\format{An object of class \code{list} of length 5.}
\usage{
supported.biome.schemes
}
\description{
The biomes schemes that are currently supported for LPJ-GUESS output.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\seealso{
\code{BiomeClassification}, \code{readHandPBiomes}
}
\keyword{datasets}

