% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome-schemes.R
\docType{data}
\name{supported.biome.schemes}
\alias{supported.biome.schemes}
\title{Currently supported biome schemes}
\format{An object of class \code{list} of length 11.}
\usage{
supported.biome.schemes
}
\description{
This is a list of all BiomeSchemes defined by DGVMTools.
}
\examples{
\donttest{
 
# In this example we derive and plot the Smith et al 2014 and the 
# Forrest et al 2015 biome classifications    
 
# First define a Source
test.dir <- system.file("extdata", "LPJ-GUESS_Runs", "CentralEurope", package = "DGVMTools")
test.Source <- defineSource(name = "LPJ-GUESS run", dir = test.dir,  format = GUESS)

# Smith et al. 2014
Smith2014.biomes <- getBiomes(source = test.Source, scheme = Smith2014BiomeScheme, 
                              year.aggregate.method = "mean")
print(plotSpatial(Smith2014.biomes))

# Forrest et al. 2014
Forrest2015.biomes <- getBiomes(source = test.Source, scheme = Forrest2015BiomeScheme, 
                                year.aggregate.method = "mean")
print(plotSpatial(Forrest2015.biomes))

}
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{datasets}
