% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biome-schemes.R
\docType{data}
\name{supported.biome.schemes}
\alias{supported.biome.schemes}
\title{Currently supported biome schemes}
\format{An object of class \code{list} of length 11.}
\usage{
supported.biome.schemes
}
\description{
This is a list of all BiomeSchemes defined by DGVMTools.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{datasets}
