% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FireMIP.R
\name{toFireMIPPFTs}
\alias{toFireMIPPFTs}
\title{Aggregate to FireMIP PFTs}
\usage{
toFireMIPPFTs(input.data, remove.agriculture = FALSE)
}
\arguments{
\item{input.data}{A ModelObject from a FireMIP model}

\item{remove.agriculture}{Logical, if TRUE, remove the cropland PFT and scale up the rest to compensate}
}
\value{
A ModelObject with the standard FireMIP PFTs
}
\description{
Aggregates the PFTs in a FireMIP ModelObject to a standard set of PFTs
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
