% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-framework-functions.R
\name{trimFieldsForPlotting}
\alias{trimFieldsForPlotting}
\title{Eubsets data from Field for plotting}
\usage{
trimFieldsForPlotting(fields, layers, years = NULL, days = NULL,
  months = NULL, seasons = NULL, gridcells = NULL)
}
\arguments{
\item{fields}{The list of Fields to be plotted (should have been check by santiseFieldsForPlotting first)}

\item{layers}{A character vector of the layers to be plotted}

\item{years}{The years to be extracted (as a numeric vector), if NULL all years are used}

\item{days}{The days to be extracted (as a numeric vector), if NULL all days are used}

\item{months}{The months to be extracted (as a numeric vector), if NULL all months are used}

\item{seasons}{The months to be extracted (as a character vector), if NULL all seasons are used}
}
\value{
Returns a list of Fields
}
\description{
This is an internal helper function which pulls out the data needed to make a plot from a bunch of Fields,
and combines them and melts them into a big data.table for plotting with ggplot2.  It also determines an appropriate title 
and subtitle
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
