% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour-palettes.R
\name{veg.palettes}
\alias{veg.palettes}
\alias{veg.palette}
\alias{lai.palette}
\alias{cmass.palette}
\alias{difference.palette}
\alias{fire.palette}
\alias{reversed.fire.palette}
\alias{reversed.tim.colors}
\alias{reversed.viridis}
\alias{reversed.magma}
\alias{reversed.inferno}
\alias{reversed.plasma}
\title{Helpful palettes for vegetation maps}
\usage{
veg.palette(n)

lai.palette(n)

cmass.palette(n)

difference.palette(n)

fire.palette(n)

reversed.fire.palette(n)

reversed.tim.colors(n)

reversed.viridis(n)

reversed.magma(n)

reversed.inferno(n)

reversed.plasma(n)
}
\arguments{
\item{n}{Number of colour shades required}
}
\description{
\describe{
\item{veg.palette}{A nice white-green-brown color scheme.  Good for vegetation cover}
\item{lai.palette}{Default colour scheme for LAI}
\item{cmass.palette}{Default colour scheme for biomass}
\item{difference.palette}{Colour scheme for differences. Symmetrical, centred on white}
\item{fire.palette}{Colour scheme for measures of fire activity like fire return time and area burned (from more-to-less fire)}
\item{reversed.fire.palette}{Colour scheme for measures of fire activity like fire return time and area burned (from less-to-more fire)}
\item{reversed.tim.colors}{The fields::tim.colors scheme reversed to go from red (hot/dry) to blue (cold/wet)}
}
}
