% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000rvc-classes.R
\docType{class}
\name{VegRun-class}
\alias{VegRun-class}
\title{An S4 class to contain metadata and, optionally, data and benchmarks for a single vegetation run}
\description{
A \code{VegRun} object contains the metadata concerning the an inherited \code{VegRunInfo} object
 and the actual model data run as \code{VegObjects} in a list in slot \code{objects} and comparisions to datasets 
 as \code{BiomeComparison} and \code{RasterComparison} in a list in slot \code{benchmarks}.
Such objects can be built by calls to \code{getVegObject()}, \code{getVegSpatial()}, \code{getVegTemporal()}, \code{calcNewVegObj}, \code{compareRunToSpatialDataset()}
 and \code{compareBiomes()}, and saved to the \code{VegRun} using \code{addToVegRun()}. 
 
Slots can be accessed by user directly, but more easily and usefully by functions \code{XXXX}
}
\section{Slots}{

\describe{
\item{\code{objects}}{List of \code{VegObjects} saved in this run}

\item{\code{benchmarks}}{List of benchmarks (\code{BiomeComparisons} and \code{RasterComparisons}) performed and saved for this run.
The following slots are inherited from \code{VegRunInfo}.}

\item{\code{id}}{A unique character string to identify this particular model run.  Recommended to be alphanumeric because it is used to construct file names. (Mandatory)}

\item{\code{model}}{A character string to identify what model produced this run.  Can currently be "LPJ-GUESS", "LPJ-GUESS-SPITFIRE" or "aDGVM". (Mandatory)}

\item{\code{pft.set}}{A list of PFT objects which includes all the PFTs used is this model run (Mandatory)}

\item{\code{description}}{A character string describing this run, ie. "LPJ-GUESS v3.1"}

\item{\code{run.dir}}{The location of this run on the file system (Mandatory)}

\item{\code{driving.data}}{A character string identifying the climate or other data used to produce this model run}

\item{\code{map.overlay}}{A character defining which map overlay to plot as standard on for this model run.  Can be:
\itemize{
 \item A single string defining one of the standard maps from the maps or mapdata package (ie "world" or "worldHires" or "italy", or "rivers")
 \item A further list containing in it's first argument a string (as above) and the further arguments interior.lines (logical), 
  lwd (numeric) and col (string representing a colour), see \code{makeOverlay} for details
  \item A string specifying some non-standard overlays, not yet defined
}
These data come from the \code{mapdata} package}

\item{\code{lonlat.offset}}{A numeric of length 1 or 2 to define the offsets to Lon and Lat to centre the modelled localities.}

\item{\code{year.offset}}{A numeric of length 1 to match be added to the simulation years to convert them to calendar years}

\item{\code{tolerance}}{The tolerance arguement when converting uneven spaced grids to regular rasters for plotting}

\item{\code{london.centre}}{If TRUE, ensure that the longitudes are (-180,180) instead of (0,360)}

\item{\code{fill.col}}{A string to define an R colour used when plotting this run as a histogram or scatter plot or so}

\item{\code{line.col}}{A string to define an R colour used when plotting this runs as a line graph}

\item{\code{line.width}}{A numeric to define the width of a line representing this model run}

\item{\code{line.type}}{A numeric to define the line style representing this model run}

\item{\code{landuseSimulated}}{If TRUE it can be assumed that land use has been simulated for this run and so no correction for land use need be applied before benchmarking.}
}}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}

